Prepare the dataset in the proper format for R2jags

data_preparation prepares the dataset in the proper format for R2jags and returns a list of elements that run_model inherits via the argument data.

data_preparation(data, measure)

Arguments

data

A data-frame of the one-trial-per-row format with arm-level data. See 'Format' in run_model.

measure

Character string indicating the effect measure with values "OR", "MD", "SMD", or "ROM" for the odds ratio, mean difference, standardised mean difference and ratio of means, respectively.

Value

A list of data-frames on the following elements to be passed to run_model:

m

The number of missing participant outcome data in each trial-arm (see 'Details').

N

The number of randomised participants in each trial-arm.

t

The intervention identifier in each trial-arm.

I

The pseudo-data-frame I (see 'Details').

r

The number of observed events of the outcome in each trial-arm, when the outcome is binary.

y0

The observed mean value of the outcome in each trial-arm, when the outcome is continuous.

se0

The observed standard deviation of the outcome in each trial-arm, when the outcome is continuous.

Details

data_preparation prepares the data for the Bayesian analysis. It checks whether the element m exists in the data argument (See 'Format' in run_model). If this element is missing, data_preparation creates a pseudo-data-frame for m that has the zero value for the observed trial-arms, and NA for the unobserved trial-arms, and the pseudo-data-frame I that is identical with the pseudo-data-frame for m. If the element m exists in the data and has values only for some trial-arms, the pseudo-data-frame for m is identical to m for the corresponding trial-arms, and the pseudo-data-frame I has the value one for these trial-arms. Both pseudo-data-frames aim to retain the trials without information on missing participant outcome data.

Furthermore, data_preparation sorts the interventions across the arms of each trial in an ascending order and correspondingly the remaining elements in data (See 'Format' in run_model). data_preparation considers the first column in t as being the control arm for every trial. Thus, this sorting ensures that interventions with a lower identifier are consistently treated as the control arm in each trial. This case is relevant in non-star-shaped networks.

See also

Author

Loukia M. Spineli