League heatmap of all possible comparisons: estimation

A function to create a heatmap with the estimated effect size of all possible comparisons of interventions in the network. league.heatmap can be used for a random-effects or fixed-effect network meta-analysis and network meta-regression. league.heatmap is applied for one outcome only.

league.heatmap(full, cov.value = NULL, drug.names)

Arguments

full
An object of S3 class run.model for network meta-analysis or run.metareg for network meta-regression. See 'Value' in run.model and run.metareg.

cov.value
A vector of two elements in the following order: a number for the covariate value of interest and a character for the name of the covariate.

drug.names
A vector of labels with the name of the interventions in the order they appear in the argument data of run.model. If the argument drug.names is not defined, the order of the interventions as they appear in data is used, instead.

Value

A league heatmap of the posterior mean and 95% credible interval of all possible comparisons in the off-diagonals, and the posterior mean of the SUCRA values in the diagonal.

Details

The rows and columns of the heatmap display the names of interventions which are sorted by decreasing order from the best to the worst based on their SUCRA value (Salanti et al., 2011). The off-diagonals contain the posterior mean and 95% credible interval of the effect measure of the corresponding comparisons. The main diagonal contains the SUCRA values of the corresponding interventions when the argument full refers to the run.model function. When the argument full refers to the run.metareg function, the p-score (Rücker and Schwarzer, 2015) is calculated for each intervention while taking into account the covariate value in the argument cov.value. P-score is the 'frequentist analogue to SUCRA' (Rücker and Schwarzer, 2015).

Results in the lower triangle refer to comparisons in the opposite direction after converting negative values into positive values (in absolute or logarithmic scale), and vice versa. Darker shades of red and blue correspond to larger treatment effects in the upper and lower triangle, respectively, for a beneficial outcome, and the other way around for a harmful outcome. Odds ratio and ratio of means are reported in the original scale after exponentiation of the logarithmic scale.

Comparisons between interventions should be read from left to right. Therefore, each cell refers to the corresponding row-defining intervention against the column-defining intervention. Results that indicate strong evidence in favor of the row-defining intervention (i.e. the respective 95% credible interval does not include the null value) are indicated in bold.

league.heatmap can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.

References

Rücker G, Schwarzer G. Ranking treatments in frequentist network meta-analysis works without resampling methods. BMC Med Res Methodol 2015;15:58. [doi: 10.1186/s12874-015-0060-8 ]

Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. J Clin Epidemiol 2011;64(2):163--71. [doi: 10.1016/j.jclinepi.2010.03.016 ]

See also

Author

Loukia M. Spineli, Chrysostomos Kalyvas, Katerina Papadimitropoulou

Examples

data("nma.liu2013")

if (FALSE) {
# Perform a random-effects network meta-analysis
res <- run.model(data = nma.liu2013,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = c(0, 0),
                 var.misspar = 1,
                 D = 1,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# The names of the interventions in the order they appear in the dataset
interv.names <- c("placebo", "pramipexole", "serotonin–norepinephrine reuptake inhibitor",
                  "serotonin reuptake inhibitor", "tricyclic antidepressant", "pergolide")

# Create the league heatmap
league.heatmap(full = res, drug.names = interv.names)
}