End-user-ready results for network meta-regression

Illustrates the effect estimates, predictions and regression coefficients of comparisons with a specified comparator intervention and also exports these results to an Excel file.

metareg_plot(full, reg, compar, cov_value, drug_names, save_xls)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

reg

An object of S3 class run_metareg. See 'Value' in run_metareg.

compar

A character to indicate the comparator intervention. It must be any name found in drug_names.

cov_value

A list of two elements in the following order: a number for the covariate value of interest (see 'Arguments' in run_metareg), and a character to indicate the name of the covariate. See also 'Details'.

drug_names

A vector of labels with the name of the interventions in the order they appear in the argument data of run_model. If drug_names is not defined, the order of the interventions as they appear in data is used, instead.

save_xls

Logical to indicate whether to export the tabulated results to an 'xlsx' file (via the write_xlsx function of the R-package writexl) at the working directory of the user. The default is FALSE (do not export).

Value

metareg_plot prints on the R console a message on the most parsimonious model (if any) based on the DIC (in red text). Furthermore, the function returns the following list of elements:

table_estimates

The posterior mean, and 95% credible interval of the summary effect measure (according to the argument measure defined in run_model) for each comparison with the selected intervention under network meta-analysis and network meta-regression based on the specified cov_value.

table_predictions

The posterior mean, and 95% prediction interval of the summary effect measure (according to the argument measure defined in run_model) for each comparison with the selected intervention under network meta-analysis and network meta-regression based on the specified cov_value.

table_model_assessment

The DIC, total residual deviance, number of effective parameters, and the posterior median and 95% credible interval of between-trial standard deviation (tau) under each model (Spiegelhalter et al., 2002). When a fixed-effect model has been performed, metareg_plot does not return results on tau. For a binary outcome, the results refer to the odds ratio scale.

table_regression_coeffients

The posterior mean and 95% credible interval of the regression coefficient(s) (according to the argument covar_assumption defined in run_metareg). For a binary outcome, the results refer to the odds ratio scale.

interval_plot

A forest plot on the estimated and predicted effect sizes of comparisons with the selected comparator intervention under network meta-analysis and network meta-regression based on the specified cov_value. See 'Details' and 'Value' in forestplot_metareg.

sucra_scatterplot

A scatterplot of the SUCRA values from the network meta-analysis against the SUCRA values from the network meta-regression based on the specified cov_value. See 'Details' and 'Value' in scatterplot_sucra.

Details

The deviance information criterion (DIC) of the network meta-analysis model is compared with the DIC of the network meta-regression model. If the difference in DIC exceeds 5, the network meta-regression model is preferred; if the difference in DIC is less than -5, the network meta-analysis model is preferred; otherwise, there is little to choose between the compared models.

When the covariate is binary, specify in the second element of cov_value the name of the level for which the output will be created.

Furthermore, metareg_plot exports all tabulated results to separate 'xlsx' files (via the write_xlsx function of the R-package writexl) to the working directory of the user.

metareg_plot can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. J Clin Epidemiol 2011;64(2):163--71. doi: 10.1016/j.jclinepi.2010.03.016

Spiegelhalter DJ, Best NG, Carlin BP, van der Linde A. Bayesian measures of model complexity and fit. J R Stat Soc B 2002;64(4):583--616. doi: 10.1111/1467-9868.00353

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# Read results from 'run_metareg' (exchangeable structure)
reg <- readRDS(system.file('extdata/reg_baker.rds', package = 'rnmamod'))

# Publication year as the covariate
pub_year <- c(1996, 1998, 1999, 2000, 2000, 2001, rep(2002, 5), 2003, 2003,
              rep(2005, 4), 2006, 2006, 2007, 2007)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "budesonide", "budesonide plus formoterol",
                  "fluticasone", "fluticasone plus salmeterol",
                  "formoterol", "salmeterol", "tiotropium")

# Plot the results from both models for all comparisons with salmeterol and
# publication year 2000
metareg_plot(full = res,
             reg = reg,
             compar = "salmeterol",
             cov_value = list(2000, "publication year"),
             drug_names = interv_names)