Indicate the mean value of the distribution of the missingness parameter

This function generates the mean value of the distribution of the missingness parameter in the proper format depending on the assumed structure of the missingness parameter. missingness.param.prior is found in the run.model function; thus, the arguments are as specified in the latter.

missingness.param.prior(assumption, mean.misspar)

Arguments

assumption
Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following: "HIE-COMMON", "HIE-TRIAL", "HIE-ARM", "IDE-COMMON", "IDE-TRIAL", "IDE-ARM", "IND-CORR", or "IND-UNCORR". The default argument is "IDE-ARM". The abbreviations "IDE", "HIE", and "IND" stand for identical, hierarchical and independent, respectively. "CORR" and "UNCORR" stand for correlated and uncorrelated, respectively.

mean.misspar
A numeric value or a vector of two numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The default argument is 0 and corresponds to the missing-at-random assumption.

Value

A value to be passed to run.model.

Details

The run.model function considers the informative missingness odds ratio in the logarithmic scale for binary outcome data (White et al., 2008; Turner et al., 2015; Spineli, 2019), the informative missingness difference of means when the effect measure is mean difference or standardised mean difference, and the informative missingness ratio of means in the logarithmic scale when the ratio of means is the effect measure (Mavridis et al., 2015; Spineli et al., 2021). When the assumption is trial-specific (i.e., "IDE-TRIAL" or "HIE-TRIAL"), or independent (i.e., "IND-CORR" or "IND-UNCORR"), only one numeric value can be assigned to mean.missingness as the same missingness scenario is applied to all trials and trial-arms of the dataset, respectively. When the assumption is "IDE-ARM" or "HIE-ARM", a maximum of two different numeric values can be assigned to mean.missingness as a vector: the first value refers to the experimental arm, and the second value refers to the control arm of each trial. In case of a network, the first value is considered for all non-reference interventions and the second value is considered for the reference intervention of the network (i.e., the intervention with identifier equal to one). This is necessary to ensure transitivity in the assumed values for the missingness parameter across the comparisons in the network (Spineli, 2019). When one numeric value is considered, the same missingness scenario is applied to all interventions in the dataset.

Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to Mavridis et al. (2015) and Spineli (2019) to determine mean.misspar for an informative missingness mechanism.

References

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. [doi: 10.1177/0962280220983544 ]

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. [doi: 10.1186/s12874-019-0731-y ]

Spineli LM. Modeling missing binary outcome data while preserving transitivity assumption yielded more credible network meta-analysis results. J Clin Epidemiol 2019;105:19--26. [doi: 10.1016/j.jclinepi.2018.09.002 ]

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--741. [doi: 10.1002/sim.6365 ]

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--2080. [doi: 10.1002/sim.6475 ]

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis—part 1: two-stage methods. Stat Med 2008;27(5):711--727. [doi: 10.1002/sim.3008 ]

See also

Author

Loukia M. Spineli