Network plot and description of the evidence base

Illustrates the network plot for one outcome and summarises the characteristics of the evidence base.

netplot(data, drug_names, save_xls, ...)

Arguments

data

A data-frame of a one-trial-per-row format containing arm-level data of each trial. See 'Format' in run_model.

drug_names

A vector of labels with the name of the interventions in the order they appear in the argument data. If drug_names is not defined, the order of the interventions as they appear in data is used, instead.

save_xls

Logical to indicate whether to export the tabulated results to an 'xlsx' file (via the write_xlsx function of the R-package writexl) at the working directory of the user. The default is FALSE (do not export).

...

Additional arguments of the nma.networkplot function of the R-package pcnetmeta.

Value

A network plot with coloured closed-loops informed by multi-arm trials. Each node indicates an intervention and each edge an observed pairwise comparison. The edge thickness is proportional to the number of trials investigating the corresponding comparison, unless specified otherwise (see nma.networkplot function of the R-package pcnetmeta). The size of the node is weighted by the total number of trials investigating the corresponding intervention, unless specified otherwise (see nma.networkplot function of the R-package pcnetmeta). netplot also returns the following data-frames that describe the evidence base:

network_description

The number of: interventions, possible comparisons, direct and indirect comparisons, number of trials in total, number of two-arm and multi-arm trials, number of randomised participants, and proportion of participants completing the trial (completers). When the outcome is binary, the number of trials with at least one zero event, and the number of trials with all zero events are also presented.

table_interventions

For each intervention, the number of trials, number of randomised participants, and proportion of completers. When the outcome is binary, the data-frame presents also the corresponding proportion of total observed events, the minimum, median and maximum proportion of observed events across the corresponding trials.

table_comparisons

Identical structure to table_interventions but for each observed comparison in the network.

Details

netplot draws the network plot using the nma.networkplot function of the R-package pcnetmeta. The mtc.data.studyrow function of the R-package hrefhttps://CRAN.R-project.org/package=gemtcgemtc is additionally used to convert data from the required one-trial-per-row format into the one-arm-per-row format.

Furthermore, netplot exports the data-frames to separate 'xlsx' files (via the write_xlsx function of the R-package writexl) at the working directory of the user.

References

Lin L, Zhang J, Hodges JS, Chu H. Performing Arm-Based Network Meta-Analysis in R with the pcnetmeta Package. J Stat Softw 2017;80(5): 1--25. doi: 10.18637/jss.v080.i05

van Valkenhoef G, Kuiper J. gemtc: Network Meta-Analysis Using Bayesian Methods. R package version 1.0-1. 2021. https://CRAN.R-project.org/package=gemtc

See also

Author

Loukia M. Spineli

Examples

data("nma.bottomley2011")

# Return the first six trials of the dataset
head(nma.bottomley2011)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("betamethasone dipropionate", "betamethasone valerate",
                  "calcipotriol", "calcipotriol plus polytar", "capasal",
                  "two-compound formulation gel", "placebo")

# Create the network plot
netplot(data = nma.bottomley2011,
        drug_names = interv_names,
        save_xls = FALSE)