Rankograms and SUCRA curves

Returns a panel of rankograms with integrated SUCRA curves for each intervention in the network.

rankosucra_plot(full, drug_names)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

drug_names
A vector of labels with the name of the interventions in the order they appear in the argument data of run_model. If the argument drug_names is not defined, the interventions are ordered as they appear in data.

Value

A panel of rankograms with integrated SUCRA curves for each intervention of the network (Salanti et al., 2011). The x-axis of each panel refers to the ranking, and the y-axis refers to the ranking probability expressed in percentage.

Details

Interventions are sorted in the descending order of their SUCRA value. The SUCRA value expressed in percentage appears on the top left corner of each panel.

rankosucra_plot can be used only for a network of interventions.

References

Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. J Clin Epidemiol 2011;64(2):163--71. [doi: 10.1016/j.jclinepi.2010.03.016 ]

See also

Author

Loukia M. Spineli, Chrysostomos Kalyvas, Katerina Papadimitropoulou

Examples

data("nma.liu2013")

if (FALSE) {
# Perform a random-effects network meta-analysis
res <- run_model(data = nma.liu2013,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 1,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "pramipexole", "serotonin-norepinephrine
                  reuptake inhibitor", "serotonin reuptake inhibitor",
                  "tricyclic antidepressant", "pergolide")

# Create the integrated rankograms and SUCRA curves
rankosucra_plot(full = res,
                drug_names = interv_names)
}