rnmamod: Bayesian Network Meta-analysis with Missing Participants

An R package for performing Bayesian network meta-analysis while handling missing participant outcome data properly.

Details

R-package rnmamod is built upon the WinBUGS program code found in the series of tutorial papers on evidence synthesis methods for decision making (Dias et al., 2013a; Dias et al., 2013b; Dias et al., 2013c) and Dias et al. (2010) that introduces the node-splitting approach. All models comprise Bayesian hierarchical models for one-stage network meta-analysis and they are implemented in JAGS through the R-package R2jags.

rnmamod comprises a suite of core models implemented in a systematic review with multiple interventions:

fixed-effect and random-effects network meta-analysis (run_model) based on Dias et al. (2013c);

fixed-effect and random-effects network meta-regression (run_metareg) based on Cooper et al. (2009), and Dias et al. (2013b);

fixed-effect and random-effects separate pairwise meta-analyses for comparisons with at least two trials (run_series_meta);

local evaluation of the consistency assumption using the fixed-effect or random-effects node-splitting approach (run_nodesplit) based on Dias et al. (2010), and van Valkenhoef et al. (2016);

global evaluation of the consistency assumption using the fixed-effect or random-effects unrelated mean effects model (run_ume) based on Dias et al. (2013a) and Spineli (2021);

comprehensive sensitivity analysis for the impact of aggregate binary and continuous missing participant outcome data (run_sensitivity) based on Spineli et al. (2021a).

rnmamod also includes a rich suite of visualisation tools to aid in the interpretation of the results and preparation of the manuscript for submission:

network plot and description of the evidence base (netplot) following the PRISMA statement for systematic reviews with network meta-analysis (Hutton et al., 2015);

tabulation of the R-hat (Gelman and Rubin, 1992) for all monitored nodes and creation of an HTML file with a panel of diagnostic plots for each monitored parameter (mcmc_diagnostics);

heatmap on the proportion of missing participants across the network (heatmap_missing_network) and across the intervention arms of each trial in the dataset (heatmap_missing_dataset);

league heatmap with the estimated and predicted summary effects of all possible pairwise comparisons in the network and integrated SUCRA or P-scores (league_heatmap and league_heatmap_pred, respectively) after performing network meta-analysis or network meta-regression (Salanti et al., 2011; Ruecker and Schwarzer, 2015);

rankograms with integrated SUCRA values for each intervention in the network (rankosucra_plot) after performing network meta-analysis (Salanti et al., 2011);

forest plot with the estimated and predicted summary effects of all comparisons with a selected intervention (forestplot) as obtained from the network meta-analysis model (Salanti et al., 2011);

tabulation of the estimated regression coefficient(s), the estimated and predicted summary effects, measures of model fit and estimated between-trial standard deviation before and after adjusting for a trial-specific covariate (metareg_plot), and visualisation of the summary effects and SUCRA values from both models (forestplot_metareg, and scatterplot_sucra, respectively--both found in metareg_plot);

tabulation of the estimated direct and indirect effects of the split nodes and corresponding inconsistency factors, measures of model fit and estimated between-trial standard deviation after each split node, and visualisation of these results (nodesplit_plot);

tabulation of the estimated summary effects of all comparisons observed in the network, measures of model fit and estimated between-trial standard deviation under the unrelated mean effects model and network meta-analysis, as well as visualisation of the summary effects from both models (intervalplot_panel_ume) and the goodness of fit of each model using a series of complementary plots (scatterplots_dev (Dias et al., 2013a), bland_altman_plot (Bland and Altman, 1999), and leverage_plot (Dias et al., 2010)--all found in ume_plot);

tabulation of the estimated summary effects and corresponding between-trial standard deviation for comparisons with at least two trials under pairwise and network meta-analysis, as well as visualisation of these results (series_meta_plot);

calculation and visualisation of the robustness index for all possible comparisons in the network (robustness_index and heatmap_robustness, respectively) (Spineli et al., 2021a);

enhanced balloon plot with the summary effects and between-trial standard deviation for a selected pairwise comparison under several scenarios about the missingness parameter (balloon_plot) (Spineli et al., 2021a);

barplot with the Kullback-Leibler divergence measure from each informative scenario to the missing-at-random assumption about the missingness parameter for a selected pairwise comparison (kld_barplot) (Spineli et al., 2021a).

Missing participant outcome data are addressed in all models of the package after extending the code to incorporate the pattern-mixture model (Spineli et al., 2021b; Spineli, 2019).

Type citation("rnmamod") on how to cite rnmamod.

To report possible bugs and errors, send an email to Loukia Spineli (Spineli.Loukia@mh-hannove.de).

The development version of rnmamod is available on GitHub under the GPL-3.0 License.

References

Bland JM, Altman DG. Measuring agreement in method comparison studies. Stat Methods Med Res 1999;8(2):135--60. doi: 10.1177/096228029900800204

Cooper NJ, Sutton AJ, Morris D, Ades AE, Welton NJ. Addressing between-study heterogeneity and inconsistency in mixed treatment comparisons: Application to stroke prevention treatments in individuals with non-rheumatic atrial fibrillation. Stat Med 2009;28(14):1861--81. doi: 10.1002/sim.3594

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013a;33(5):641--56. doi: 10.1177/0272989X12455847

Dias S, Sutton AJ, Welton NJ, Ades AE. Evidence synthesis for decision making 3: heterogeneity--subgroups, meta-regression, bias, and bias-adjustment. Med Decis Making 2013b;33(5):618--40. doi: 10.1177/0272989X13485157

Dias S, Sutton AJ, Ades AE, Welton NJ. Evidence synthesis for decision making 2: a generalized linear modeling framework for pairwise and network meta-analysis of randomized controlled trials. Med Decis Making 2013c;33(5):607--17. doi: 10.1177/0272989X12458724

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Hutton B, Salanti G, Caldwell DM, Chaimani A, Schmid CH, Cameron C, et al. The PRISMA extension statement for reporting of systematic reviews incorporating network meta-analyses of health care interventions: checklist and explanations. Ann Intern Med 2015;162(11):777--84. doi: 10.7326/M14-2385

Ruecker G, Schwarzer G. Ranking treatments in frequentist network meta-analysis works without resampling methods. BMC Med Res Methodol 2015;15:58. doi: 10.1186/s12874-015-0060-8

Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. J Clin Epidemiol 2011;64(2):163--71. doi: 10.1016/j.jclinepi.2010.03.016

Spineli LM. A revised framework to evaluate the consistency assumption globally in a network of interventions. Med Decis Making 2021. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021a;12(4):475--90. doi: 10.1002/jrsm.1478

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021b;30(4):958--75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

Author

Loukia M. Spineli