Perform the node-splitting approach

Performs the Bayesian node-splitting approach of Dias et al. (2010) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al., 2021; Spineli, 2019). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al., 2010).

Usage

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct

The summary effect measure (according to the argument measure defined in run_model) of each split node based on the corresponding trials.

indirect

The indirect summary effect measure (according to the argument measure defined in run_model) of each split node based on the remaining network after removing (splitting) the corresponding node.

diff

The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau

The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model_assessment

A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, ref, and indic from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see 'Examples').

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_nodesplit currently performs node-splitting using the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al. (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model).

run_nodesplit runs the Bayesian node-splitting approach in JAGS. The progress of the simulation appears on the R console. The number of times run_nodesplit is used appears on the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed on the R console. The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

run_nodesplit uses the mtc.nodesplit.comparisons function of the R-package gemtc to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al. (2016). run_nodesplit uses the option (1) in van Valkenhoef et al. (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, mtc.nodesplit.comparisons uses the option (3) in van Valkenhoef et al. (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function inherits the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects node-splitting approach
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 1000,
              n_burnin = 100,
              n_thin = 1)
#> JAGS generates initial values for the parameters.
#> 1 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 1 until convergence
#> 2 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 2 until convergence
#> 3 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 3 until convergence
#> 4 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2534
#> 
#> Initializing model
#> 
#> Updating model for split node 4 until convergence
#> 5 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2534
#> 
#> Initializing model
#> 
#> Updating model for split node 5 until convergence
#> 6 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2534
#> 
#> Initializing model
#> 
#> Updating model for split node 6 until convergence
#> 7 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 7 until convergence
#> 8 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2530
#> 
#> Initializing model
#> 
#> Updating model for split node 8 until convergence
#> $direct
#>   treat1 treat2       mean        sd       2.5%        97.5%     Rhat n.eff
#> 1      6      1 -0.4051748 0.2881095 -0.8277626  0.258595609 1.160710    19
#> 2      7      1 -0.7061183 0.2048064 -1.1062416 -0.237698997 1.321727    11
#> 3      8      1 -0.3939128 0.1868743 -0.7637187 -0.022303677 1.147058    20
#> 4      5      4 -0.2541271 0.2292428 -0.7155505  0.166248247 1.008492   250
#> 5      7      4 -0.5127616 0.2354707 -0.9668011 -0.048588736 1.072882    34
#> 6      7      5 -0.4162774 0.2267083 -0.8336832 -0.005106965 1.055457    42
#> 7      8      6 -1.6887592 0.7882375 -3.1482169 -0.085437373 1.060952    50
#> 8      8      7 -0.8245797 0.3383449 -1.4029114 -0.072943391 1.010965   200
#> 
#> $indirect
#>   treat1 treat2        mean        sd       2.5%       97.5%     Rhat n.eff
#> 1      6      1  1.21307590 0.7227542 -0.2949785  2.71312222 1.023088   330
#> 2      7      1 -0.38600154 0.6087917 -1.3687411  1.04931900 1.389008    10
#> 3      8      1 -0.98310736 0.5676435 -2.2247075 -0.13040627 1.020644   140
#> 4      5      4 -0.83814233 0.4776175 -1.8849520  0.06428168 1.059472    49
#> 5      7      4 -0.23434623 0.1650901 -0.5549826  0.09367799 1.051047    45
#> 6      7      5 -0.33552809 0.3022950 -0.9328942  0.25532323 1.100729    26
#> 7      8      6 -0.08077327 0.2073334 -0.4833248  0.32036144 1.033531    76
#> 8      8      7 -0.01011442 0.1964527 -0.4019750  0.33934927 1.008633   620
#> 
#> $diff
#>   treat1 treat2        mean        sd       2.5%       97.5%     Rhat n.eff
#> 1      6      1 -1.61825068 0.7908899 -3.1139536 0.039242813 1.032931    87
#> 2      7      1 -0.32011679 0.6359012 -1.7704306 0.636843943 1.332284    11
#> 3      8      1  0.58919452 0.5194541 -0.2623170 1.722718020 1.011117   780
#> 4      5      4  0.58401518 0.5082450 -0.4347414 1.635375554 1.026496    89
#> 5      7      4 -0.27841539 0.2667468 -0.8189545 0.247394578 1.007975   290
#> 6      7      5 -0.08074927 0.4400514 -0.9174754 0.791481971 1.086301    31
#> 7      8      6 -1.60798596 0.8002984 -3.1029704 0.001399872 1.078260    36
#> 8      8      7 -0.81446533 0.4193489 -1.5318569 0.081104474 1.009178   270
#> 
#> $tau
#>   treat1 treat2        50%         sd         2.5%     97.5%     Rhat n.eff
#> 1      6      1 0.12850034 0.10146983 0.0026951306 0.3644712 1.332412    12
#> 2      7      1 0.09440514 0.10086801 0.0005205508 0.3576373 1.867228     6
#> 3      8      1 0.12137280 0.07845900 0.0150229895 0.2922614 1.042283    81
#> 4      5      4 0.14543350 0.09585586 0.0036653203 0.3996323 1.183107    54
#> 5      7      4 0.09376025 0.09052217 0.0030909292 0.3351126 1.131206    25
#> 6      7      5 0.14199747 0.08247375 0.0382641931 0.3416285 1.032216    73
#> 7      8      6 0.12591997 0.08338546 0.0184798181 0.3351017 1.199441    16
#> 8      8      7 0.14001323 0.09538700 0.0339289532 0.3794030 1.012826   200
#> 
#> $model
#> [1] "RE"
#> 
#> $model_assessment
#>   treat1 treat2 deviance      DIC       pD
#> 1      6      1 54.91672 91.76099 36.84427
#> 2      7      1 55.27202 91.58647 36.31445
#> 3      8      1 53.74910 89.14903 35.39993
#> 4      5      4 53.95318 91.98579 38.03261
#> 5      7      4 53.60225 89.80769 36.20544
#> 6      7      5 54.57756 92.68134 38.10378
#> 7      8      6 53.52367 89.48341 35.95974
#> 8      8      7 53.65634 90.56977 36.91344
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> attr(,"class")
#> [1] "run_nodesplit"
# }