Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al. (2013) that has been refined (Spineli, 2021) and extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al. 2021; Spineli, 2019). This model offers a global evaluation of the plausibility of the consistency assumption in the network.

Usage

run_ume(full, n_iter, n_burnin, n_chains, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM

The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o

The deviance contribution of each trial-arm based on the observed outcome.

hat_par

The fitted outcome at each trial-arm.

tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o

The leverage for the observed outcome at each trial-arm.

sign_dev_o

The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment

A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit

An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in

ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then use run_ume (see 'Examples').

The run_ume function also returns the arguments data, model, measure, assumption, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm or multi-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_ume currently considers the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has also been refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the underlying trial-specific effect size of comparisons with the baseline arm of the multi-arm trial (Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console. The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013;33(5):641--56. doi: 10.1177/0272989X12455847

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Spineli LM. A Revised Framework to Evaluate the Consistency Assumption Globally in a Network of Interventions. Med Decis Making 2021. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> JAGS generates initial values for the parameters.
#> Running the model ...
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 1291
#> 
#> Initializing model
#> 
#> ... Updating the model until convergence
#> $EM
#>               mean        sd       2.5%        25%        50%         75%
#> EM[2,1]  0.5966885 0.8342222 -1.0994215  0.1569220  0.5945228  1.05122744
#> EM[3,1]  0.2304546 0.7113043 -1.2355160 -0.1763918  0.2247365  0.65396064
#> EM[4,1]  0.8338454 0.5659844 -0.3553165  0.4926443  0.8679732  1.22794589
#> EM[5,1]  1.9537071 0.7891748  0.4253632  1.4097994  1.9601821  2.50630471
#> EM[4,2] -0.4746563 1.0085621 -2.5602172 -1.0607855 -0.4176659  0.09031537
#> EM[6,2] -1.2885208 1.1526491 -3.6531767 -1.9960014 -1.2934126 -0.57287953
#> EM[4,3]  1.1880354 0.9397013 -0.7466540  0.6703262  1.2081067  1.70459760
#> EM[5,4] -0.7504635 1.4555264 -3.9071829 -1.7972332 -0.5663716  0.25938551
#>            97.5%     Rhat n.eff
#> EM[2,1] 2.297911 1.005960  3000
#> EM[3,1] 1.619610 1.007649   460
#> EM[4,1] 1.860852 1.036058    61
#> EM[5,1] 3.398116 1.020633   120
#> EM[4,2] 1.517655 1.002253  1100
#> EM[6,2] 1.015504 1.009320   340
#> EM[4,3] 3.077021 1.011269   470
#> EM[5,4] 1.756041 1.037026    74
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  0.976959727 1.33873878 0.0007010644 0.09913790 0.4634169
#> dev.o[2,1]  0.841933929 1.17895429 0.0008668805 0.08808460 0.4034739
#> dev.o[3,1]  1.124782071 1.49645439 0.0010866462 0.11264729 0.5046061
#> dev.o[4,1]  1.005666777 1.41520570 0.0010088342 0.09488830 0.4518879
#> dev.o[5,1]  0.958727227 1.40689517 0.0009685106 0.10123914 0.4434823
#> dev.o[6,1]  1.116963315 1.49075461 0.0011118169 0.12954707 0.5418241
#> dev.o[7,1]  1.060789276 1.39478494 0.0011421647 0.12219928 0.5089456
#> dev.o[8,1]  0.004549947 0.04513835 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.825071794 1.14343064 0.0005893364 0.08827685 0.3972898
#> dev.o[10,1] 1.091084148 1.38624941 0.0014840765 0.12967395 0.5448985
#> dev.o[11,1] 0.743251787 1.02815941 0.0005762650 0.07156105 0.3549021
#> dev.o[1,2]  1.016819538 1.44417075 0.0010147430 0.11182901 0.4702504
#> dev.o[2,2]  1.125121403 1.46434446 0.0011694425 0.12458716 0.5533803
#> dev.o[3,2]  1.140564021 1.58545175 0.0011604765 0.12366243 0.5245611
#> dev.o[4,2]  0.985637634 1.40007442 0.0008816534 0.10077529 0.4321237
#> dev.o[5,2]  1.042813817 1.50407412 0.0009334041 0.10488314 0.4453120
#> dev.o[6,2]  1.331656521 1.60781417 0.0017094019 0.19729038 0.7591602
#> dev.o[7,2]  0.836841391 1.21279298 0.0007982162 0.08123854 0.3669786
#> dev.o[8,2]  0.010579408 0.13177468 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.065315488 1.49604865 0.0011738714 0.10958637 0.4741480
#> dev.o[10,2] 1.384706597 1.69192787 0.0017876939 0.17855494 0.7383296
#> dev.o[11,2] 1.109070618 1.55812964 0.0009797543 0.11765442 0.5144245
#> dev.o[9,3]  1.113379811 1.54224879 0.0013741441 0.11626973 0.5278750
#> dev.o[10,3] 0.898685123 1.25842976 0.0008391889 0.08826259 0.4245276
#> dev.o[11,3] 0.947384388 1.32933954 0.0014770029 0.09935093 0.4077821
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.318853e+00 4.87956980 1.001971  1400
#> dev.o[2,1]  1.140005e+00 4.16941155 1.002615   950
#> dev.o[3,1]  1.546504e+00 5.41738612 1.004698   630
#> dev.o[4,1]  1.346898e+00 5.02815344 1.001076  3000
#> dev.o[5,1]  1.252319e+00 4.78899808 1.001296  3000
#> dev.o[6,1]  1.505215e+00 5.43816063 1.001776  1800
#> dev.o[7,1]  1.467946e+00 4.91971813 1.002086  1300
#> dev.o[8,1]  3.197442e-14 0.01194100 1.001845  3000
#> dev.o[9,1]  1.093501e+00 3.97374188 1.000551  3000
#> dev.o[10,1] 1.556344e+00 4.83462541 1.003914   590
#> dev.o[11,1] 9.869966e-01 3.83590301 1.012787   160
#> dev.o[1,2]  1.380173e+00 4.86271731 1.001928  1400
#> dev.o[2,2]  1.563192e+00 5.23160029 1.001094  3000
#> dev.o[3,2]  1.519281e+00 5.61726807 1.001422  2200
#> dev.o[4,2]  1.296707e+00 5.10096096 1.000798  3000
#> dev.o[5,2]  1.381207e+00 5.40851926 1.001906  1400
#> dev.o[6,2]  1.880094e+00 5.85242042 1.015107   170
#> dev.o[7,2]  1.055929e+00 4.56899040 1.007833  3000
#> dev.o[8,2]  6.661338e-14 0.02296173 1.034488  3000
#> dev.o[9,2]  1.399660e+00 5.20042788 1.001800  1500
#> dev.o[10,2] 1.968092e+00 6.12868583 1.001425  2200
#> dev.o[11,2] 1.424576e+00 5.62691224 1.006794   480
#> dev.o[9,3]  1.489340e+00 5.42249494 1.001592  1800
#> dev.o[10,3] 1.168885e+00 4.58814855 1.003838   600
#> dev.o[11,3] 1.241395e+00 4.63090286 1.005051   800
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  26.820064 4.55336501 18.8731812 23.612983 26.566530 29.837952
#> hat.par[2,1]   2.778605 1.28778624  0.7909060  1.823799  2.606373  3.530044
#> hat.par[3,1]   9.905309 1.27641179  6.9938513  9.110023 10.094598 10.895403
#> hat.par[4,1]  23.039742 2.47745988 17.8746581 21.388247 23.173350 24.778097
#> hat.par[5,1]   7.901707 2.01659604  4.0967519  6.503109  7.869749  9.294705
#> hat.par[6,1]   3.232346 0.96748997  1.3258808  2.555353  3.241378  3.949236
#> hat.par[7,1]   2.489779 1.27812835  0.6526770  1.535982  2.282450  3.213850
#> hat.par[8,1]  11.997746 0.02227675 11.9940310 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.232906 2.54639423 10.3444768 13.474596 15.253362 17.000291
#> hat.par[10,1]  3.214069 1.23926464  1.2309379  2.306492  3.087075  3.963881
#> hat.par[11,1]  4.548701 1.45048150  1.9744118  3.500671  4.425689  5.533585
#> hat.par[1,2]  38.173715 5.15021812 28.4586396 34.623784 38.101457 41.683619
#> hat.par[2,2]   4.212494 1.63635657  1.5948183  2.997128  4.042273  5.261739
#> hat.par[3,2]   8.103625 1.44554618  4.8835116  7.199660  8.304100  9.213096
#> hat.par[4,2]  15.989873 2.46689424 10.9990516 14.359363 15.955599 17.711360
#> hat.par[5,2]   3.030003 1.54184540  0.7156971  1.888900  2.778331  3.884277
#> hat.par[6,2]   3.842171 0.92797592  1.8362456  3.259371  3.960572  4.527742
#> hat.par[7,2]   2.544306 1.23481886  0.6562592  1.646182  2.385550  3.257073
#> hat.par[8,2]  14.994852 0.06300360 14.9885235 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.964116 2.79172269 11.3058989 15.110626 17.025774 18.825615
#> hat.par[10,2]  3.201172 1.32122247  0.9021838  2.228609  3.107864  4.067031
#> hat.par[11,2]  6.948819 1.52163347  3.8777874  5.937579  6.962234  8.043357
#> hat.par[9,3]  21.697641 2.27074588 17.0170797 20.236587 21.820706 23.269812
#> hat.par[10,3]  8.422304 1.37624237  5.4933230  7.500214  8.538103  9.427714
#> hat.par[11,3] 11.478078 1.57399730  8.0895584 10.457689 11.609590 12.598347
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.618639 1.001471  2100
#> hat.par[2,1]   5.746459 1.007539   320
#> hat.par[3,1]  11.744294 1.010036   250
#> hat.par[4,1]  27.521467 1.006613   330
#> hat.par[5,1]  11.834576 1.003507   740
#> hat.par[6,1]   5.030862 1.005376   410
#> hat.par[7,1]   5.549282 1.008013   270
#> hat.par[8,1]  12.000000 1.001845  3000
#> hat.par[9,1]  20.069941 1.014195   160
#> hat.par[10,1]  6.050436 1.010975   190
#> hat.par[11,1]  7.564720 1.019757   210
#> hat.par[1,2]  48.417919 1.003213   840
#> hat.par[2,2]   7.839870 1.002915   830
#> hat.par[3,2]  10.271553 1.023865   160
#> hat.par[4,2]  20.685402 1.002789   880
#> hat.par[5,2]   6.558441 1.010520   230
#> hat.par[6,2]   5.357775 1.018950   120
#> hat.par[7,2]   5.540656 1.012838   170
#> hat.par[8,2]  15.000000 1.034488  3000
#> hat.par[9,2]  22.177987 1.004999   710
#> hat.par[10,2]  5.980506 1.008472   360
#> hat.par[11,2]  9.790050 1.007255  1200
#> hat.par[9,3]  25.806624 1.008809   240
#> hat.par[10,3] 10.759889 1.005342   480
#> hat.par[11,3] 14.121672 1.008994   240
#> 
#> $leverage_o
#>  [1]  0.975450136  0.567125228  1.119533253  1.005422176  0.956636444
#>  [6]  0.711685418  0.944566490 -0.009762510  0.818452578  0.830746409
#> [11]  0.648347096  1.015676460  0.950911340  1.135569276  0.985621904
#> [16]  1.042451356  0.839150080  0.683216177  0.007003067  0.955892276
#> [21]  0.687665838  0.763309204  0.820822387  0.761184884  0.878337499
#> 
#> $sign_dev_o
#>  [1]  1 -1  1 -1  1  1  1 -1 -1  1 -1 -1  1 -1  1 -1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>         mean           sd         2.5%          25%          50%          75% 
#> 6.901601e-01 4.262690e-01 3.900282e-02 3.799128e-01 6.207667e-01 9.336426e-01 
#>        97.5%         Rhat        n.eff 
#> 1.762270e+00 1.053054e+00 2.400000e+03 
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.67522107  0.51017458  0.01510335  0.26055755  0.57117256  0.97597289 
#>       97.5%        Rhat       n.eff 
#>  1.86061300  1.14522129 40.00000000 
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 43.85337 20.09501 23.75836
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "9", fit using jags,
#>  3 chains, each with 1000 iterations (first 0 discarded)
#>  n.sims = 3000 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.597   0.834 -1.099  0.157  0.595  1.051  2.298 1.006  3000
#> EM[3,1]         0.230   0.711 -1.236 -0.176  0.225  0.654  1.620 1.008   460
#> EM[4,1]         0.834   0.566 -0.355  0.493  0.868  1.228  1.861 1.036    61
#> EM[5,1]         1.954   0.789  0.425  1.410  1.960  2.506  3.398 1.021   120
#> EM[4,2]        -0.475   1.009 -2.560 -1.061 -0.418  0.090  1.518 1.002  1100
#> EM[6,2]        -1.289   1.153 -3.653 -1.996 -1.293 -0.573  1.016 1.009   340
#> EM[4,3]         1.188   0.940 -0.747  0.670  1.208  1.705  3.077 1.011   470
#> EM[5,4]        -0.750   1.456 -3.907 -1.797 -0.566  0.259  1.756 1.037    74
#> dev.o[1,1]      0.977   1.339  0.001  0.099  0.463  1.319  4.880 1.002  1400
#> dev.o[2,1]      0.842   1.179  0.001  0.088  0.403  1.140  4.169 1.003   950
#> dev.o[3,1]      1.125   1.496  0.001  0.113  0.505  1.547  5.417 1.005   630
#> dev.o[4,1]      1.006   1.415  0.001  0.095  0.452  1.347  5.028 1.001  3000
#> dev.o[5,1]      0.959   1.407  0.001  0.101  0.443  1.252  4.789 1.001  3000
#> dev.o[6,1]      1.117   1.491  0.001  0.130  0.542  1.505  5.438 1.002  1800
#> dev.o[7,1]      1.061   1.395  0.001  0.122  0.509  1.468  4.920 1.002  1300
#> dev.o[8,1]      0.005   0.045  0.000  0.000  0.000  0.000  0.012 1.002  3000
#> dev.o[9,1]      0.825   1.143  0.001  0.088  0.397  1.094  3.974 1.001  3000
#> dev.o[10,1]     1.091   1.386  0.001  0.130  0.545  1.556  4.835 1.004   590
#> dev.o[11,1]     0.743   1.028  0.001  0.072  0.355  0.987  3.836 1.013   160
#> dev.o[1,2]      1.017   1.444  0.001  0.112  0.470  1.380  4.863 1.002  1400
#> dev.o[2,2]      1.125   1.464  0.001  0.125  0.553  1.563  5.232 1.001  3000
#> dev.o[3,2]      1.141   1.585  0.001  0.124  0.525  1.519  5.617 1.001  2200
#> dev.o[4,2]      0.986   1.400  0.001  0.101  0.432  1.297  5.101 1.001  3000
#> dev.o[5,2]      1.043   1.504  0.001  0.105  0.445  1.381  5.409 1.002  1400
#> dev.o[6,2]      1.332   1.608  0.002  0.197  0.759  1.880  5.852 1.015   170
#> dev.o[7,2]      0.837   1.213  0.001  0.081  0.367  1.056  4.569 1.008  3000
#> dev.o[8,2]      0.011   0.132  0.000  0.000  0.000  0.000  0.023 1.034  3000
#> dev.o[9,2]      1.065   1.496  0.001  0.110  0.474  1.400  5.200 1.002  1500
#> dev.o[10,2]     1.385   1.692  0.002  0.179  0.738  1.968  6.129 1.001  2200
#> dev.o[11,2]     1.109   1.558  0.001  0.118  0.514  1.425  5.627 1.007   480
#> dev.o[9,3]      1.113   1.542  0.001  0.116  0.528  1.489  5.422 1.002  1800
#> dev.o[10,3]     0.899   1.258  0.001  0.088  0.425  1.169  4.588 1.004   600
#> dev.o[11,3]     0.947   1.329  0.001  0.099  0.408  1.241  4.631 1.005   800
#> hat.par[1,1]   26.820   4.553 18.873 23.613 26.567 29.838 36.619 1.001  2100
#> hat.par[2,1]    2.779   1.288  0.791  1.824  2.606  3.530  5.746 1.008   320
#> hat.par[3,1]    9.905   1.276  6.994  9.110 10.095 10.895 11.744 1.010   250
#> hat.par[4,1]   23.040   2.477 17.875 21.388 23.173 24.778 27.521 1.007   330
#> hat.par[5,1]    7.902   2.017  4.097  6.503  7.870  9.295 11.835 1.004   740
#> hat.par[6,1]    3.232   0.967  1.326  2.555  3.241  3.949  5.031 1.005   410
#> hat.par[7,1]    2.490   1.278  0.653  1.536  2.282  3.214  5.549 1.008   270
#> hat.par[8,1]   11.998   0.022 11.994 12.000 12.000 12.000 12.000 1.002  3000
#> hat.par[9,1]   15.233   2.546 10.344 13.475 15.253 17.000 20.070 1.014   160
#> hat.par[10,1]   3.214   1.239  1.231  2.306  3.087  3.964  6.050 1.011   190
#> hat.par[11,1]   4.549   1.450  1.974  3.501  4.426  5.534  7.565 1.020   210
#> hat.par[1,2]   38.174   5.150 28.459 34.624 38.101 41.684 48.418 1.003   840
#> hat.par[2,2]    4.212   1.636  1.595  2.997  4.042  5.262  7.840 1.003   830
#> hat.par[3,2]    8.104   1.446  4.884  7.200  8.304  9.213 10.272 1.024   160
#> hat.par[4,2]   15.990   2.467 10.999 14.359 15.956 17.711 20.685 1.003   880
#> hat.par[5,2]    3.030   1.542  0.716  1.889  2.778  3.884  6.558 1.011   230
#> hat.par[6,2]    3.842   0.928  1.836  3.259  3.961  4.528  5.358 1.019   120
#> hat.par[7,2]    2.544   1.235  0.656  1.646  2.386  3.257  5.541 1.013   170
#> hat.par[8,2]   14.995   0.063 14.989 15.000 15.000 15.000 15.000 1.034  3000
#> hat.par[9,2]   16.964   2.792 11.306 15.111 17.026 18.826 22.178 1.005   710
#> hat.par[10,2]   3.201   1.321  0.902  2.229  3.108  4.067  5.981 1.008   360
#> hat.par[11,2]   6.949   1.522  3.878  5.938  6.962  8.043  9.790 1.007  1200
#> hat.par[9,3]   21.698   2.271 17.017 20.237 21.821 23.270 25.807 1.009   240
#> hat.par[10,3]   8.422   1.376  5.493  7.500  8.538  9.428 10.760 1.005   480
#> hat.par[11,3]  11.478   1.574  8.090 10.458 11.610 12.598 14.122 1.009   240
#> m.tau           0.675   0.510  0.015  0.261  0.571  0.976  1.861 1.145    40
#> resdev.o[1]     1.994   1.972  0.051  0.572  1.418  2.740  7.311 1.006   420
#> resdev.o[2]     1.967   1.817  0.064  0.653  1.450  2.734  6.649 1.007   510
#> resdev.o[3]     2.265   2.150  0.064  0.701  1.619  3.191  7.830 1.008   290
#> resdev.o[4]     1.991   2.020  0.040  0.535  1.347  2.790  7.294 1.001  3000
#> resdev.o[5]     2.002   2.077  0.046  0.534  1.362  2.777  7.568 1.002  1200
#> resdev.o[6]     2.449   1.963  0.119  1.042  2.004  3.291  7.565 1.027    97
#> resdev.o[7]     1.898   1.834  0.062  0.582  1.352  2.564  6.672 1.005   590
#> resdev.o[8]     0.015   0.160  0.000  0.000  0.000  0.000  0.047 1.017  3000
#> resdev.o[9]     3.004   2.394  0.225  1.231  2.451  4.027  9.542 1.004   750
#> resdev.o[10]    3.374   2.334  0.290  1.621  2.936  4.624  9.379 1.002  1100
#> resdev.o[11]    2.800   2.304  0.209  1.136  2.182  3.787  8.840 1.004   560
#> tau             0.690   0.426  0.039  0.380  0.621  0.934  1.762 1.053  2400
#> totresdev.o    23.758   6.692 12.159 19.083 23.232 27.772 38.204 1.007   300
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.67522107  0.51017458  0.01510335  0.26055755  0.57117256  0.97597289 
#>       97.5%        Rhat       n.eff 
#>  1.86061300  1.14522129 40.00000000 
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
#> attr(,"class")
#> [1] "run_ume"
# }