End-user-ready results for series of pairwise meta-analyses

Facilitates comparing the estimated effects and between-trial standard deviation obtained from corresponding pairwise meta-analyses with those from a network meta-analysis.

series_meta_plot(full, meta, drug_names, save_xls)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

meta
An object of S3 class run_series_meta. See 'Value' in run_series_meta.

drug_names
A vector of labels with the name of the interventions in the order they appear in the argument data of run_model. If drug_names is not defined, the order of the interventions as they appear in data is used, instead.

save_xls
Logical to indicate whether to export the tabulated results to an 'xlsx' file at the working directory of the user. The default is FALSE (do not export).

Value

series_meta_plot returns a panel of two forest plots: (1) a forest plot on the posterior mean and 95% credible interval of the effect size of the observed comparisons from network meta-analysis and the corresponding pairwise meta-analyses, and (2) a forest plot on the posterior median and 95% credible interval of the between-trial standard deviation (\(\tau\)) for the observed comparisons. The estimated \(\tau\) from network meta-analysis appears as a rectangle in the forest plot. When a fixed-effect model has been fitted, only the forest plot on the estimated effects is shown.

The user can detect any inconsistencies in the estimated effects from the compared models and explore the gains in precision stemming from network meta-analysis. Furthermore, the user can investigate the plausibility of the common between-trial heterogeneity assumption which is typically considered in network meta-analysis.

The R console prints the data-frame with the estimated effects and \(\tau\)s of the observed comparisons under both models. In the case of a fixed-effect model, the data-frame is printed without results on \(\tau\). Furthermore, series_meta_plot the data-frame to an 'xlsx' file at the working directory of the user.

Details

series_meta_plot can be used only for a network of interventions. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

See also

Author

Loukia M. Spineli

Examples

data("nma.dogliotti2014")

if (FALSE) {
# Perform a random-effects network meta-analysis
res <- run_model(data = nma.dogliotti2014,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 1,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# Run separate random-effects pairwise meta-analyses
meta <- run_series_meta(full = res,
                        n_chains = 3,
                        n_iter = 10000,
                        n_burnin = 1000,
                        n_thin = 1)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "aspirin", "aspirin plus clopidogrel",
                  "dabigatran 110 mg", "dabigatran 150 mg", "rivaroxaban",
                  "vitamin K antagonist", "apixaban")

# Plot the results from both models
series_meta_plot(full = res,
                 meta = meta,
                 drug_names = interv_names)
}