Rating the quality of transitivity evaluation

Classifies a systematic review with multiple interventions as having low, unclear or high quality regarding the transitivity evaluation.

Usage

trans_quality(
  plan_protocol,
  plan_methods,
  report_results,
  discuss_trans,
  proper_table
)

Arguments

plan_protocol

Character string that indicates how the systematic review planned in the protocol to assess the transitivity assumption. The following values can be considered: "Both", "No", "No protocol", "Only direct methods", and "Only indirect methods". "Both" implies that at least one direct and one indirect method has been used.

plan_methods

Character string that indicates whether the systematic review described in the methods section a strategy to assess the transitivity assumption. The following values can be considered: "Yes", and "No".

report_results

Character string that indicates whether the systematic review reported in the results section the transitivity evaluation and which strategy was employed. The following values can be considered: "Both", "No", "Only direct methods", and "Only indirect methods". "Both" implies that at least one direct and one indirect method has been used.

discuss_trans

Character string that indicates whether the systematic review discussed the transitivity assumption and which model parameters where considered. The following values can be considered: "Both", "No", "Only treatment effects", "Other parameter", and "NMA not conducted". "Both" implies that the treatment effect and other model parameters were considered.

proper_table

Character string that indicates whether the systematic review reported a proper table of characteristics. The following values can be considered: "No", "No table", and "Yes".

Value

A character with value "Low", "Unclear", or

"High" to indicate low, unclear, or high-quality of transitivity evaluation.

Details

A systematic review with "Low" quality of transitivity evaluation does not provide a protocol, nor describe the evaluation strategy in the methods section, does not report the evaluation results, nor discusses the transitivity evaluation and does not provide a table of characteristics. On the contrary, a systematic review with "High" quality of transitivity evaluation provides an evaluation plan in the protocol (including at least one direct method), describes the evaluation strategy in the methods section (including at least one direct method), reports the evaluation results in the results section, discusses the transitivity evaluation while considering at least one model parameter (if NMA has been conducted), and provides a proper table of characteristics. Otherwise, the systematic review is judged to have an "Unclear" quality of transitivity evaluation.

References

Spineli LM, Kalyvas C, Seide SE, Papadimitropoulou K. Low awareness of the transitivity assumption in complex networks of interventions: empirical evidence from 356 network meta-analyses. 2022 submitted

Author

Loukia M. Spineli