Unrelated trials effects model: a panel of interval plots

Perform the unrelated trial effects model and create a panel of interval plots on the results of each trial and corresponding pairwise comparison.

unrelated.effects.plot(
  data,
  measure,
  char,
  drug.names,
  trial.names,
  mean.value,
  var.value,
  rho,
  save.xls
)

Arguments

data
A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models. See 'Format' in run.model function for the specification of the columns.

measure
Character string indicating the effect measure with values "OR", "MD", "SMD", or "ROM" for the odds ratio, mean difference, standardised mean difference and ratio of means, respectively.

char
A data-frame of three columns and number of rows equal to the number of trials in data. Each column refers to a trial-characteristic with nominal elements.

drug.names
A vector of labels with the name of the interventions in the order they appear in the argument data. If drug.names is not defined, the order of the interventions as they appear in data is used, instead.

trial.names
A vector of labels with the name of the trials in the order they appear in the argument data. If the argument drug.names is not defined, the order of the interventions as they appear in data is used, instead.

mean.value
A numeric value for the mean of the normal distribution of the informative missingness parameter. The same value is considered for all trial-arms of the dataset. The default argument is 0 and corresponds to the missing-at-random assumption. For the informative missingness odds ratio and the informative missingness ratio of means, the mean value is defined in the logarithmic scale.

var.value
A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure is "OR", "MD", or "SMD" the default argument is 1; When the measure is "ROM" the default argument is 0.04

rho
A numeric value in the interval [-1, 1] that indicates the correlation coefficient between two missingness parameters in a trial. The same value is considered across all trials of the dataset. The default argument is 0 and corresponds to uncorrelated missingness parameters.

save.xls
Logical to indicate whether to export the tabulated results to an Excel 'xlsx' format (via the write_xlsx function) to the working directory of the user. The default is FALSE (do not export to an Excel format).

Value

A panel of interval plots for each observed comparison in the network, when there are up to 21 trials in the data. Otherwise, an Excel file with the arm-level data of each trial, the corresponding effect measure and 95% confidence interval (lower, upper), the intervention compared, and the three characteristics (as defined in char). For larger datasets, the plot becomes cluttered and it is difficult to identify the trial names. Hence, exporting the results in an Excel file is an viable alternative.

Details

The unrelated trial effects model may be an alternative to network meta-analysis, when the latter is not deemed appropriate (e.g., considerable statistical heterogeneity, or substantial intransitivity). In the presence of missing participant outcome data (MOD), the effect size and standard error are adjusted by applying the pattern-mixture model with Taylor series in trial-arms with MOD (White et al., 2008; Mavridis et al., 2015). The unrelated.effects.plot function calls the Taylor.IMOR and Taylor.IMDoM.IMRoM functions (for binary and continuous outcome, respectively) to employ pattern-mixture model with Taylor series. The unrelated.effects.plot function considers the informative missingness odds ratio in the logarithmic scale for binary outcome data (White et al., 2008), the informative missingness difference of means when measure is "MD" or "SMD", and the informative missingness ratio of means in the logarithmic scale when "ROM" is the effect measure (Mavridis et al., 2015).

The number of interval plots equals the number of observed comparisons in the network. In each interval plot, the y-axis refers to all trials of the network and x-axis refers to the selected effect measure. The odds ratio and ratio of means are calculated in the logarithmic scale but they are reported in their original scale.

unrelated.effects.plot depicts all three characteristics for each trial using different colours, line-types and point-shapes for the corresponding interval and point estimate. Ideally, each characterstic should have no more than three categories; otherwise, the plot becomes cluttered. For now, the unrelated.effects.plot function uses the default color palette, line-types and point-shapes.

References

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--41. [doi: 10.1002/sim.6365 ]

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--27. [doi: 10.1002/sim.3008 ]

See also

Author

Loukia M. Spineli