Preparing the study weights based on Gower's dissimilarities (Transitivity evaluation)

Using the Gower's dissimilarities to prepare the study weights in a network of interventions to be used by the run_model function in the context of the transitivity evaluation as a sensitivity analysis.

Usage

weight_defined(diss_res)

Arguments

diss_res

A list of two elements with the following order: 1) an object of S3 class 'comp_clustering' (comp_clustering), and 2) a character with values "uniform", "fixed", or "index" to define the usage of the similarities (see 'Details') as being sampled from a uniform distribution, between-comparisons similarity (fixed weight) or a ratio of their between-comparisons similarity and total similarity (percentage index). See 'Details'.

Value

A list of the following two elements:

weights

A vector of study weights if "fixed" or "index" has been specified or a matrix with the minimum and maximum similarity values for each study if "uniform" has been specified.

type

A character indicating the weight type considered: "uniform", "fixed", or "index".

Details

The function receives the matrix Trials_diss_table found in the results of comp_clustering. This matrix contains the Gower's dissimilarities of all study pairs in the network for a specific set of clinical and methodological characteristics that act as effect modifiers. Gower's dissimilarities take values from 0 to 1, with 0 and 1 implying perfect similarity and perfect dissimilarity, respectively. Hence, subtracting each value from 1 yields the similarities of all study pairs in the network.

For each study, the Gower's similarities can be either transformed into a fixed value ("fixed") or percentage index ("index") based on the root mean square of the similarities or restricted to the minimum and maximum similarity values ("uniform"). This depends on the second element specified in the argument diss_res. When the "uniform" choice has been specified in the second element of the argument diss_res, the function checks whether the bounds are the same value. If this is the case for at least one study, the uniform distribution cannot be defined and, hence, the function returns the between-comparisons similarity values ("fixed") for each study.

weight_defined can be used in run_model via the argument adjust_wgt. See 'Details' in run_model.

References

Gower J. General Coefficient of Similarity and Some of Its Properties. Biometrics 1971;27(4):857–71. doi: 10.2307/2528823

See also

Author

Loukia M. Spineli