Comparator-specific forest-plot for network meta-regression

This function illustrates a forest plot of the posterior mean and 95% credible and predictive interval of comparisons with the selected intervention of the network under the network meta-analysis and network meta-regression.

forestplot_metareg(full, reg, compar, cov_value, drug_names)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

reg
An object of S3 class run_metareg. See 'Value' in run_metareg.

compar
A character to indicate the comparator intervention. It must be any name found in drug_names.

cov_value
A vector of two elements in the following order: a number for the covariate value of interest (see 'Arguments' in run_metareg), and a character to indicate the name of the covariate. See also 'Details'.

drug_names
A vector of labels with the name of the interventions in the order they appear in the argument data of run_model. If drug_names is not defined, the order of the interventions as they appear in data is used, instead.

Value

A panel of two forest plots: (1) a forest plot on the estimated effect size of comparisons with the selected intervention of the network, and (2) a forest plot on the predicted effect size of comparisons with the selected intervention of the network. Both forest plots illustrate the results from network meta-analysis and network meta-regression using different colours for the corresponding lines.

Details

In both plots, the y-axis displays all interventions in the network; the selected intervention that comprises the compar is indicated in the plot with a homonymous label. The numerical results are displayed above each line. Odds ratio and ratio of means are reported in the original scale after exponentiation of the logarithmic scale.

When the covariate is binary, specify in the second element of cov_value the name of the level for which the forest plot will be created.

In both plots, the interventions are sorted in the descending order of their SUCRA values based on the network meta-analysis.

forestplot can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. J Clin Epidemiol 2011;64(2):163--71. doi: 10.1016/j.jclinepi.2010.03.016

See also

Author

Loukia M. Spineli