Markov Chain Monte Carlo diagnostics

Evaluates whether convergence has been achieved for the monitored parameters of the Bayesian models. The Gelman-Rubin convergence diagnostic and relevant diagnostic plots are applied.

mcmc_diagnostics(net, par = NULL)

Arguments

net
An object of S3 class run_model, run_series_meta, run_nodesplit, run_ume, run_sensitivity and run_metareg. See 'Value' in the functions above.

par
A vector of three character strings that refer to three monitored parameters in jagsfit which is an object of S3 class run_model, run_ume and run_metareg. These three selected parameters will be considered in the diagnostic plots (see 'Value'). This argument will be ignored for objects of S3 class run_series_meta, run_nodesplit, or run_sensitivity.

Value

mcmc_diagnostics returns a data-frame that contains the Gelman-Rubin convergence diagnostic, R-hat, and convergence status of the following monitored parameters:

EM	The estimated summary effect measure.

EM_pred	The predicted summary effect measure.

delta	The estimated trial-specific effect measure.

tau	The between-trial standard deviation.

direct	The direct estimate of the split node (see 'Value' in run_nodesplit).

indirect	The indirect estimate of the split node (see 'Value' in run_nodesplit).

IF	The inconsistency factor of the split node (see 'Value' in run_nodesplit).

phi	The informative missingness parameter.

beta	The regression coefficient.

mcmc_diagnostics also uses the mcmcplot function to create an HTML file with a panel of diagnostic plots (trace, density, and autocorrelation) for each monitored parameter.

Details

For each monitored parameter, mcmc_diagnostics considers the maximum R-hat and compares it with the threshold 1.1: convergence is achieved for the monitored parameter, when the maximum R-hat is below that threshold; otherwise, the Markov Chain Monte Carlo algorithm has not converged for that parameter. If the monitored parameter is a vector with the posterior results, there is only one R-hat. If the monitored parameter is a matrix of the posterior results, there are as many R-hats as the number of rows for that parameter.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_sensitivity' (using the default arguments)
res <- readRDS(system.file('extdata/res_sens_baker.rds',
               package = 'rnmamod'))

# Obtain the diagnostic plots and check convergence based on R-hat
mcmc_diagnostics(net = res,
                 par = c("tau", "EM[2,1]", "EM.pred[2,1]"))
#> The argument 'par' is ignored as it is used only for the functions 'run_model', 'run_ume' and 'run_metareg'
#> 
#>                                                R.hat max convergence status
#> Effect estimates (EM)                           1.074498           achieved
#> Predictions (EM_pred)                                 NA     Not applicable
#> Within-trial estimates (delta)                        NA     Not applicable
#> Between-trial standard deviation (tau)          1.128297             failed
#> Direct effects (node-splitting; direct)               NA     Not applicable
#> Indirect effect(s) (node-splitting; indirect)         NA     Not applicable
#> Inconsistency factor(s) (node-splitting; diff)        NA     Not applicable
#> Informative missingness parameter(s) (phi)            NA     Not applicable
#> Regression coefficient(s) (beta)                      NA     Not applicable