Markov Chain Monte Carlo diagnostics

Evaluates whether convergence has been achieved for the monitored parameters of the Bayesian models. The Gelman-Rubin convergence diagnostic, the Markov Chain Monte Carl (MCMC) error and relevant diagnostic plots are applied.

mcmc_diagnostics(net, par = NULL)

Arguments

net

An object of S3 class run_metareg, run_model, run_nodesplit, run_sensitivity, run_series_meta, and run_ume. See 'Value' in the functions above.

par

A vector of three character strings that refer to three monitored parameters in jagsfit which is an object of S3 class run_metareg, run_model, and run_ume. These three selected parameters will be considered in the diagnostic plots (see 'Value'). This argument will be ignored for objects of S3 class run_nodesplit, run_sensitivity, and run_series_meta.

Value

mcmc_diagnostics returns a data-frame that contains the Gelman-Rubin convergence diagnostic, R-hat, the MCMC error, and the convergence status of the following monitored parameters: EMThe estimated summary effect measure.EM_predThe predicted summary effect measure.deltaThe estimated trial-specific effect measure.tauThe between-trial standard deviation.directThe direct estimate of the split node (see 'Value' in run_nodesplit).indirectThe indirect estimate of the split node (see 'Value' in run_nodesplit).diffThe inconsistency factor of the split node (see 'Value' in run_nodesplit).phiThe informative missingness parameter.betaThe regression coefficient.mcmc_diagnostics also uses the mcmcplot function of the R-package mcmcplots to create an HTML file with a panel of diagnostic plots (trace, density, and autocorrelation) for each monitored parameter.

Details

For each monitored parameter, mcmc_diagnostics considers the maximum R-hat and MCMC error and compares them with the thresholds 1.1 and 5%, respectively. Convergence is achieved for the monitored parameter, when the maximum R-hat and MCMC error are below the corresponding thresholds; otherwise, the MCMC algorithm has not converged for that parameter. If the monitored parameter is a vector with the posterior results, there is only one R-hat and one MCMC error. If the monitored parameter is a matrix of the posterior results, there are as many R-hats and MCMC errors as the number of rows for that parameter. In that case, the maximum R-hat and MCMC error are considered.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_nodesplit' (using the default arguments)
res <- readRDS(system.file('extdata/node_baker.rds', package = 'rnmamod'))

# Check convergence based on R-hat
mcmc_diagnostics(net = res,
                 par = c("tau", "EM[2,1]", "EM.pred[2,1]"))