Robustness index

Calculates the robustness index, a novel index that quantifies the overall divergence of the sensitivity analysis results from the primary analysis results. The robustness index considers objective decision rules to infer the presence or lack of robustness of the primary analysis results when conducting a sensitivity analysis (Spineli et al., 2021).

Usage

robustness_index(sens, threshold)

Arguments

sens

An object of S3 class run_sensitivity when sensitivity analysis refers to different scenarios about the average missingness parameter. See 'Value' in run_sensitivity. For a general sensitivity analysis, insert a list of at least two objects of S3 class run_model indicating different re-analyses: the first object (of class run_model) in the list should refer to the primary analysis.

threshold

A number indicating the threshold of robustness, that is, the minimally allowed deviation between the primary analysis and re-analysis results. See 'Details' below.

Value

robustness_index prints on the R console a message in green text on the threshold of robustness determined by the user. Then, the function returns the following list of elements:

robust_index

A numeric scalar or vector on the robustness index values. In the case of a pairwise meta-analysis, robust_index is scalar as only one summary effect size is obtained. In the case of network meta-analysis, robust_index is a vector with length equal to the number of possible pairwise comparisons; one robustness index per possible comparison.

robust

A character or character vector (of same length with robust_index) on whether the primary analysis results are robust or frail to the different re-analyses.

kld

A vector or matrix on the Kullback-Leibler divergence (KLD) measure in the summary effect size from a subsequent re-analysis to the primary analysis. In the case of a pairwise meta-analysis, kld is a vector with length equal to the number of total analyses (one KLD value is obtained per analysis). The number of total analyses equals the square of the number of scenarios indicated in the argument mean_scenarios of run_sensitivity, in the case of missing participant outcome data; otherwise, the length of the character vector in argument sens. In the case of network meta-analysis, robust_index is a matrix with number of rows equal to the number of total analyses and number of columns equal to the number of possible pairwise comparisons; one KLD value per analysis and possible comparison.

threshold

The threshold used to be inherited by the heatmap_robustness function.

scenarios

The scenarios considered to be inherited by the kld_barplot function.

Details

Thresholds of robustness have been proposed only for the odds ratio and standardised mean difference (Spineli et al., 2021). The user may consider the values 0.28 and 0.17 in the argument threshold for the odds ratio and standardised mean difference effect measures (the default values), respectively, or consider other plausible values. When the argument threshold has not been defined, robustness_index considers the default values 0.28 and 0.17 as threshold for robustness for binary and continuous outcome, respectively, regardless of the effect measure (the default thresholds may not be proper choices for other effect measures; hence, use these threshold with great caution in this case). Spineli et al. (2021) offers a discussion on specifying the threshold of robustness.

In the case of binary outcome, robustness_index considers the results in the odds ratio scale to calculate the robustness index. This is because, the odds ratio is used as the 'best-case' effect measure in run_model. Then, relative risk, and risk difference are functions of the odds ratio and the selected baseline risk (See 'Details' in run_model).

In the case of missing participant outcome data, the primary analysis is considered to be the middle of the numbers in the argument mean_scenarios of run_sensitivity (see 'Arguments' and 'Details' in run_sensitivity).

In robust, the value "robust" appears when robust_index is less than threshold; otherwise, the value "frail" appears.

In the case of missing participant outcome data, robustness_index can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed in the R console.

References

Kullback S, Leibler RA. On information and sufficiency. Ann Math Stat 1951;22(1):79–86. doi: 10.1214/aoms/1177729694

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475–90. doi: 10.1002/jrsm.1478

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_sensitivity' (using the default arguments)
res_sens <- readRDS(system.file('extdata/res_sens_baker.rds',
                    package = 'rnmamod'))

# Calculate the robustness index
robustness_index(sens = res_sens,
                 threshold = 0.28)
#> The value 0.28 was assigned as 'threshold' for odds ratio.
#> Warning: data length differs from size of matrix: [700 != 25 x 350]
#> Warning: data length differs from size of matrix: [700 != 25 x 350]
#> $robust_index
#>  [1] 13.3385157 17.2941030  3.7720021  5.1970184  0.4324350  6.8941279
#>  [7]  3.1344856  5.6904306  2.1961330  6.7431447 18.2273695 20.9173851
#> [13] 24.2149541 18.3936674 14.4826448 18.7026469  8.4704781 10.1785559
#> [19]  0.1095445 22.9767926 10.1742322  2.7179036  1.5915401  3.1181725
#> [25]  8.0182916  7.2901989  7.3584645  4.4529765
#> 
#> $robust
#>  [1] "frail"  "frail"  "frail"  "frail"  "frail"  "frail"  "frail"  "frail" 
#>  [9] "frail"  "frail"  "frail"  "frail"  "frail"  "frail"  "frail"  "frail" 
#> [17] "frail"  "frail"  "robust" "frail"  "frail"  "frail"  "frail"  "frail" 
#> [25] "frail"  "frail"  "frail"  "frail" 
#> 
#> $kld
#>       [,1]       [,2] [,3]       [,4] [,5]       [,6] [,7]       [,8] [,9]
#>  [1,]    0 3.85050067    0 3.85050067    0 3.85050067    0 3.85050067    0
#>  [2,]    0 4.99237617    0 4.99237617    0 4.99237617    0 4.99237617    0
#>  [3,]    0 1.08887436    0 1.08887436    0 1.08887436    0 1.08887436    0
#>  [4,]    0 1.50024733    0 1.50024733    0 1.50024733    0 1.50024733    0
#>  [5,]    0 0.12480459    0 0.12480459    0 0.12480459    0 0.12480459    0
#>  [6,]    0 1.99016594    0 1.99016594    0 1.99016594    0 1.99016594    0
#>  [7,]    0 0.90484332    0 0.90484332    0 0.90484332    0 0.90484332    0
#>  [8,]    0 1.64268062    0 1.64268062    0 1.64268062    0 1.64268062    0
#>  [9,]    0 0.63397987    0 0.63397987    0 0.63397987    0 0.63397987    0
#> [10,]    0 1.94657557    0 1.94657557    0 1.94657557    0 1.94657557    0
#> [11,]    0 5.26179065    0 5.26179065    0 5.26179065    0 5.26179065    0
#> [12,]    0 6.03833055    0 6.03833055    0 6.03833055    0 6.03833055    0
#> [13,]    0 6.99025606    0 6.99025606    0 6.99025606    0 6.99025606    0
#> [14,]    0 5.30979284    0 5.30979284    0 5.30979284    0 5.30979284    0
#> [15,]    0 4.18078284    0 4.18078284    0 4.18078284    0 4.18078284    0
#> [16,]    0 5.39899012    0 5.39899012    0 5.39899012    0 5.39899012    0
#> [17,]    0 2.44520974    0 2.44520974    0 2.44520974    0 2.44520974    0
#> [18,]    0 2.93829882    0 2.93829882    0 2.93829882    0 2.93829882    0
#> [19,]    0 0.03215141    0 0.03215141    0 0.03215141    0 0.03215141    0
#> [20,]    0 6.63283047    0 6.63283047    0 6.63283047    0 6.63283047    0
#> [21,]    0 2.93704258    0 2.93704258    0 2.93704258    0 2.93704258    0
#> [22,]    0 0.78456944    0 0.78456944    0 0.78456944    0 0.78456944    0
#> [23,]    0 0.45945020    0 0.45945020    0 0.45945020    0 0.45945020    0
#> [24,]    0 0.90015396    0 0.90015396    0 0.90015396    0 0.90015396    0
#> [25,]    0 2.31468552    0 2.31468552    0 2.31468552    0 2.31468552    0
#> [26,]    0 2.10450750    0 2.10450750    0 2.10450750    0 2.10450750    0
#> [27,]    0 2.12420793    0 2.12420793    0 2.12420793    0 2.12420793    0
#> [28,]    0 1.28547931    0 1.28547931    0 1.28547931    0 1.28547931    0
#>            [,10] [,11]      [,12] [,13]      [,14] [,15]      [,16] [,17]
#>  [1,] 3.85050067     0 3.85050067     0 3.85050067     0 3.85050067     0
#>  [2,] 4.99237617     0 4.99237617     0 4.99237617     0 4.99237617     0
#>  [3,] 1.08887436     0 1.08887436     0 1.08887436     0 1.08887436     0
#>  [4,] 1.50024733     0 1.50024733     0 1.50024733     0 1.50024733     0
#>  [5,] 0.12480459     0 0.12480459     0 0.12480459     0 0.12480459     0
#>  [6,] 1.99016594     0 1.99016594     0 1.99016594     0 1.99016594     0
#>  [7,] 0.90484332     0 0.90484332     0 0.90484332     0 0.90484332     0
#>  [8,] 1.64268062     0 1.64268062     0 1.64268062     0 1.64268062     0
#>  [9,] 0.63397987     0 0.63397987     0 0.63397987     0 0.63397987     0
#> [10,] 1.94657557     0 1.94657557     0 1.94657557     0 1.94657557     0
#> [11,] 5.26179065     0 5.26179065     0 5.26179065     0 5.26179065     0
#> [12,] 6.03833055     0 6.03833055     0 6.03833055     0 6.03833055     0
#> [13,] 6.99025606     0 6.99025606     0 6.99025606     0 6.99025606     0
#> [14,] 5.30979284     0 5.30979284     0 5.30979284     0 5.30979284     0
#> [15,] 4.18078284     0 4.18078284     0 4.18078284     0 4.18078284     0
#> [16,] 5.39899012     0 5.39899012     0 5.39899012     0 5.39899012     0
#> [17,] 2.44520974     0 2.44520974     0 2.44520974     0 2.44520974     0
#> [18,] 2.93829882     0 2.93829882     0 2.93829882     0 2.93829882     0
#> [19,] 0.03215141     0 0.03215141     0 0.03215141     0 0.03215141     0
#> [20,] 6.63283047     0 6.63283047     0 6.63283047     0 6.63283047     0
#> [21,] 2.93704258     0 2.93704258     0 2.93704258     0 2.93704258     0
#> [22,] 0.78456944     0 0.78456944     0 0.78456944     0 0.78456944     0
#> [23,] 0.45945020     0 0.45945020     0 0.45945020     0 0.45945020     0
#> [24,] 0.90015396     0 0.90015396     0 0.90015396     0 0.90015396     0
#> [25,] 2.31468552     0 2.31468552     0 2.31468552     0 2.31468552     0
#> [26,] 2.10450750     0 2.10450750     0 2.10450750     0 2.10450750     0
#> [27,] 2.12420793     0 2.12420793     0 2.12420793     0 2.12420793     0
#> [28,] 1.28547931     0 1.28547931     0 1.28547931     0 1.28547931     0
#>            [,18] [,19]      [,20] [,21]      [,22] [,23]      [,24] [,25]
#>  [1,] 3.85050067     0 3.85050067     0 3.85050067     0 3.85050067     0
#>  [2,] 4.99237617     0 4.99237617     0 4.99237617     0 4.99237617     0
#>  [3,] 1.08887436     0 1.08887436     0 1.08887436     0 1.08887436     0
#>  [4,] 1.50024733     0 1.50024733     0 1.50024733     0 1.50024733     0
#>  [5,] 0.12480459     0 0.12480459     0 0.12480459     0 0.12480459     0
#>  [6,] 1.99016594     0 1.99016594     0 1.99016594     0 1.99016594     0
#>  [7,] 0.90484332     0 0.90484332     0 0.90484332     0 0.90484332     0
#>  [8,] 1.64268062     0 1.64268062     0 1.64268062     0 1.64268062     0
#>  [9,] 0.63397987     0 0.63397987     0 0.63397987     0 0.63397987     0
#> [10,] 1.94657557     0 1.94657557     0 1.94657557     0 1.94657557     0
#> [11,] 5.26179065     0 5.26179065     0 5.26179065     0 5.26179065     0
#> [12,] 6.03833055     0 6.03833055     0 6.03833055     0 6.03833055     0
#> [13,] 6.99025606     0 6.99025606     0 6.99025606     0 6.99025606     0
#> [14,] 5.30979284     0 5.30979284     0 5.30979284     0 5.30979284     0
#> [15,] 4.18078284     0 4.18078284     0 4.18078284     0 4.18078284     0
#> [16,] 5.39899012     0 5.39899012     0 5.39899012     0 5.39899012     0
#> [17,] 2.44520974     0 2.44520974     0 2.44520974     0 2.44520974     0
#> [18,] 2.93829882     0 2.93829882     0 2.93829882     0 2.93829882     0
#> [19,] 0.03215141     0 0.03215141     0 0.03215141     0 0.03215141     0
#> [20,] 6.63283047     0 6.63283047     0 6.63283047     0 6.63283047     0
#> [21,] 2.93704258     0 2.93704258     0 2.93704258     0 2.93704258     0
#> [22,] 0.78456944     0 0.78456944     0 0.78456944     0 0.78456944     0
#> [23,] 0.45945020     0 0.45945020     0 0.45945020     0 0.45945020     0
#> [24,] 0.90015396     0 0.90015396     0 0.90015396     0 0.90015396     0
#> [25,] 2.31468552     0 2.31468552     0 2.31468552     0 2.31468552     0
#> [26,] 2.10450750     0 2.10450750     0 2.10450750     0 2.10450750     0
#> [27,] 2.12420793     0 2.12420793     0 2.12420793     0 2.12420793     0
#> [28,] 1.28547931     0 1.28547931     0 1.28547931     0 1.28547931     0
#> 
#> $measure
#> [1] "OR"
#> 
#> $threshold
#> [1] 0.28
#> 
#> $scenarios
#> [1] -1.098612289 -0.693147181 -0.000100005  0.693147181  1.098612289
#> 
#> attr(,"class")
#> [1] "robustness_index"