Perform Bayesian pairwise or network meta-regression

Performs a one-stage pairwise or network meta-regression while addressing aggregate binary or continuous missing participant outcome data via the pattern-mixture model.

run_metareg(
  full,
  covariate,
  covar_assumption,
  n_chains,
  n_iter,
  n_burnin,
  n_thin
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

covariate

A numeric vector or matrix for a trial-specific covariate that is a potential effect modifier. See 'Details'.

covar_assumption

Character string indicating the structure of the intervention-by-covariate interaction, as described in Cooper et al. (2009). Set covar_assumption equal to "exchangeable", "independent", or "common".

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) for the following monitored parameters for a fixed-effect pairwise meta-analysis:

EM

The estimated summary effect measure (according to the argument measure defined in run_model).

beta_all

The estimated regression coefficient for all possible pairwise comparisons according to the argument covar_assumption.

dev_o

The deviance contribution of each trial-arm based on the observed outcome.

hat_par

The fitted outcome at each trial-arm.

phi

The informative missingness parameter.

For a fixed-effect network meta-analysis, the output additionally includes:

SUCRA

The surface under the cumulative ranking (SUCRA) curve for each intervention.

effectiveneness

The ranking probability of each intervention for every rank.

For a random-effects pairwise meta-analysis, the output additionally includes the following elements:

EM_pred

The predicted summary effect measure (according to the argument measure defined in run_model).

delta

The estimated trial-specific effect measure (according to the argument measure defined in run_model). For a multi-arm trial, we estimate T-1 effects, where T is the number of interventions in the trial.

tau

The between-trial standard deviation.

In network meta-analysis, EM and EM_pred refer to all possible pairwise comparisons of interventions in the network. Furthermore, tau is typically assumed to be common for all observed comparisons in the network. For a multi-arm trial, we estimate a total T-1 of delta for comparisons with the baseline intervention of the trial (found in the first column of the element t), with T being the number of interventions in the trial. Furthermore, the output includes the following elements:

leverage_o

The leverage for the observed outcome at each trial-arm.

sign_dev_o

The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment

A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit

An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

The run_metareg function also returns the arguments data, measure, model, assumption, covariate, covar_assumption, n_chains, n_iter, n_burnin, and n_thin to be inherited by other relevant functions of the package.

Details

run_metareg inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, D, ref, indic, and base_risk from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_metareg (see 'Examples').

The model runs in JAGS and the progress of the simulation appears on the R console. The output of run_metareg is used as an S3 object by other functions of the package to be processed further and provide an end-user-ready output.

The models described in Spineli et al. (2021), and Spineli (2019) have been extended to incorporate one study-level covariate variable following the assumptions of Cooper et al. (2009) for the structure of the intervention-by-covariate interaction. The covariate can be either a numeric vector or matrix with columns equal to the maximum number of arms in the dataset.

References

Cooper NJ, Sutton AJ, Morris D, Ades AE, Welton NJ. Addressing between-study heterogeneity and inconsistency in mixed treatment comparisons: Application to stroke prevention treatments in individuals with non-rheumatic atrial fibrillation. Stat Med 2009;28(14):1861--81. doi: 10.1002/sim.3594

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--975. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# Publication year
pub_year <- c(1996, 1998, 1999, 2000, 2000, 2001, rep(2002, 5), 2003, 2003,
              rep(2005, 4), 2006, 2006, 2007, 2007)

# \donttest{
# Perform a random-effects network meta-regression (exchangeable structure)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_metareg(full = res,
            covariate = pub_year,
            covar_assumption = "exchangeable",
            n_chains = 3,
            n_iter = 1000,
            n_burnin = 100,
            n_thin = 1)
# }