Perform the node-splitting approach

Performs the Bayesian node-splitting approach of Dias et al. (2010) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al., 2021; Spineli, 2019). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al., 2010).

Usage

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct

The summary effect measure (according to the argument measure defined in run_model) of each split node based on the corresponding trials.

indirect

The indirect summary effect measure (according to the argument measure defined in run_model) of each split node based on the remaining network after removing (splitting) the corresponding node.

diff

The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau

The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model_assessment

A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, ref, and indic from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see 'Examples').

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_nodesplit currently performs node-splitting using the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al. (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model).

run_nodesplit runs the Bayesian node-splitting approach in JAGS. The progress of the simulation appears on the R console. The number of times run_nodesplit is used appears on the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed on the R console. The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

run_nodesplit uses the mtc.nodesplit.comparisons function of the R-package gemtc to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al. (2016). run_nodesplit uses the option (1) in van Valkenhoef et al. (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, mtc.nodesplit.comparisons uses the option (3) in van Valkenhoef et al. (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function inherits the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects node-splitting approach
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 1000,
              n_burnin = 100,
              n_thin = 1)
#> JAGS generates initial values for the parameters.
#> 1 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 1 until convergence
#> 2 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 2 until convergence
#> 3 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 3 until convergence
#> 4 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2534
#> 
#> Initializing model
#> 
#> Updating model for split node 4 until convergence
#> 5 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2534
#> 
#> Initializing model
#> 
#> Updating model for split node 5 until convergence
#> 6 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2534
#> 
#> Initializing model
#> 
#> Updating model for split node 6 until convergence
#> 7 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2525
#> 
#> Initializing model
#> 
#> Updating model for split node 7 until convergence
#> 8 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 2530
#> 
#> Initializing model
#> 
#> Updating model for split node 8 until convergence
#> $direct
#>   treat1 treat2       mean        sd       2.5%       97.5%     Rhat n.eff
#> 1      6      1 -0.4118009 0.2379709 -0.9367185  0.07538669 1.053771    96
#> 2      7      1 -0.4977904 0.2472717 -1.2173362 -0.11976094 1.963469     5
#> 3      8      1 -0.3068949 0.1637373 -0.7229880 -0.06945310 1.324814    10
#> 4      5      4 -0.1662375 0.2009210 -0.5967559  0.22658327 1.004724   480
#> 5      7      4 -0.5620111 0.2423137 -1.0485343 -0.06154875 1.040517    58
#> 6      7      5 -0.3882015 0.2394913 -0.8650817  0.06477497 1.031592    93
#> 7      8      6 -1.6979965 0.8156387 -3.2673145 -0.27788530 1.059618    40
#> 8      8      7 -0.8828357 0.3495282 -1.5638817 -0.20704741 1.315256    11
#> 
#> $indirect
#>   treat1 treat2        mean        sd       2.5%      97.5%     Rhat n.eff
#> 1      6      1  1.23680347 0.8128773 -0.4118940  2.6294910 1.377411     9
#> 2      7      1  0.23449328 0.6197991 -0.9729792  1.4340783 1.359914     9
#> 3      8      1 -0.99428942 0.4215193 -1.9193491 -0.3047727 1.291854    11
#> 4      5      4 -0.52323992 0.5113064 -1.5678527  0.4247992 1.028821   140
#> 5      7      4 -0.20038920 0.2089812 -0.5954085  0.2271275 1.103734    24
#> 6      7      5 -0.29177040 0.3393078 -1.0204228  0.3305946 1.078478    46
#> 7      8      6 -0.08636863 0.2090346 -0.5117019  0.3241281 1.008156   300
#> 8      8      7 -0.06674555 0.1877759 -0.4885816  0.2791635 1.031403    73
#> 
#> $diff
#>   treat1 treat2        mean        sd        2.5%       97.5%     Rhat n.eff
#> 1      6      1 -1.64860435 0.8538752 -3.16060542  0.15578772 1.376179     9
#> 2      7      1 -0.73228368 0.7170582 -2.23528898  0.46806632 1.791653     6
#> 3      8      1  0.68739455 0.3944100 -0.05223545  1.53969276 1.170643    23
#> 4      5      4  0.35700242 0.5288460 -0.63320914  1.43920320 1.041657   110
#> 5      7      4 -0.36162191 0.3324475 -0.98086211  0.27016685 1.004146   550
#> 6      7      5 -0.09643106 0.4491935 -0.94550295  0.80532198 1.034079   170
#> 7      8      6 -1.61162792 0.8366833 -3.15339847 -0.10883549 1.050049    46
#> 8      8      7 -0.81609012 0.4330921 -1.65800685  0.08428173 1.287608    11
#> 
#> $tau
#>   treat1 treat2        50%         sd         2.5%     97.5%     Rhat n.eff
#> 1      6      1 0.15231074 0.10608094 0.0076732407 0.3892762 1.244054    15
#> 2      7      1 0.06869423 0.09087569 0.0005481785 0.3072761 2.211114     5
#> 3      8      1 0.09088988 0.10220045 0.0003085378 0.3336325 2.953375     4
#> 4      5      4 0.10292000 0.10374351 0.0066007311 0.3792919 1.036409   190
#> 5      7      4 0.10942186 0.08057161 0.0251331578 0.3269146 1.191910    15
#> 6      7      5 0.16560856 0.10099800 0.0483893615 0.4205952 1.049133    53
#> 7      8      6 0.14441903 0.08720261 0.0254429917 0.3566555 1.025982   130
#> 8      8      7 0.10423762 0.09375392 0.0116255152 0.3571677 1.002642  1900
#> 
#> $model
#> [1] "RE"
#> 
#> $model_assessment
#>   treat1 treat2 deviance      DIC       pD
#> 1      6      1 53.55050 90.14570 36.59520
#> 2      7      1 55.48235 90.04201 34.55966
#> 3      8      1 53.93050 88.96305 35.03255
#> 4      5      4 53.12954 91.44609 38.31654
#> 5      7      4 53.19020 89.87446 36.68426
#> 6      7      5 52.98297 92.25532 39.27236
#> 7      8      6 53.05971 89.45791 36.39820
#> 8      8      7 54.73180 91.81966 37.08785
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> attr(,"class")
#> [1] "run_nodesplit"
# }