Perform the node-splitting approach

Performs the Bayesian node-splitting approach of Dias et al. (2010) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al., 2021; Spineli, 2019). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al., 2010).

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct	The summary effect measure (according to the argument measure defined in run_model) of each split node based on the corresponding trials.

indirect	The indirect summary effect measure (according to the argument measure defined in run_model) of each split node based on the remaining network after removing (splitting) the corresponding node.

diff	The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau	The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model_assessment	A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see 'Examples').

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al. (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model).

run_nodesplit runs the Bayesian node-splitting approach in JAGS. The progress of the simulation appears on the R console. The number of times run_nodesplit is used appears on the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed on the R console.

run_nodesplit uses the mtc.nodesplit.comparisons function of the R-package gemtc to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al. (2016). run_nodesplit uses the option (1) in van Valkenhoef et al. (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, mtc.nodesplit.comparisons uses the option (3) in van Valkenhoef et al. (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function inherits the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--975. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects node-splitting approach
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 1000,
              n_burnin = 100,
              n_thin = 1)
#> 1 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 2 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 3 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 4 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5267
#> 
#> Initializing model
#> 
#> 5 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5267
#> 
#> Initializing model
#> 
#> 6 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5268
#> 
#> Initializing model
#> 
#> 7 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5259
#> 
#> Initializing model
#> 
#> 8 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5264
#> 
#> Initializing model
#> 
#> $direct
#>   treat1 treat2       mean        sd       2.5%       97.5%     Rhat n.eff
#> 1      6      1 -0.3873277 0.2652071 -0.8850601  0.14721105 1.219244    15
#> 2      7      1 -0.3821269 0.2115759 -0.7496827  0.03083592 1.070070    34
#> 3      8      1 -0.5095922 0.2257975 -0.9582121 -0.10202194 1.058869    97
#> 4      5      4 -0.2016000 0.2028216 -0.5572348  0.24760477 1.010899   340
#> 5      7      4 -0.5552726 0.2400957 -0.9940370 -0.09525698 1.076570    37
#> 6      7      5 -0.3662018 0.2293609 -0.8236182  0.02330352 1.274012    11
#> 7      8      6 -1.8184461 0.9012724 -3.7395398 -0.34102766 1.141065    20
#> 8      8      7 -0.9230884 0.3733013 -1.6453089 -0.21158625 1.249862    12
#> 
#> $indirect
#>   treat1 treat2        mean        sd       2.5%      97.5%     Rhat n.eff
#> 1      6      1  1.11199032 0.8348653 -0.2751997  2.9237964 1.099119    35
#> 2      7      1 -0.11396826 0.6454612 -1.6030600  1.1359320 1.134041    23
#> 3      8      1 -1.28932269 0.5389905 -2.6470268 -0.4390495 1.090383    43
#> 4      5      4 -0.65594832 0.5199095 -1.6044752  0.2993950 1.337477    10
#> 5      7      4 -0.17875473 0.1924316 -0.5713523  0.2171103 1.149972    18
#> 6      7      5 -0.25810269 0.2626253 -0.8003076  0.2136960 1.328514    10
#> 7      8      6 -0.08281475 0.2134203 -0.5480264  0.3000090 1.015870   190
#> 8      8      7 -0.05217033 0.2136210 -0.4970179  0.3620076 1.068706    37
#> 
#> $diff
#>   treat1 treat2       mean        sd       2.5%       97.5%     Rhat n.eff
#> 1      6      1 -1.4993180 0.8849486 -3.3701256  0.02144620 1.079643    37
#> 2      7      1 -0.2681587 0.6015260 -1.4115475  1.13106402 1.093922    36
#> 3      8      1  0.7797305 0.5217343 -0.1151459  1.97934652 1.095209    37
#> 4      5      4  0.4543483 0.5715536 -0.6500159  1.53748044 1.261940    13
#> 5      7      4 -0.3765179 0.3047378 -0.9824054  0.22804060 1.040973   200
#> 6      7      5 -0.1080991 0.3928108 -0.8648647  0.60567711 1.542617     7
#> 7      8      6 -1.7356314 0.9460156 -3.6621207 -0.16801150 1.147205    19
#> 8      8      7 -0.8709181 0.4651929 -1.7156156  0.07989799 1.204026    14
#> 
#> $tau
#>   treat1 treat2       50%         sd        2.5%     97.5%     Rhat n.eff
#> 1      6      1 0.1324154 0.09329003 0.010349812 0.3472648 1.621362     7
#> 2      7      1 0.1707689 0.08746981 0.024244409 0.3589821 1.364009    11
#> 3      8      1 0.1623430 0.08539552 0.024193983 0.3356758 1.046820    69
#> 4      5      4 0.1026571 0.07632202 0.013461669 0.2968045 1.133453    27
#> 5      7      4 0.1120171 0.08363083 0.024057362 0.3477892 1.011957   390
#> 6      7      5 0.1072796 0.08705009 0.005372873 0.3402534 1.294016    16
#> 7      8      6 0.1324767 0.08561449 0.011583450 0.3332715 1.276182    14
#> 8      8      7 0.1886068 0.09575736 0.037947102 0.3941237 1.247859    14
#> 
#> $model_assessment
#>   treat1 treat2      DIC deviance       pD
#> 1      6      1 89.99540 53.45429 36.54111
#> 2      7      1 90.16538 53.22395 36.94143
#> 3      8      1 89.96739 53.06466 36.90274
#> 4      5      4 91.18939 53.66989 37.51950
#> 5      7      4 90.78190 52.97499 37.80692
#> 6      7      5 91.54991 54.40724 37.14267
#> 7      8      6 89.12296 53.31364 35.80932
#> 8      8      7 92.01071 53.59451 38.41620
#> 
# }