Perform the node-splitting approach

Performs the Bayesian node-splitting approach of Dias et al. (2010) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al., 2021; Spineli, 2019). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al., 2010).

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct

The summary effect measure (according to the argument measure defined in run_model) of each split node based on the corresponding trials.

indirect

The indirect summary effect measure (according to the argument measure defined in run_model) of each split node based on the remaining network after removing (splitting) the corresponding node.

diff

The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau

The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model_assessment

A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see 'Examples').

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al. (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model).

run_nodesplit runs the Bayesian node-splitting approach in JAGS. The progress of the simulation appears on the R console. The number of times run_nodesplit is used appears on the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed on the R console.

run_nodesplit uses the mtc.nodesplit.comparisons function of the R-package gemtc to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al. (2016). run_nodesplit uses the option (1) in van Valkenhoef et al. (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, mtc.nodesplit.comparisons uses the option (3) in van Valkenhoef et al. (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function inherits the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--975. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects node-splitting approach
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 1000,
              n_burnin = 100,
              n_thin = 1)
#> 1 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 2 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 3 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 4 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5267
#> 
#> Initializing model
#> 
#> 5 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5267
#> 
#> Initializing model
#> 
#> 6 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5268
#> 
#> Initializing model
#> 
#> 7 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5259
#> 
#> Initializing model
#> 
#> 8 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5264
#> 
#> Initializing model
#> 
#> $direct
#>   treat1 treat2       mean        sd       2.5%        97.5%     Rhat n.eff
#> 1      6      1 -0.4260244 0.2043189 -0.8240061  0.006905684 1.013987   450
#> 2      7      1 -0.6207286 0.2235990 -1.0348453 -0.181520765 1.254332    13
#> 3      8      1 -0.6142802 0.2319593 -1.0296245 -0.182016282 1.576723     7
#> 4      5      4 -0.1513790 0.2125605 -0.6223254  0.194622369 1.124793    23
#> 5      7      4 -0.5449496 0.2383361 -1.0235659 -0.061871259 1.026904    87
#> 6      7      5 -0.4348256 0.2151137 -0.8805343  0.005947830 1.015599   360
#> 7      8      6 -1.7122923 0.9021977 -3.5299128 -0.019909152 1.088107    33
#> 8      8      7 -0.8720894 0.3660333 -1.5619485 -0.149797443 1.242308    12
#> 
#> $indirect
#>   treat1 treat2        mean        sd       2.5%      97.5%     Rhat n.eff
#> 1      6      1  0.98660854 0.7786255 -0.4655886  2.6367982 1.069042    35
#> 2      7      1 -0.39044838 0.6204541 -1.6515931  0.8532592 1.031054    82
#> 3      8      1 -1.79141830 0.7924242 -3.2766879 -0.3965316 1.264677    12
#> 4      5      4 -0.74717765 0.3571373 -1.4090449  0.0419487 1.225468    13
#> 5      7      4 -0.21916942 0.2237198 -0.6607624  0.2167690 1.063500    38
#> 6      7      5 -0.31521412 0.2711974 -0.8301784  0.1978488 1.171475    16
#> 7      8      6 -0.05908925 0.1909420 -0.4692070  0.2860590 1.050192    57
#> 8      8      7 -0.01764719 0.2191612 -0.5182623  0.3474076 1.062693    43
#> 
#> $diff
#>   treat1 treat2       mean        sd        2.5%      97.5%     Rhat n.eff
#> 1      6      1 -1.4126329 0.8010519 -3.10919025 0.04961144 1.077400    34
#> 2      7      1 -0.2302802 0.5778000 -1.37900392 0.91948821 1.008686   930
#> 3      8      1  1.1771381 0.7179978 -0.04428144 2.49742262 1.196623    15
#> 4      5      4  0.5957987 0.3683580 -0.27520987 1.24006205 1.092431    28
#> 5      7      4 -0.3257802 0.3175845 -0.93142804 0.29067934 1.051187    44
#> 6      7      5 -0.1196114 0.3772940 -0.91996426 0.60513913 1.078007    34
#> 7      8      6 -1.6532031 0.9348473 -3.46137428 0.19970768 1.103918    27
#> 8      8      7 -0.8544422 0.4528424 -1.70537985 0.15497461 1.241893    12
#> 
#> $tau
#>   treat1 treat2        50%         sd        2.5%     97.5%     Rhat n.eff
#> 1      6      1 0.13295610 0.07716093 0.016907175 0.3122890 1.096902    73
#> 2      7      1 0.14251880 0.09836835 0.008810835 0.3589458 1.270602    15
#> 3      8      1 0.11233709 0.08244377 0.013534923 0.3004815 1.452799     9
#> 4      5      4 0.06935735 0.08972232 0.001879428 0.3069354 1.490436     8
#> 5      7      4 0.15600574 0.07416634 0.055951774 0.3332944 1.011398   350
#> 6      7      5 0.14518258 0.09000010 0.023210835 0.3634770 1.094541    33
#> 7      8      6 0.11202252 0.08041243 0.008025215 0.3102506 1.197712    16
#> 8      8      7 0.14707475 0.11867082 0.010382611 0.4340239 1.049308    65
#> 
#> $model_assessment
#>   treat1 treat2      DIC deviance       pD
#> 1      6      1 88.17248 53.49266 34.67982
#> 2      7      1 90.83132 54.19436 36.63696
#> 3      8      1 88.45031 52.51950 35.93081
#> 4      5      4 88.81425 52.71329 36.10096
#> 5      7      4 91.48515 53.15663 38.32852
#> 6      7      5 91.39985 53.65983 37.74002
#> 7      8      6 89.34540 54.08642 35.25898
#> 8      8      7 92.93502 54.81065 38.12437
#> 
# }