Perform the node-splitting approach

Performs the Bayesian node-splitting approach of Dias et al. (2010) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al., 2021; Spineli, 2019). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al., 2010).

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct	The summary effect measure (according to the argument measure defined in run_model) of each split node based on the corresponding trials.

indirect	The indirect summary effect measure (according to the argument measure defined in run_model) of each split node based on the remaining network after removing (splitting) the corresponding node.

diff	The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau	The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model_assessment	A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see 'Examples').

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al. (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model).

run_nodesplit runs the Bayesian node-splitting approach in JAGS. The progress of the simulation appears on the R console. The number of times run_nodesplit is used appears on the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed on the R console.

run_nodesplit uses the mtc.nodesplit.comparisons function of the R-package gemtc to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al. (2016). run_nodesplit uses the option (1) in van Valkenhoef et al. (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, mtc.nodesplit.comparisons uses the option (3) in van Valkenhoef et al. (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function inherits the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--975. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects node-splitting approach
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 1000,
              n_burnin = 100,
              n_thin = 1)
#> 1 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 2 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 3 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5258
#> 
#> Initializing model
#> 
#> 4 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5267
#> 
#> Initializing model
#> 
#> 5 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5267
#> 
#> Initializing model
#> 
#> 6 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5268
#> 
#> Initializing model
#> 
#> 7 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5259
#> 
#> Initializing model
#> 
#> 8 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5264
#> 
#> Initializing model
#> 
#> $direct
#>   treat1 treat2       mean        sd       2.5%       97.5%     Rhat n.eff
#> 1      6      1 -0.4249323 0.3309013 -0.9781062  0.22158154 2.112721     5
#> 2      7      1 -0.5172382 0.2276804 -0.9052110 -0.06867215 1.806120     6
#> 3      8      1 -0.4411137 0.2787469 -1.1069254  0.07745253 1.403387     9
#> 4      5      4 -0.2290733 0.2221004 -0.6935540  0.19633618 1.012345   200
#> 5      7      4 -0.5420435 0.2325042 -1.0126370 -0.12338790 1.107960    23
#> 6      7      5 -0.4052476 0.2216776 -0.8789262  0.03037011 1.148107    23
#> 7      8      6 -1.3182838 0.7562662 -2.7713492  0.53613681 1.100666    28
#> 8      8      7 -0.8024307 0.3735362 -1.5539485 -0.04958035 1.142612    19
#> 
#> $indirect
#>   treat1 treat2        mean        sd       2.5%      97.5%     Rhat n.eff
#> 1      6      1  0.72741079 1.0523048 -0.8922358 2.59281504 2.050121     5
#> 2      7      1 -0.05304679 0.8176189 -1.4004478 1.74964386 1.299585    11
#> 3      8      1 -1.04381932 0.6601272 -2.4219946 0.08164612 1.250748    13
#> 4      5      4 -0.73868337 0.5456398 -1.8297584 0.54740629 1.105059    96
#> 5      7      4 -0.25103557 0.2311784 -0.7304773 0.17883377 1.212784    14
#> 6      7      5 -0.39076667 0.3414399 -0.9974828 0.32607178 1.133397    27
#> 7      8      6 -0.08982424 0.2091228 -0.4724975 0.31120104 1.176050    16
#> 8      8      7 -0.04902258 0.2000638 -0.4338437 0.34842541 1.013607   170
#> 
#> $diff
#>   treat1 treat2        mean        sd       2.5%     97.5%     Rhat n.eff
#> 1      6      1 -1.15234306 0.8507732 -2.6632673 0.3172376 1.643976     6
#> 2      7      1 -0.46419143 0.7618396 -2.2228809 0.7926852 1.117886    22
#> 3      8      1  0.60270560 0.5696508 -0.4090246 1.8699963 1.089317    29
#> 4      5      4  0.50961010 0.5523296 -0.8295904 1.5661156 1.109545    88
#> 5      7      4 -0.29100793 0.2931307 -0.8539935 0.3101689 1.036447   190
#> 6      7      5 -0.01448089 0.4684542 -1.0424391 0.8022100 1.197364    19
#> 7      8      6 -1.22845951 0.7670925 -2.6565340 0.6603301 1.051956    73
#> 8      8      7 -0.75340812 0.4574637 -1.6279238 0.1719979 1.112283    23
#> 
#> $tau
#>   treat1 treat2       50%         sd        2.5%     97.5%     Rhat n.eff
#> 1      6      1 0.0718814 0.07846926 0.001979634 0.2735281 1.434586     9
#> 2      7      1 0.1372810 0.11170408 0.001748279 0.4515977 1.419728    11
#> 3      8      1 0.1521524 0.09318821 0.027250814 0.3688429 1.055002    77
#> 4      5      4 0.1329093 0.08672263 0.022709340 0.3706712 1.138803    21
#> 5      7      4 0.1340352 0.09102338 0.021370269 0.3691923 1.253372    15
#> 6      7      5 0.1472006 0.10774996 0.002135033 0.3980344 1.565648     8
#> 7      8      6 0.1032133 0.09149371 0.005540987 0.3189888 1.053891    55
#> 8      8      7 0.1816427 0.10326931 0.044138866 0.4594903 1.042821    75
#> 
#> $model_assessment
#>   treat1 treat2      DIC deviance       pD
#> 1      6      1 89.46580 53.90188 35.56393
#> 2      7      1 91.35886 54.15266 37.20619
#> 3      8      1 90.39423 52.91666 37.47757
#> 4      5      4 92.71720 54.41059 38.30661
#> 5      7      4 91.04503 53.25208 37.79296
#> 6      7      5 91.34404 53.77520 37.56884
#> 7      8      6 88.18801 53.37948 34.80854
#> 8      8      7 91.43272 53.15697 38.27575
#> 
# }