Perform the node-splitting approach

Performs the Bayesian node-splitting approach of Dias et al., (2010) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli, 2019; Spineli et al., 2021). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al., 2010).

Usage,

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct	The summary effect measure (according to the argument measure defined in run_model) of each split node based on the corresponding trials.

indirect	The indirect summary effect measure (according to the argument measure defined in run_model) of each split node based on the remaining network after removing (splitting) the corresponding node.

diff	The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau	The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model_assessment	A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var_misspar from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see 'Examples').

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al., (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model).

run_nodesplit runs the Bayesian node-splitting approach in JAGS. The progress of the simulation appears on the R console. The number of times run_nodesplit is used appears on the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed on the R console.

run_nodesplit uses the mtc.nodesplit.comparisons function of the R-package gemtc to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al., (2016). run_nodesplit uses the option (1) in van Valkenhoef et al., (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, mtc.nodesplit.comparisons uses the option (3) in van Valkenhoef et al., (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials. Hence, the corresponding subnetworks obtained after splitting the node differ, and by extend, the results of the corresponding split nodes found in multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function uses the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_baker.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects node-splitting approach
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 1000,
              n_burnin = 100,
              n_thin = 1)
#> 1 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5083
#> 
#> Initializing model
#> 
#> 2 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5083
#> 
#> Initializing model
#> 
#> 3 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5083
#> 
#> Initializing model
#> 
#> 4 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5090
#> 
#> Initializing model
#> 
#> 5 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5090
#> 
#> Initializing model
#> 
#> 6 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5092
#> 
#> Initializing model
#> 
#> 7 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5084
#> 
#> Initializing model
#> 
#> 8 out of 8 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 79
#>    Unobserved stochastic nodes: 141
#>    Total graph size: 5089
#> 
#> Initializing model
#> 
#> $direct
#>   treat1 treat2       mean        sd       2.5%        97.5%     Rhat n.eff
#> 1      6      1 -0.4171952 0.2742219 -0.8327364  0.216972839 1.442538     8
#> 2      7      1 -0.4681402 0.2279647 -0.9655031 -0.081728203 1.427280     8
#> 3      8      1 -0.4395447 0.2139879 -0.9449624 -0.109171466 1.205967    17
#> 4      5      4 -0.2066307 0.2032534 -0.6332277  0.196509450 1.029936   100
#> 5      7      4 -0.5608216 0.1970737 -0.9848288 -0.190448595 1.061055    46
#> 6      7      5 -0.4247075 0.2233930 -0.8517251  0.007950549 1.045950   110
#> 7      8      6 -1.4135188 0.7640011 -3.0951093  0.006398113 1.005698   380
#> 8      8      7 -0.8406550 0.3948944 -1.5660600  0.011281168 1.066277    36
#> 
#> $indirect
#>   treat1 treat2        mean        sd       2.5%     97.5%     Rhat n.eff
#> 1      6      1  0.92978021 0.7638990 -0.3668793 2.6018227 1.112345    28
#> 2      7      1  0.08898774 0.6637699 -1.1997910 1.4127703 1.063503    48
#> 3      8      1 -0.90311820 0.5184888 -1.9322438 0.1149535 1.118947    24
#> 4      5      4 -0.62730128 0.4083636 -1.3846009 0.1501757 1.121164    25
#> 5      7      4 -0.16943073 0.2255933 -0.6315812 0.2033143 1.051922    55
#> 6      7      5 -0.33027671 0.2673588 -0.8818607 0.1928913 1.250005    12
#> 7      8      6 -0.09118663 0.1937407 -0.4837850 0.2929146 1.014063  1900
#> 8      8      7 -0.04893434 0.1975734 -0.4541301 0.3489574 1.009840   490
#> 
#> $diff
#>   treat1 treat2        mean        sd       2.5%       97.5%     Rhat n.eff
#> 1      6      1 -1.34697542 0.7154272 -2.9266432 -0.07024535 1.070751    68
#> 2      7      1 -0.55712794 0.6641563 -1.8433895  0.74236346 1.030897   160
#> 3      8      1  0.46357348 0.4784014 -0.4434107  1.38836175 1.069804    79
#> 4      5      4  0.42067059 0.4642730 -0.5202529  1.32790193 1.065048    42
#> 5      7      4 -0.39139086 0.2932681 -0.9424030  0.21291250 1.010123  2700
#> 6      7      5 -0.09443078 0.3761121 -0.8121403  0.62823720 1.171538    17
#> 7      8      6 -1.32233216 0.7995573 -2.9785402  0.18212922 1.004954   450
#> 8      8      7 -0.79172062 0.4551886 -1.5616203  0.20138021 1.044709    54
#> 
#> $tau
#>   treat1 treat2        50%         sd        2.5%     97.5%     Rhat n.eff
#> 1      6      1 0.14389586 0.10101486 0.011905066 0.3820974 1.399820    10
#> 2      7      1 0.13967364 0.08981056 0.013154304 0.3413205 1.167209    25
#> 3      8      1 0.14053040 0.09170506 0.012127188 0.3539291 1.029624   140
#> 4      5      4 0.13276697 0.08908458 0.004542895 0.3265989 1.351959    11
#> 5      7      4 0.08923853 0.08282554 0.007814448 0.2938232 1.189818    15
#> 6      7      5 0.15409848 0.09501836 0.006013537 0.3383281 1.364012    10
#> 7      8      6 0.11660650 0.10105463 0.006749674 0.3603156 1.118580    27
#> 8      8      7 0.17332407 0.12522953 0.028395875 0.4946055 1.023015   180
#> 
#> $model_assessment
#>   treat1 treat2      DIC deviance       pD
#> 1      6      1 88.59064 52.51665 36.07399
#> 2      7      1 90.91992 54.25003 36.66989
#> 3      8      1 89.84727 53.58568 36.26160
#> 4      5      4 90.99412 53.36135 37.63277
#> 5      7      4 90.51328 53.19927 37.31401
#> 6      7      5 91.95833 54.02889 37.92943
#> 7      8      6 89.36484 53.74085 35.62399
#> 8      8      7 92.55384 54.10096 38.45287
#> 
# }