Perform the node-splitting approach

Performs the Bayesian node-splitting approach of Dias et al., (2010) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli, 2019; Spineli et al., 2021). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al., 2010).

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

n_chains
Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter
Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin
Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin
Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct	The summary effect measure (according to the argument measure defined in run_model) of each split node based on the corresponding trials.

indirect	The indirect summary effect measure (according to the argument measure defined in run_model) of each split node based on the remaining network after removing (splitting) the corresponding node.

diff	The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau	The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model_assessment	A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var_misspar from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see 'Examples').

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al., (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model).

run_nodesplit runs the Bayesian node-splitting approach in JAGS. The progress of the simulation appears on the R console. The number of times run_nodesplit is used appears on the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed on the R console.

run_nodesplit uses the mtc.nodesplit.comparisons function of the R-package gemtc to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al., (2016). run_nodesplit uses the option (1) in van Valkenhoef et al., (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, mtc.nodesplit.comparisons uses the option (3) in van Valkenhoef et al., (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials. Hence, the corresponding subnetworks obtained after splitting the node differ, and by extend, the results of the corresponding split nodes found in multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function uses the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. doi: 10.1002/jrsm.1167

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. doi: 10.1002/sim.3767

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects node-splitting approach
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 1000,
              n_burnin = 100,
              n_thin = 1)
#> 1 out of 5 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 39
#>    Unobserved stochastic nodes: 77
#>    Total graph size: 2538
#> 
#> Initializing model
#> 
#> 2 out of 5 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 39
#>    Unobserved stochastic nodes: 77
#>    Total graph size: 2535
#> 
#> Initializing model
#> 
#> 3 out of 5 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 39
#>    Unobserved stochastic nodes: 77
#>    Total graph size: 2539
#> 
#> Initializing model
#> 
#> 4 out of 5 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 39
#>    Unobserved stochastic nodes: 77
#>    Total graph size: 2538
#> 
#> Initializing model
#> 
#> 5 out of 5 split nodes
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 39
#>    Unobserved stochastic nodes: 77
#>    Total graph size: 2544
#> 
#> Initializing model
#> 
#> $direct
#>   treat1 treat2       mean        sd       2.5%    97.5%     Rhat n.eff
#> 1      2      1  0.4569921 1.0996127 -1.7810438 2.675252 1.042961    64
#> 2      4      1  0.3481521 0.6664940 -0.9150557 1.693172 1.016116   140
#> 3      5      1  1.8095744 0.9148393  0.2085485 3.822889 1.048216    56
#> 4      4      2 -0.5370164 1.1669469 -2.9010817 1.582377 1.008067   410
#> 5      4      3  1.6052275 1.0519354 -0.7011241 3.691963 1.024002   100
#> 
#> $indirect
#>   treat1 treat2        mean       sd      2.5%    97.5%     Rhat n.eff
#> 1      2      1  1.22468392 1.471158 -1.694242 3.981058 1.018064   130
#> 2      4      1  1.02072590 1.211544 -1.330777 3.525290 1.000731  2700
#> 3      5      1 -0.41099491 1.821607 -4.437378 2.804795 1.042699    83
#> 4      4      2  0.09074088 1.092031 -2.236228 2.164116 1.019629   390
#> 5      4      3 -0.23526565 1.056744 -2.402181 1.801434 1.029798    76
#> 
#> $diff
#>   treat1 treat2       mean       sd      2.5%    97.5%     Rhat n.eff
#> 1      2      1 -0.7676918 1.700495 -3.873946 2.715280 1.003346   730
#> 2      4      1 -0.6725738 1.263741 -3.465549 1.726317 1.007412   320
#> 3      5      1  2.2205693 1.763608 -1.113354 6.000685 1.015116   390
#> 4      4      2 -0.6277572 1.578240 -3.830823 2.671740 1.005035  2700
#> 5      4      3  1.8404931 1.633223 -1.502139 5.035177 1.042758    53
#> 
#> $tau
#>   treat1 treat2       50%        sd       2.5%    97.5%     Rhat n.eff
#> 1      2      1 0.7763761 0.4095607 0.20501082 1.739858 1.007725   280
#> 2      4      1 0.7880710 0.4020449 0.11809192 1.715408 1.136265    37
#> 3      5      1 0.5490102 0.3779794 0.05814441 1.496177 1.045221    80
#> 4      4      2 0.7041921 0.4354579 0.06643632 1.732713 1.035404   110
#> 5      4      3 0.6252761 0.4105260 0.09242337 1.637255 1.049247    63
#> 
#> $model_assessment
#>   treat1 treat2      DIC deviance       pD
#> 1      2      1 44.51124 24.16676 20.34447
#> 2      4      1 44.13187 23.51070 20.62117
#> 3      5      1 44.54135 24.36482 20.17653
#> 4      4      2 44.44220 24.56095 19.88125
#> 5      4      3 44.57968 24.39544 20.18424
#> 
# }