Bayesian node-splitting approach to explore (in)consistency

Performs a Bayesian network meta-analysis based on the node-splitting approach of Dias et al. (2010) extended to address aggregate binary and continuous participant outcome data via the pattern-mixture model (Spineli, 2019; Spineli et al., 2021). This model offers a local evaluation of the plausibility of the consistency assumption in the network (Dias et al. (2010)).

run_nodesplit(full, n_chains, n_iter, n_burnin, n_thin)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

n_chains
Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter
Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin
Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin
Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

direct	The (summary) direct effect estimate of each split node based on the corresponding trials.

indirect	The indirect effect estimate of each split node based on the remaining network after removing (splitting) the corresponding node.

diff	The inconsistency parameter for each split node defined as the difference between the direct and indirect effect of the corresponding split node.

tau	The between-trial standard deviation after each split node, when the random-effects model has been specified.

Furthermore, the output includes the following element:

model.assessment	A data-frame on the measures of model assessment after each split node: deviance information criterion, total residual deviance, and number of effective parameters.

Details

run_nodesplit does not contain the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var.misspar that are found in run_model. This is to prevent misspecifying the Bayesian model that has been considered in run_model. Instead, these arguments are contained in the argument full of the function. Therefore, the user needs first to apply run_model, and then use run_nodesplit (see, 'Examples').

To perform the Bayesian node-splitting approach, the prepare_nodesplit function is called which contains the WinBUGS code as written by Dias et al. (2010) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_nodesplit has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see, 'Details' in run_model).

run_nodesplit runs Bayesian network meta-analysis based on the node-splitting approach in JAGS. The progress of the simulation appears in the R console. The number of times run_nodesplit is used appears in the R console as a text in red and it equals the number of split nodes (see 'Examples'). If there are no split nodes in the network, the execution of the function will be stopped and an error message will be printed in the R console.

run_nodesplit uses the mtc.nodesplit.comparisons function of the gemtc package to obtain automatically the nodes to split based on the decision rule of van Valkenhoef et al. (2016). run_nodesplit uses the option (1) in van Valkenhoef et al. (2016) to parameterise multi-arm trials that contain the node-to-split. In contrast, the mtc.nodesplit function of the gemtc package uses the option (3) in van Valkenhoef et al. (2016). Option (1) keeps the baseline arm of the node-to-split in the corresponding multi-arms. Option (3) excludes both arms of the node-to-split from the corresponding multi-arm trials. Hence, the corresponding subnetworks obtained after splitting the node differ, and by extend, the results that correspond to split nodes found in multi-arm trials.

The output of run_nodesplit is not end-user-ready. The nodesplit_plot function uses the output of run_nodesplit as an S3 object and processes it further to provide an end-user-ready output.

run_nodesplit can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.

References

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. [doi: 10.1177/0962280220983544 ]

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. [doi: 10.1186/s12874-019-0731-y ]

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed treatment comparison meta-analysis. Stat Med 2010;29(7-8):932--44. [doi: 10.1002/sim.3767 ]

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472. [doi: 10.1214/ss/1177011136 ]

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of node-splitting models for assessment of inconsistency in network meta-analysis. Res Synth Methods 2016;7(1):80--93. [doi: 10.1002/jrsm.1167 ]

See also

Author

Loukia M. Spineli

Examples

data("nma.baker2009")

if (FALSE) {
# Perform a random-effects network meta-analysis
res <- run_model(data = nma.baker2009,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 1,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# Run random-effects network meta-analysis with node-splitting approach
run_nodesplit(full = res,
              n_chains = 3,
              n_iter = 10000,
              n_burnin = 1000,
              n_thin = 1)
}