Perform sensitivity analysis for missing participant outcome data

Performs a sensitivity analysis by applying pairwise meta-analysis or network meta-analysis for a series of different scenarios about the informative missingness parameter,

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length of 5 or larger. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default arguments are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects pairwise meta-analysis:

EM

The estimated summary effect measure (according to the argument measure defined in run_model).

tau

The between-trial standard deviation. This element does not appear in the case of a fixed-effect pairwise meta-analysis.

In a random-effects network meta-analysis, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore, tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the vector specified in mean_scenarios (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package function to be processed further and provide an end-user-ready output.

In the case of pairwise meta-analysis, EM and tau are estimated as many times as the number of scenarios considered. In the case of network meta-analysis, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console. Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to Spineli (2019), Mavridis et al. (2015), Turner et al. (2015), and White et al. (2008) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, heter_prior, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model). Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci. 1992;7:457--472.

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--741. doi: 10.1002/sim.6365

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--490. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--2080. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--727. doi: 10.1002/sim.3008

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (default arguments)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The following vector of scenarios was considered by default: c(-2, -1, 0, 1, 2)
#> 
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> $EM
#>                mean        sd       2.5%     97.5%      Rhat n.eff
#>  [1,] -0.1143233450 0.3038885 -0.7112382 0.4385605 1.0058510   260
#>  [2,] -0.1136730249 0.3497224 -0.8164001 0.5784351 1.0056469   540
#>  [3,] -0.1453408197 0.3406688 -0.8633990 0.5692992 1.0064979   540
#>  [4,] -0.1955085458 0.3397800 -0.9229207 0.4838299 1.0012468   540
#>  [5,] -0.1990557279 0.3552751 -0.9923372 0.4582267 1.0202941   120
#>  [6,] -0.0930722726 0.3489889 -0.7966542 0.5401301 1.0363401   540
#>  [7,] -0.0853192319 0.3095645 -0.6915525 0.5441034 1.0117489   290
#>  [8,] -0.1180848901 0.3428750 -0.8770234 0.4727859 1.0152966   540
#>  [9,] -0.1275082398 0.3171325 -0.7223087 0.4769922 1.0130789   140
#> [10,] -0.1376560026 0.3602184 -0.9312075 0.6100578 1.0028933   540
#> [11,] -0.0883044249 0.3413872 -0.7785074 0.6016023 1.0037221   540
#> [12,] -0.0829886427 0.3393888 -0.7129343 0.5991814 0.9999549   540
#> [13,] -0.0696467911 0.3204388 -0.7032158 0.5331390 1.0178137   180
#> [14,] -0.1150293703 0.3955918 -0.8179169 0.7040457 1.0498143   540
#> [15,] -0.1283480113 0.2823664 -0.7281587 0.4057107 1.0003413   540
#> [16,] -0.0388056127 0.3104717 -0.6607481 0.5942945 1.0041861   340
#> [17,] -0.0457913756 0.3207268 -0.7195066 0.6157734 1.0274790   100
#> [18,] -0.0744394348 0.3398067 -0.8394006 0.5863599 1.0098455   540
#> [19,] -0.0964590001 0.3480953 -0.7120842 0.6382576 1.0121464   340
#> [20,] -0.1193117546 0.3464866 -0.8078749 0.6967191 1.0137959   540
#> [21,] -0.0138861561 0.3081762 -0.6948123 0.6114241 1.0065299   390
#> [22,] -0.0001034023 0.3728113 -0.7099140 0.7530102 0.9998997   540
#> [23,] -0.0417786886 0.3485763 -0.7568870 0.6327906 1.0117016   540
#> [24,] -0.0484176908 0.3330919 -0.6809987 0.5564207 1.0047958   540
#> [25,] -0.0955138651 0.3452589 -0.6995031 0.6468030 1.0093584   540
#> 
#> $tau
#>             50%        sd        2.5%    97.5%     Rhat n.eff
#>  [1,] 0.3637003 0.3218933 0.008203490 1.260953 1.032552   200
#>  [2,] 0.3742772 0.3660393 0.014139072 1.285409 1.064644    59
#>  [3,] 0.3754220 0.3410894 0.036000680 1.272131 1.011145   260
#>  [4,] 0.3455308 0.3589318 0.024815184 1.405330 1.016884   250
#>  [5,] 0.4010464 0.3639443 0.017295436 1.351402 1.002246   540
#>  [6,] 0.3523784 0.3643193 0.012595288 1.324935 1.039429    70
#>  [7,] 0.3230249 0.2916129 0.024093569 1.118097 1.011453   540
#>  [8,] 0.3591957 0.3150973 0.012670947 1.144954 1.084875    44
#>  [9,] 0.3306590 0.3126683 0.020648976 1.164005 1.009175   280
#> [10,] 0.4019707 0.3418640 0.015211052 1.196860 1.066006    66
#> [11,] 0.3248147 0.3248158 0.017050856 1.166910 1.081265    83
#> [12,] 0.3479151 0.3532467 0.007054801 1.347320 1.103223    44
#> [13,] 0.3074142 0.3394693 0.006088491 1.228283 1.045014   350
#> [14,] 0.4009849 0.4097838 0.019021701 1.634853 1.116051    22
#> [15,] 0.3385987 0.2929029 0.022008748 1.045862 1.039379    62
#> [16,] 0.3344093 0.3192332 0.012806257 1.156753 1.055429   320
#> [17,] 0.3194741 0.3128551 0.028046427 1.201906 1.013203   540
#> [18,] 0.3600262 0.3396028 0.015431708 1.288485 1.054722    52
#> [19,] 0.3806232 0.3355189 0.027252426 1.251623 1.008300   470
#> [20,] 0.3448923 0.3695138 0.020865376 1.537245 1.018252   150
#> [21,] 0.2912831 0.2888834 0.010177618 1.091354 1.050532    53
#> [22,] 0.3436301 0.3655080 0.043151915 1.291432 1.011717   180
#> [23,] 0.3482885 0.3509060 0.016994940 1.300556 1.024206    83
#> [24,] 0.3592972 0.3263815 0.038066917 1.169917 1.029240   140
#> [25,] 0.3694771 0.3551261 0.025687291 1.316510 1.048271   180
#> 
#> $measure
#> [1] "SMD"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] 0 1 1
#> 
# }