Perform sensitivity analysis for aggregate missing outcome participant data

Performs a sensitivity analysis by applying pairwise meta-analysis (PMA) or network meta-analysis (NMA) for a series of different scenarios about the informative missingness parameter,

Usage,

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length of 5 or larger positive odd integer. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default arguments are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects PMA:

EM	The estimated summary effect measure (according to the argument measure defined in run_model).

tau	The between-trial standard deviation. This element does not appear in the case of a fixed-effect PMA.

In a random-effects NMA, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore, tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the mean_scenarios vector (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package function to be processed further and provide an end-user-ready output.

In the case of PMA, EM and tau have as many rows as the number of scenarios considered. In the case of NMA, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console. Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to White et al., (2008), Mavridis et al., (2015), Turner et al., (2015) and Spineli (2019) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, and heter_prior from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model). Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--490. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--741. doi: 10.1002/sim.6365

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--2080. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--727. doi: 10.1002/sim.3008

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci. 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (missing-at-random assumption)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The following vector of scenarios was considered by default: c(-2, -1, 0, 1, 2)
#> 
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> $EM
#>               mean        sd       2.5%     97.5%     Rhat n.eff
#>  [1,] -0.119893241 0.3685249 -0.9088419 0.6377648 1.004832   540
#>  [2,] -0.104417666 0.3309644 -0.7638519 0.5886600 1.001303   540
#>  [3,] -0.127219834 0.3182832 -0.7733581 0.5207493 1.024210   540
#>  [4,] -0.156500525 0.3588872 -0.9359761 0.5689154 1.006121   540
#>  [5,] -0.174990274 0.3761623 -0.8553781 0.6117313 1.036183   540
#>  [6,] -0.064575788 0.3447200 -0.7462637 0.6348227 1.014871   170
#>  [7,] -0.122084109 0.3502896 -0.7833387 0.5510343 1.016198   540
#>  [8,] -0.143022645 0.3110131 -0.7932578 0.4577146 1.024255   410
#>  [9,] -0.128791506 0.3456341 -0.8076051 0.5880939 1.009699   180
#> [10,] -0.163531630 0.3419341 -0.8462660 0.5234321 1.000788   540
#> [11,] -0.062004351 0.3148918 -0.7281061 0.4765463 1.003740   540
#> [12,] -0.074124931 0.3605351 -0.7868941 0.6350898 1.001498   540
#> [13,] -0.092518964 0.3518898 -0.8049151 0.5531263 1.002048   520
#> [14,] -0.126703212 0.3350137 -0.7457144 0.5635215 1.001782   540
#> [15,] -0.141012151 0.3554389 -0.9073636 0.5759955 1.012243   540
#> [16,] -0.030581982 0.3195192 -0.6570607 0.6398148 1.004069   540
#> [17,] -0.053864174 0.3413322 -0.6856778 0.5904572 1.001427   540
#> [18,] -0.068264509 0.3320227 -0.7136043 0.4527978 1.050937    80
#> [19,] -0.080997800 0.3455610 -0.7697947 0.5102602 1.008920   540
#> [20,] -0.119264599 0.3108864 -0.7524038 0.5490882 1.011098   540
#> [21,]  0.008685001 0.3380453 -0.6979558 0.7472470 1.006957   410
#> [22,] -0.027116716 0.3438177 -0.8123260 0.6809324 1.018648   380
#> [23,] -0.053543023 0.3106376 -0.6712949 0.5325794 1.006098   540
#> [24,] -0.045154840 0.3455568 -0.7630557 0.6102939 1.005687   350
#> [25,] -0.098466262 0.3464263 -0.6899871 0.6720330 1.001470   540
#> 
#> $tau
#>             50%        sd        2.5%    97.5%     Rhat n.eff
#>  [1,] 0.3660508 0.4088546 0.014687164 1.589614 1.063336    46
#>  [2,] 0.3494056 0.3345529 0.010212456 1.296872 1.111666    36
#>  [3,] 0.3193976 0.3484744 0.002624766 1.252180 1.165057    64
#>  [4,] 0.4255242 0.3807480 0.045090968 1.469935 1.088233    39
#>  [5,] 0.3733927 0.3467616 0.032137122 1.345188 1.061235    51
#>  [6,] 0.3531398 0.3452319 0.034135583 1.320969 1.013815   260
#>  [7,] 0.3905228 0.3379946 0.033923611 1.358248 1.079000   150
#>  [8,] 0.3474821 0.2877534 0.031174873 1.057108 1.022942   540
#>  [9,] 0.4010625 0.3453214 0.031017981 1.323253 1.059023    68
#> [10,] 0.3793832 0.3333521 0.020439556 1.296396 1.115833    71
#> [11,] 0.3035266 0.3491330 0.011646150 1.252397 1.029512   330
#> [12,] 0.3958687 0.3624047 0.042422687 1.457937 1.085284    71
#> [13,] 0.3298260 0.3668705 0.018245765 1.409908 1.068263   280
#> [14,] 0.3722521 0.3041678 0.022795608 1.199338 1.011500   280
#> [15,] 0.3948727 0.3579650 0.015430750 1.375513 1.134297    75
#> [16,] 0.3375628 0.3368394 0.021530381 1.348472 1.034221    71
#> [17,] 0.3510145 0.3353337 0.008197244 1.291084 1.051941    92
#> [18,] 0.2495769 0.2906578 0.015672277 1.061323 1.047113    46
#> [19,] 0.3278715 0.3241972 0.017904040 1.196750 1.071601    42
#> [20,] 0.3212365 0.3201065 0.011732464 1.177220 1.038196    83
#> [21,] 0.3423693 0.3062909 0.014579947 1.180313 1.051877   120
#> [22,] 0.3232916 0.3371782 0.005487719 1.299682 1.160708    33
#> [23,] 0.3255761 0.3571095 0.004130013 1.303222 1.086277    42
#> [24,] 0.4003042 0.3449394 0.029797815 1.369456 1.011593   290
#> [25,] 0.3556000 0.3283055 0.020705534 1.315278 1.018489   150
#> 
#> $measure
#> [1] "SMD"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] 0 1 1
#> 
# }