Perform sensitivity analysis for missing participant outcome data

Performs a sensitivity analysis by applying pairwise meta-analysis or network meta-analysis for a series of different scenarios about the informative missingness parameter.

Usage

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin,
  inits = NULL
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following two: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length equal to 5 or larger. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default scenarios are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04.

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects pairwise meta-analysis:

EM

The estimated summary effect measure (according to the argument measure defined in run_model).

EM_LOR

The estimated summary odd ratio in the logarithmic scale when measure = "RR" or measure = "RD".

tau

The between-trial standard deviation. This element does not appear in the case of a fixed-effect pairwise meta-analysis.

In a random-effects network meta-analysis, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore,

tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the vector specified in mean_scenarios (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package to be processed further and provide an end-user-ready output.

In the case of pairwise meta-analysis, EM and tau are estimated as many times as the number of scenarios considered. In the case of network meta-analysis, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console.

Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to Spineli (2019), Mavridis et al. (2015), Turner et al. (2015), and White et al. (2008) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, heter_prior, D, indic, base_risk, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model)--an exception in the assumption argument as it is restricted to only two character strings. Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

The run_sensitivity function also returns the arguments measure, scenarios, D, heter, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--41. doi: 10.1002/sim.6365

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--90. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--80. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--27. doi: 10.1002/sim.3008

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (default arguments)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The default scenarios were considered: c(-2, -1, 0, 1, 2).
#> JAGS generates initial values for the parameters.
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 1 until convergence
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 2 until convergence
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 3 until convergence
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 4 until convergence
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 5 until convergence
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 6 until convergence
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 7 until convergence
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 8 until convergence
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 9 until convergence
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 10 until convergence
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 11 until convergence
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 12 until convergence
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 13 until convergence
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 14 until convergence
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 15 until convergence
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 16 until convergence
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 17 until convergence
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 18 until convergence
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 19 until convergence
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 20 until convergence
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 21 until convergence
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 22 until convergence
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 23 until convergence
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 24 until convergence
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 25 until convergence
#> $EM
#>               mean        sd       2.5%        25%          50%           75%
#>  [1,] -0.083767263 0.2660897 -0.5871490 -0.2553942 -0.083441407  0.0835259239
#>  [2,] -0.135373191 0.2500720 -0.6039926 -0.2997241 -0.130400483  0.0255714434
#>  [3,] -0.156571209 0.2458816 -0.6217541 -0.3053434 -0.152549025 -0.0099799074
#>  [4,] -0.169645775 0.2557590 -0.7478970 -0.3112498 -0.157961571 -0.0196033369
#>  [5,] -0.184007973 0.2523523 -0.6783044 -0.3389226 -0.178944274 -0.0173934378
#>  [6,] -0.085257996 0.2623760 -0.6162234 -0.2399078 -0.081229272  0.0773023037
#>  [7,] -0.102998272 0.2603404 -0.6041621 -0.2745696 -0.097456761  0.0691334098
#>  [8,] -0.123479467 0.2619242 -0.6422429 -0.2680094 -0.131704385  0.0364263282
#>  [9,] -0.159042253 0.2701517 -0.7842676 -0.2944719 -0.143911473  0.0020470485
#> [10,] -0.144296664 0.2734450 -0.7057398 -0.2877330 -0.152682262 -0.0006615422
#> [11,] -0.017262325 0.2670952 -0.4761185 -0.1811843 -0.016252339  0.1330684670
#> [12,] -0.077194556 0.2630689 -0.5565626 -0.2306088 -0.086969849  0.0873986084
#> [13,] -0.118347223 0.2621878 -0.5905706 -0.2641378 -0.124469188  0.0348279578
#> [14,] -0.117517119 0.2596380 -0.6538857 -0.2713124 -0.112236051  0.0636435951
#> [15,] -0.123291194 0.2321049 -0.6144391 -0.2413051 -0.120469059  0.0155817437
#> [16,] -0.028967765 0.2374524 -0.4569183 -0.1956619 -0.035976472  0.1080036141
#> [17,] -0.033278582 0.2393568 -0.5090799 -0.1876118 -0.032801472  0.1069492587
#> [18,] -0.041006169 0.2704348 -0.6299697 -0.2015224 -0.016276617  0.1369826806
#> [19,] -0.076615331 0.2833405 -0.5950304 -0.2429014 -0.069663141  0.0732077683
#> [20,] -0.105457060 0.2582874 -0.6177022 -0.2685833 -0.092145060  0.0539712222
#> [21,] -0.006523787 0.2441667 -0.4986136 -0.1664051  0.009223824  0.1444706392
#> [22,] -0.003757884 0.2539070 -0.4550146 -0.1902224  0.018879189  0.1619549574
#> [23,] -0.053176183 0.2779952 -0.5171080 -0.2275229 -0.060890400  0.1080152886
#> [24,] -0.064139485 0.2567573 -0.5034721 -0.2413353 -0.070952482  0.0961894542
#> [25,] -0.079532787 0.2514462 -0.6160589 -0.2273448 -0.075447492  0.0723120874
#>           97.5%      Rhat n.eff
#>  [1,] 0.4255672 1.0127876   260
#>  [2,] 0.3597753 1.0087629   240
#>  [3,] 0.3043359 1.0015006   600
#>  [4,] 0.2659810 1.0015976   600
#>  [5,] 0.3200412 1.0132486   600
#>  [6,] 0.4453039 1.0033652   410
#>  [7,] 0.3655557 1.0003083   600
#>  [8,] 0.4071642 1.0143294   600
#>  [9,] 0.3318797 1.0181557   160
#> [10,] 0.3573721 1.0148546   380
#> [11,] 0.4818761 1.0095298   190
#> [12,] 0.3887805 0.9988638   600
#> [13,] 0.3770747 0.9990815   600
#> [14,] 0.3543751 1.0067869   600
#> [15,] 0.3113384 1.0084148   340
#> [16,] 0.4391679 1.0026613   480
#> [17,] 0.4274605 1.0047682   320
#> [18,] 0.4420873 1.0022778   600
#> [19,] 0.4648740 1.0080744   410
#> [20,] 0.3802614 1.0024037   600
#> [21,] 0.4456060 1.0160127   240
#> [22,] 0.4939045 1.0021145   550
#> [23,] 0.5273180 1.0024649   500
#> [24,] 0.4284508 1.0123574   320
#> [25,] 0.3724498 1.0076068   490
#> 
#> $tau
#>            mean        sd        2.5%        25%       50%       75%     97.5%
#>  [1,] 0.2140638 0.2113879 0.010503957 0.07453078 0.1534500 0.2789046 0.7232777
#>  [2,] 0.2533746 0.2373810 0.013461874 0.08853451 0.1882525 0.3263761 0.9028062
#>  [3,] 0.2285916 0.2536187 0.008528579 0.08257352 0.1669957 0.3026568 0.8066088
#>  [4,] 0.2407443 0.1984416 0.012930526 0.10170221 0.1832201 0.3424973 0.7224866
#>  [5,] 0.2381770 0.2279607 0.011169537 0.09202832 0.1866613 0.3102785 0.8100560
#>  [6,] 0.2222402 0.1973194 0.009847599 0.07596633 0.1661309 0.3188870 0.7103917
#>  [7,] 0.2190736 0.1980064 0.007348719 0.08283416 0.1665810 0.2923180 0.7053675
#>  [8,] 0.2410638 0.2175789 0.012990213 0.08503178 0.1921475 0.3193150 0.9039997
#>  [9,] 0.2407883 0.2244062 0.007543975 0.08001305 0.1740062 0.3308362 0.8157748
#> [10,] 0.2434443 0.2573367 0.009324447 0.08083154 0.1649251 0.3155368 0.8460848
#> [11,] 0.2231697 0.2306704 0.015722844 0.08161744 0.1596850 0.2764771 0.8799915
#> [12,] 0.2453558 0.2520949 0.008573179 0.09023378 0.1867532 0.3332556 0.7673977
#> [13,] 0.2324852 0.2181758 0.013723350 0.08361607 0.1737863 0.3268266 0.8205538
#> [14,] 0.2250647 0.2092751 0.012082158 0.07774385 0.1655311 0.3080694 0.8512765
#> [15,] 0.2256121 0.2134246 0.017168535 0.07939105 0.1673514 0.2929849 0.7267891
#> [16,] 0.1957288 0.1724942 0.013531851 0.07240843 0.1491581 0.2700078 0.6703720
#> [17,] 0.2194047 0.2207308 0.017404072 0.07924889 0.1653957 0.2867934 0.7003202
#> [18,] 0.2380276 0.2317014 0.006788260 0.08152290 0.1723599 0.3222653 0.8387800
#> [19,] 0.2387290 0.2453093 0.012160294 0.08865527 0.1741473 0.3134959 0.8421131
#> [20,] 0.2490903 0.2597332 0.020681725 0.09492346 0.1820908 0.3026568 0.9741500
#> [21,] 0.2252942 0.1937180 0.021046557 0.08838774 0.1714358 0.3038312 0.7460939
#> [22,] 0.2181207 0.2000347 0.010640094 0.07727056 0.1699001 0.2869206 0.7585545
#> [23,] 0.2528960 0.2395219 0.018176468 0.10275769 0.1882447 0.3233727 0.8577428
#> [24,] 0.2429186 0.2053908 0.017816191 0.10276279 0.1902846 0.3297238 0.7495462
#> [25,] 0.2374162 0.2151635 0.010690259 0.09236478 0.1809893 0.3250247 0.7203618
#>           Rhat n.eff
#>  [1,] 1.008330   490
#>  [2,] 1.034649   100
#>  [3,] 1.008304   350
#>  [4,] 1.020929   120
#>  [5,] 1.072795    50
#>  [6,] 1.002455   600
#>  [7,] 1.041891   220
#>  [8,] 1.066613    40
#>  [9,] 1.022801   140
#> [10,] 1.012524   600
#> [11,] 1.009736   180
#> [12,] 1.013631   600
#> [13,] 1.030074   140
#> [14,] 1.031110   150
#> [15,] 1.018159   180
#> [16,] 1.010821   410
#> [17,] 1.011519   260
#> [18,] 1.038130    67
#> [19,] 1.019720   230
#> [20,] 1.036726   100
#> [21,] 1.031107   100
#> [22,] 1.007445   600
#> [23,] 1.024303   480
#> [24,] 1.011560   300
#> [25,] 1.046659    89
#> 
#> $measure
#> [1] "SMD"
#> 
#> $model
#> [1] "RE"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] -2.9900000  0.2143347  4.0000000
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 5
#> 
#> attr(,"class")
#> [1] "run_sensitivity"
# }