Perform sensitivity analysis for missing participant outcome data

Performs a sensitivity analysis by applying pairwise meta-analysis or network meta-analysis for a series of different scenarios about the informative missingness parameter.

Usage

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin,
  inits = NULL
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following two: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length equal to 5 or larger. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default scenarios are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04.

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects pairwise meta-analysis:

EM

The estimated summary effect measure (according to the argument measure defined in run_model).

EM_LOR

The estimated summary odd ratio in the logarithmic scale when measure = "RR" or measure = "RD".

tau

The between-trial standard deviation. This element does not appear in the case of a fixed-effect pairwise meta-analysis.

In a random-effects network meta-analysis, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore,

tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the vector specified in mean_scenarios (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package to be processed further and provide an end-user-ready output.

In the case of pairwise meta-analysis, EM and tau are estimated as many times as the number of scenarios considered. In the case of network meta-analysis, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console.

Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to Spineli (2019), Mavridis et al. (2015), Turner et al. (2015), and White et al. (2008) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, heter_prior, D, indic, base_risk, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model)--an exception in the assumption argument as it is restricted to only two character strings. Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

The run_sensitivity function also returns the arguments measure, scenarios, D, heter, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--41. doi: 10.1002/sim.6365

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--90. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--80. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--27. doi: 10.1002/sim.3008

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (default arguments)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The default scenarios were considered: c(-2, -1, 0, 1, 2).
#> JAGS generates initial values for the parameters.
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 1 until convergence
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 2 until convergence
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 3 until convergence
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 4 until convergence
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 5 until convergence
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 6 until convergence
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 7 until convergence
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 8 until convergence
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 9 until convergence
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 10 until convergence
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 11 until convergence
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 12 until convergence
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 13 until convergence
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 14 until convergence
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 15 until convergence
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 16 until convergence
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 17 until convergence
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 18 until convergence
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 19 until convergence
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 20 until convergence
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 21 until convergence
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 22 until convergence
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 23 until convergence
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 24 until convergence
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 25 until convergence
#> $EM
#>               mean        sd       2.5%        25%          50%          75%
#>  [1,] -0.126868061 0.2483807 -0.5874043 -0.2914170 -0.126838836  0.028910852
#>  [2,] -0.156652587 0.2731867 -0.6559197 -0.3239737 -0.160249946 -0.005865364
#>  [3,] -0.147190724 0.2632825 -0.6734893 -0.3040218 -0.145639931  0.002315553
#>  [4,] -0.137366563 0.2639888 -0.7076403 -0.2893586 -0.133342071  0.013455758
#>  [5,] -0.147448979 0.2922438 -0.7382995 -0.3098753 -0.144530869  0.015784605
#>  [6,] -0.098134663 0.2248005 -0.5198390 -0.2504843 -0.108809156  0.052290042
#>  [7,] -0.110566343 0.2345203 -0.5568618 -0.2606091 -0.110675518  0.031001067
#>  [8,] -0.112632147 0.2653473 -0.6052136 -0.2742903 -0.109498125  0.046252995
#>  [9,] -0.140513406 0.2459981 -0.5810040 -0.3110653 -0.139159368  0.002565925
#> [10,] -0.147281770 0.2229947 -0.6308304 -0.2889199 -0.140972290 -0.004898091
#> [11,] -0.088647698 0.2488149 -0.5480487 -0.2466818 -0.092662356  0.050842782
#> [12,] -0.081998377 0.2419031 -0.5559767 -0.2379612 -0.058857984  0.079118319
#> [13,] -0.094144487 0.2545997 -0.5618141 -0.2411930 -0.103693857  0.049077535
#> [14,] -0.122279296 0.2684896 -0.5976421 -0.2964534 -0.116124525  0.055219629
#> [15,] -0.107711301 0.2546929 -0.5960821 -0.2554727 -0.096433774  0.060240961
#> [16,] -0.028219597 0.2572366 -0.5729169 -0.1714281 -0.016086579  0.126579659
#> [17,] -0.030531502 0.2662812 -0.5391686 -0.1844322 -0.029688490  0.140821669
#> [18,] -0.072805625 0.2511698 -0.5766781 -0.2211247 -0.074620229  0.080415044
#> [19,] -0.094105355 0.2532346 -0.6045385 -0.2450031 -0.090793522  0.040551730
#> [20,] -0.101284999 0.2908543 -0.6086152 -0.2724504 -0.113622017  0.057569343
#> [21,]  0.003687165 0.2598106 -0.4690264 -0.1390402 -0.003532221  0.151666347
#> [22,]  0.017375064 0.2557016 -0.4389851 -0.1561326  0.007207158  0.190169603
#> [23,] -0.042890357 0.2362887 -0.5309148 -0.1847718 -0.049623302  0.104154183
#> [24,] -0.087606028 0.2343723 -0.5770373 -0.2385587 -0.089281046  0.052084404
#> [25,] -0.102180486 0.2310989 -0.5356530 -0.2615373 -0.090591671  0.060187260
#>           97.5%      Rhat n.eff
#>  [1,] 0.3612938 1.0086206   600
#>  [2,] 0.4146834 1.0091545   310
#>  [3,] 0.3895244 1.0101417   600
#>  [4,] 0.3625747 1.0039228   590
#>  [5,] 0.4513369 1.0062037   600
#>  [6,] 0.3409378 1.0203528    97
#>  [7,] 0.3649391 1.0027477   600
#>  [8,] 0.3513826 1.0078006   600
#>  [9,] 0.3910688 1.0199430   130
#> [10,] 0.2644625 1.0096875   250
#> [11,] 0.3935290 1.0192568   180
#> [12,] 0.3418544 1.0091487   240
#> [13,] 0.4063994 1.0036813   600
#> [14,] 0.3533458 1.0030554   600
#> [15,] 0.3526535 1.0041928   390
#> [16,] 0.4599752 1.0066982   320
#> [17,] 0.4537877 1.0161145   260
#> [18,] 0.4294313 1.0016274   600
#> [19,] 0.3838061 0.9992933   600
#> [20,] 0.4547121 1.0108344   370
#> [21,] 0.5185603 1.0179995   600
#> [22,] 0.5111636 1.0016699   600
#> [23,] 0.4290211 1.0076891   430
#> [24,] 0.3729560 1.0075155   500
#> [25,] 0.3352876 1.0100016   220
#> 
#> $tau
#>            mean        sd        2.5%        25%       50%       75%     97.5%
#>  [1,] 0.2353274 0.2393375 0.012379643 0.08622826 0.1772553 0.3069768 0.8514724
#>  [2,] 0.2702278 0.2352457 0.015966882 0.10474786 0.2090684 0.3708562 0.8520584
#>  [3,] 0.2475160 0.2500663 0.009594359 0.08381794 0.1826293 0.3413602 0.7477012
#>  [4,] 0.2585011 0.2359767 0.013225225 0.08850714 0.1996518 0.3600033 0.8707806
#>  [5,] 0.3022205 0.2900896 0.021495693 0.11298631 0.2248686 0.3954077 1.0142296
#>  [6,] 0.2200616 0.2073074 0.007043508 0.07617129 0.1690944 0.2886187 0.7176699
#>  [7,] 0.2171498 0.1879602 0.012094637 0.07811882 0.1669114 0.3002449 0.7133262
#>  [8,] 0.2405239 0.2385161 0.011224803 0.08770636 0.1701764 0.3214205 0.8223803
#>  [9,] 0.2508116 0.2284168 0.012999969 0.09927421 0.1956019 0.3288813 0.8527621
#> [10,] 0.2102437 0.1931612 0.011693555 0.06725771 0.1579977 0.2971573 0.6721487
#> [11,] 0.2142818 0.1990255 0.015500232 0.08352856 0.1594617 0.2740434 0.7148224
#> [12,] 0.2224845 0.1893000 0.009521880 0.08480907 0.1782677 0.3117029 0.6975844
#> [13,] 0.2426882 0.2280687 0.014539442 0.08822304 0.1878094 0.3143724 0.8036938
#> [14,] 0.2259916 0.2333470 0.008933877 0.07240211 0.1646009 0.2981061 0.8172443
#> [15,] 0.2613757 0.2566222 0.018794676 0.09545860 0.1909491 0.3384522 0.9039183
#> [16,] 0.2520568 0.2221634 0.008586992 0.09733473 0.1885281 0.3313080 0.8459220
#> [17,] 0.2337712 0.2243652 0.010885164 0.07370715 0.1736608 0.3086304 0.8516948
#> [18,] 0.2289584 0.2212926 0.004446070 0.07699892 0.1784542 0.3149394 0.8368228
#> [19,] 0.2371755 0.2153407 0.026806142 0.09681053 0.1878187 0.3052490 0.8283111
#> [20,] 0.2555035 0.2329203 0.018399284 0.10474883 0.1883110 0.3432139 0.8288999
#> [21,] 0.2353204 0.2027090 0.017396713 0.09654750 0.1863620 0.3348966 0.6769980
#> [22,] 0.2312851 0.2096131 0.014645088 0.08072063 0.1714076 0.3189449 0.7624886
#> [23,] 0.2164837 0.1981168 0.018658853 0.07697346 0.1614202 0.2841582 0.7500869
#> [24,] 0.2335187 0.2223538 0.012106704 0.08949196 0.1715633 0.3057768 0.8380116
#> [25,] 0.2175742 0.2076305 0.011591307 0.08360004 0.1679920 0.2881772 0.7027327
#>           Rhat n.eff
#>  [1,] 1.042887    84
#>  [2,] 1.004608   600
#>  [3,] 1.047657   180
#>  [4,] 1.006899   580
#>  [5,] 1.015582   210
#>  [6,] 1.031163   120
#>  [7,] 1.031388   100
#>  [8,] 1.017141   470
#>  [9,] 1.083574    53
#> [10,] 1.016063   160
#> [11,] 1.008694   230
#> [12,] 1.002295   600
#> [13,] 1.006042   380
#> [14,] 1.069196   130
#> [15,] 1.019773   410
#> [16,] 1.035124   230
#> [17,] 1.079036    65
#> [18,] 1.002449   600
#> [19,] 1.020792    94
#> [20,] 1.027202   100
#> [21,] 1.021312   220
#> [22,] 1.093949    30
#> [23,] 1.003992   400
#> [24,] 1.056357    59
#> [25,] 1.025507   150
#> 
#> $measure
#> [1] "SMD"
#> 
#> $model
#> [1] "RE"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] -2.9900000  0.2143347  4.0000000
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 5
#> 
#> attr(,"class")
#> [1] "run_sensitivity"
# }