Perform sensitivity analysis for aggregate missing outcome participant data

Performs a sensitivity analysis by applying pairwise meta-analysis (PMA) or network meta-analysis (NMA) for a series of different scenarios about the informative missingness parameter,

Usage,

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length of 5 or larger positive odd integer. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default arguments are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects PMA:

EM	The estimated summary effect measure (according to the argument measure defined in run_model).

tau	The between-trial standard deviation. This element does not appear in the case of a fixed-effect PMA.

In a random-effects NMA, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore, tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the mean_scenarios vector (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package function to be processed further and provide an end-user-ready output.

In the case of PMA, EM and tau have as many rows as the number of scenarios considered. In the case of NMA, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console. Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to White et al., (2008), Mavridis et al., (2015), Turner et al., (2015) and Spineli (2019) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, and heter_prior from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model). Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--490. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--741. doi: 10.1002/sim.6365

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--2080. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--727. doi: 10.1002/sim.3008

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci. 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (missing-at-random assumption)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The following vector of scenarios was considered by default: c(-2, -1, 0, 1, 2)
#> 
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 579
#> 
#> Initializing model
#> 
#> $EM
#>               mean        sd       2.5%     97.5%      Rhat n.eff
#>  [1,] -0.108803584 0.3272870 -0.7133082 0.6110102 1.0162731   300
#>  [2,] -0.116111243 0.3828628 -0.8563185 0.6903806 1.0157946   130
#>  [3,] -0.161448896 0.3506679 -0.8993845 0.5674655 1.0030569   540
#>  [4,] -0.167424568 0.3227193 -0.7903322 0.4784678 1.0019165   540
#>  [5,] -0.165449303 0.3471830 -0.8316135 0.5454656 0.9991067   540
#>  [6,] -0.069078097 0.3274673 -0.8041599 0.5001052 1.0042058   540
#>  [7,] -0.070934984 0.3758080 -0.7751873 0.6097643 1.0037976   540
#>  [8,] -0.111825851 0.3289134 -0.7228483 0.4680301 1.0098474   540
#>  [9,] -0.138129962 0.3131168 -0.7834489 0.3686549 1.0033738   510
#> [10,] -0.172791327 0.3399765 -0.8595749 0.4415606 1.0054141   540
#> [11,] -0.041287516 0.3803176 -0.7725274 0.6063716 1.0020873   520
#> [12,] -0.096729086 0.3321786 -0.8237123 0.6032001 1.0090293   190
#> [13,] -0.100508861 0.3224800 -0.7250295 0.4844738 1.0067415   540
#> [14,] -0.146701390 0.3913503 -1.0291685 0.6735324 1.0073441   270
#> [15,] -0.130852158 0.3510039 -0.8091057 0.5379830 1.0091528   540
#> [16,] -0.021191209 0.3247978 -0.6503623 0.5943710 1.0047106   310
#> [17,] -0.054776769 0.3601420 -0.7389206 0.6577104 1.0131983   540
#> [18,] -0.044106725 0.3451842 -0.7198784 0.6854731 1.0006466   540
#> [19,] -0.092173157 0.3276107 -0.7904491 0.5727071 1.0151388   540
#> [20,] -0.117440398 0.3295779 -0.8072133 0.5387512 1.0172763   120
#> [21,]  0.005220164 0.3537156 -0.7532428 0.6739723 1.0026351   540
#> [22,] -0.040261704 0.3642946 -0.7814001 0.6631947 1.0152996   540
#> [23,] -0.044338559 0.3264769 -0.6503771 0.6209831 1.0009065   540
#> [24,] -0.090032853 0.3545060 -0.8458850 0.5654784 1.0002727   540
#> [25,] -0.080096160 0.3355294 -0.7604731 0.6246676 0.9991101   540
#> 
#> $tau
#>             50%        sd        2.5%    97.5%      Rhat n.eff
#>  [1,] 0.3859109 0.3219764 0.029084205 1.208457 1.0844292    33
#>  [2,] 0.3982033 0.3720631 0.060957338 1.490563 1.0355847    74
#>  [3,] 0.3904163 0.3763115 0.052803979 1.479407 1.0176786   170
#>  [4,] 0.3521211 0.3376999 0.033393676 1.305101 0.9991542   540
#>  [5,] 0.3943660 0.3434800 0.037927240 1.346006 1.0685965    53
#>  [6,] 0.3388850 0.3607345 0.025791463 1.375591 1.0120107   150
#>  [7,] 0.3712631 0.3510807 0.021032184 1.416171 0.9987102   540
#>  [8,] 0.3527803 0.3381296 0.022551708 1.316366 1.0087874   190
#>  [9,] 0.3269240 0.3254637 0.009401379 1.200243 1.0336255   140
#> [10,] 0.4177102 0.3275255 0.035944864 1.246057 1.0835343    55
#> [11,] 0.3867310 0.3869928 0.031456192 1.592117 1.0005464   540
#> [12,] 0.3759842 0.3912007 0.020626713 1.501968 1.0126025   160
#> [13,] 0.3470181 0.3363449 0.031111602 1.307874 1.0333947    89
#> [14,] 0.3421189 0.4192514 0.011479391 1.568383 1.0170004   130
#> [15,] 0.3674272 0.3756348 0.027400509 1.426960 1.0154854   120
#> [16,] 0.3409362 0.3199313 0.031260355 1.257801 1.0093805   540
#> [17,] 0.3549927 0.3904326 0.023272247 1.447510 1.0101779   170
#> [18,] 0.3041262 0.3252388 0.009153277 1.085484 1.0253680   150
#> [19,] 0.3408002 0.3409928 0.015060018 1.292369 1.1040525    29
#> [20,] 0.3734348 0.3103619 0.036837902 1.178908 1.0151187   190
#> [21,] 0.3322135 0.3084005 0.023091832 1.183026 1.0149720   540
#> [22,] 0.3533150 0.3354422 0.031581906 1.248543 1.0237728   150
#> [23,] 0.3482832 0.3785793 0.009728342 1.401872 1.0089652   540
#> [24,] 0.3723118 0.3468312 0.023474270 1.304160 1.0794303    47
#> [25,] 0.3413319 0.3274066 0.012552168 1.231973 1.0374753   140
#> 
#> $measure
#> [1] "SMD"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] 0 1 1
#> 
# }