Perform sensitivity analysis for missing participant outcome data

Performs a sensitivity analysis by applying pairwise meta-analysis or network meta-analysis for a series of different scenarios about the informative missingness parameter.

Usage

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin,
  inits = NULL
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following two: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length equal to 5 or larger. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default scenarios are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04.

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects pairwise meta-analysis:

EM

The estimated summary effect measure (according to the argument measure defined in run_model).

EM_LOR

The estimated summary odd ratio in the logarithmic scale when measure = "RR" or measure = "RD".

tau

The between-trial standard deviation. This element does not appear in the case of a fixed-effect pairwise meta-analysis.

In a random-effects network meta-analysis, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore,

tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the vector specified in mean_scenarios (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package to be processed further and provide an end-user-ready output.

In the case of pairwise meta-analysis, EM and tau are estimated as many times as the number of scenarios considered. In the case of network meta-analysis, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console.

Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to Spineli (2019), Mavridis et al. (2015), Turner et al. (2015), and White et al. (2008) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, heter_prior, D, indic, base_risk, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model)--an exception in the assumption argument as it is restricted to only two character strings. Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

The run_sensitivity function also returns the arguments measure, scenarios, D, heter, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--41. doi: 10.1002/sim.6365

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--90. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--80. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--27. doi: 10.1002/sim.3008

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (default arguments)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The default scenarios were considered: c(-2, -1, 0, 1, 2).
#> JAGS generates initial values for the parameters.
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 1 until convergence
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 2 until convergence
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 3 until convergence
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 4 until convergence
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 5 until convergence
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 6 until convergence
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 7 until convergence
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 8 until convergence
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 9 until convergence
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 10 until convergence
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 11 until convergence
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 12 until convergence
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 13 until convergence
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 14 until convergence
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 15 until convergence
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 16 until convergence
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 17 until convergence
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 18 until convergence
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 19 until convergence
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 20 until convergence
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 21 until convergence
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 22 until convergence
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 23 until convergence
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 24 until convergence
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 339
#> 
#> Initializing model
#> 
#> Updating model for scenario 25 until convergence
#> $EM
#>               mean        sd       2.5%        25%          50%          75%
#>  [1,] -0.123410739 0.2624823 -0.5908456 -0.2801769 -0.113520285  0.040614870
#>  [2,] -0.142360678 0.2384828 -0.5865343 -0.3105050 -0.161290079  0.022419022
#>  [3,] -0.125136784 0.2319786 -0.6163227 -0.2717930 -0.129475365  0.024113836
#>  [4,] -0.151936203 0.2657977 -0.6799140 -0.3141361 -0.151949078 -0.001165675
#>  [5,] -0.188243874 0.2417977 -0.6404716 -0.3431395 -0.191920586 -0.039347075
#>  [6,] -0.091665658 0.2600287 -0.5982461 -0.2462623 -0.086617501  0.052583247
#>  [7,] -0.125380626 0.2626243 -0.6355222 -0.2775848 -0.142144999  0.026137932
#>  [8,] -0.117708533 0.2588260 -0.5998263 -0.2896732 -0.099980714  0.035127610
#>  [9,] -0.138609273 0.2759603 -0.6182894 -0.3084618 -0.144003948  0.018805177
#> [10,] -0.139790041 0.2623561 -0.6878858 -0.3032018 -0.129558331  0.029959531
#> [11,] -0.055286151 0.2457820 -0.5055359 -0.2235960 -0.067632220  0.115317158
#> [12,] -0.058203635 0.2476484 -0.5348826 -0.1937499 -0.055619801  0.091536727
#> [13,] -0.139229915 0.2510009 -0.5903143 -0.2777408 -0.158073053  0.026374691
#> [14,] -0.106514711 0.2563482 -0.5925325 -0.2754599 -0.099343011  0.059270411
#> [15,] -0.143864590 0.2633377 -0.6875889 -0.3018264 -0.140400834  0.005516509
#> [16,] -0.023815033 0.2501131 -0.4797322 -0.1753484 -0.030315848  0.125594519
#> [17,] -0.069760779 0.2384688 -0.5581775 -0.2212006 -0.060329445  0.086894371
#> [18,] -0.086906477 0.2332074 -0.5903589 -0.2324846 -0.084871209  0.062388755
#> [19,] -0.082003581 0.2567979 -0.6039356 -0.2177873 -0.068106931  0.064560711
#> [20,] -0.106466842 0.2497097 -0.5709392 -0.2660129 -0.112195804  0.044900783
#> [21,]  0.018043740 0.2657343 -0.4750835 -0.1368244 -0.008834574  0.165025667
#> [22,]  0.002338685 0.2565763 -0.4800746 -0.1611044 -0.002349104  0.158300738
#> [23,] -0.052475100 0.2801628 -0.6292366 -0.2207755 -0.044660326  0.115433601
#> [24,] -0.034962188 0.2472967 -0.5129293 -0.1845139 -0.049629947  0.101758601
#> [25,] -0.069307698 0.2668649 -0.5636252 -0.2383966 -0.089116879  0.073206892
#>           97.5%      Rhat n.eff
#>  [1,] 0.3316879 1.0156230   540
#>  [2,] 0.3181202 1.0026872   600
#>  [3,] 0.3196181 1.0017656   600
#>  [4,] 0.3950377 1.0260161   110
#>  [5,] 0.2994264 1.0040198   360
#>  [6,] 0.3833743 1.0222239   120
#>  [7,] 0.3937248 1.0165106   140
#>  [8,] 0.3762619 1.0132237   190
#>  [9,] 0.4025799 1.0115311   400
#> [10,] 0.3378291 1.0105044   440
#> [11,] 0.4323942 1.0013979   600
#> [12,] 0.4024076 1.0036694   410
#> [13,] 0.3192981 1.0129569   200
#> [14,] 0.3595255 1.0070479   300
#> [15,] 0.3580489 1.0111666   300
#> [16,] 0.4604506 1.0144115   170
#> [17,] 0.3579769 1.0030166   460
#> [18,] 0.3613182 1.0052086   300
#> [19,] 0.3554673 1.0053551   600
#> [20,] 0.3732656 1.0000882   600
#> [21,] 0.5592587 1.0391890    67
#> [22,] 0.5003945 1.0032668   600
#> [23,] 0.4689314 1.0076793   370
#> [24,] 0.4662939 0.9986754   600
#> [25,] 0.4988956 1.0059894   270
#> 
#> $tau
#>            mean        sd        2.5%        25%       50%       75%     97.5%
#>  [1,] 0.2411508 0.2324079 0.015489481 0.08688814 0.1824736 0.3166276 0.8245045
#>  [2,] 0.2465842 0.1973090 0.021869778 0.10800153 0.1953613 0.3362576 0.7195314
#>  [3,] 0.2168785 0.2053213 0.016872975 0.08850182 0.1641831 0.2799730 0.7386435
#>  [4,] 0.2409531 0.2026575 0.010784970 0.09087212 0.1855950 0.3470224 0.7431461
#>  [5,] 0.2356008 0.1986779 0.015279212 0.09148202 0.1766207 0.3278563 0.7419125
#>  [6,] 0.2560399 0.2178586 0.010679450 0.09645729 0.2016976 0.3532917 0.8714202
#>  [7,] 0.2430204 0.2092834 0.012793373 0.09963418 0.1900955 0.3261266 0.7753394
#>  [8,] 0.2262102 0.1965060 0.012517650 0.08256628 0.1771229 0.2977152 0.7915866
#>  [9,] 0.2376424 0.2989342 0.019611394 0.08539900 0.1801120 0.2961634 0.7721159
#> [10,] 0.2482062 0.2252105 0.010741141 0.08687758 0.1866415 0.3388781 0.7843995
#> [11,] 0.2493179 0.2473753 0.019213527 0.09413299 0.1781801 0.3277404 0.8228511
#> [12,] 0.2081066 0.1909172 0.012606426 0.07522018 0.1520496 0.2791916 0.7011768
#> [13,] 0.2182642 0.2103717 0.009097648 0.07182843 0.1566615 0.3012547 0.7939756
#> [14,] 0.2519082 0.2372194 0.013933579 0.09545510 0.1917106 0.3374826 0.9032712
#> [15,] 0.2442384 0.2402715 0.019922985 0.08444390 0.1751508 0.3283337 0.9460265
#> [16,] 0.2243152 0.2799305 0.016968954 0.08343413 0.1582209 0.2965523 0.7184983
#> [17,] 0.2363119 0.2116697 0.009189573 0.08574189 0.1784029 0.3209222 0.7637594
#> [18,] 0.2085531 0.1909043 0.007437119 0.07587946 0.1561639 0.2750506 0.7316995
#> [19,] 0.2338700 0.2124298 0.006717186 0.08548037 0.1772311 0.3032444 0.8207581
#> [20,] 0.2496360 0.2144635 0.014429177 0.10030830 0.1968482 0.3232600 0.8506654
#> [21,] 0.2286444 0.2367394 0.012568607 0.08334353 0.1609740 0.3000530 0.8264884
#> [22,] 0.2262944 0.2124349 0.017383401 0.08204244 0.1650555 0.3085140 0.7935618
#> [23,] 0.2404374 0.2597586 0.017236147 0.08579974 0.1766629 0.3104541 0.8106209
#> [24,] 0.2144877 0.2145049 0.008373151 0.07784919 0.1515944 0.2860897 0.7127945
#> [25,] 0.2227022 0.1803482 0.021031133 0.09518765 0.1760990 0.2907525 0.6976148
#>           Rhat n.eff
#>  [1,] 1.032145    79
#>  [2,] 1.006705   470
#>  [3,] 1.015162   550
#>  [4,] 1.019060   460
#>  [5,] 1.006489   310
#>  [6,] 1.080927    49
#>  [7,] 1.035933   110
#>  [8,] 1.005077   310
#>  [9,] 1.011107   170
#> [10,] 1.028212   110
#> [11,] 1.014608   130
#> [12,] 1.012965   300
#> [13,] 1.047520    64
#> [14,] 1.045464    66
#> [15,] 1.040942    53
#> [16,] 1.049024    66
#> [17,] 1.061708   120
#> [18,] 1.022200   170
#> [19,] 1.071374    85
#> [20,] 1.025511   180
#> [21,] 1.019926   200
#> [22,] 1.001664   600
#> [23,] 1.004365   600
#> [24,] 1.029764   220
#> [25,] 1.002217   540
#> 
#> $measure
#> [1] "SMD"
#> 
#> $model
#> [1] "RE"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] -2.9900000  0.2143347  4.0000000
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 5
#> 
#> attr(,"class")
#> [1] "run_sensitivity"
# }