Perform sensitivity analysis for missing outcome participant data

Performs a sensitivity analysis by applying pairwise meta-analysis (PMA) or network meta-analysis (NMA) for a series of different scenarios about the informative missingness parameter,

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length of 5 or larger. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default arguments are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects PMA:

EM	The estimated summary effect measure (according to the argument measure defined in run_model).

tau	The between-trial standard deviation. This element does not appear in the case of a fixed-effect PMA.

In a random-effects NMA, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore, tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the vector specified in mean_scenarios (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package function to be processed further and provide an end-user-ready output.

In the case of PMA, EM and tau are estimated as many times as the number of scenarios considered. In the case of NMA, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console. Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to Spineli (2019), Mavridis et al. (2015), Turner et al. (2015), and White et al. (2008) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, heter_prior, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model). Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci. 1992;7:457--472.

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--741. doi: 10.1002/sim.6365

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--490. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--2080. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--727. doi: 10.1002/sim.3008

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (default arguments)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The following vector of scenarios was considered by default: c(-2, -1, 0, 1, 2)
#> 
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 624
#> 
#> Initializing model
#> 
#> $EM
#>              mean        sd       2.5%     97.5%     Rhat n.eff
#>  [1,] -0.10876559 0.3062351 -0.7346564 0.5265658 1.019081   270
#>  [2,] -0.12122650 0.3334839 -0.7412519 0.5333821 1.003062   440
#>  [3,] -0.15734575 0.3227913 -0.7976501 0.5767675 1.009190   540
#>  [4,] -0.16614366 0.3511493 -0.8606588 0.4493203 1.006820   540
#>  [5,] -0.20277057 0.3343087 -0.8431813 0.4733912 1.009363   280
#>  [6,] -0.07815457 0.3537199 -0.8293072 0.6137303 1.006320   310
#>  [7,] -0.09487643 0.3110728 -0.6665889 0.4974820 1.017370   200
#>  [8,] -0.10365558 0.3461746 -0.8021838 0.6609228 1.015356   150
#>  [9,] -0.13407103 0.3642485 -0.7601491 0.5704566 1.019130   170
#> [10,] -0.16751567 0.3617316 -0.8603582 0.5772357 1.006346   540
#> [11,] -0.04248056 0.3349004 -0.7082829 0.6511442 1.005579   540
#> [12,] -0.05611864 0.3585943 -0.7060783 0.6692161 1.004493   320
#> [13,] -0.09930374 0.3761554 -0.9039624 0.6275074 1.004288   450
#> [14,] -0.12291298 0.3470416 -0.7737745 0.5913097 1.011772   330
#> [15,] -0.14715716 0.3434453 -0.8694204 0.5039177 1.001514   540
#> [16,] -0.07081587 0.3287605 -0.7488994 0.6286831 1.007692   210
#> [17,] -0.03892547 0.3626280 -0.6806437 0.6871652 1.045611   500
#> [18,] -0.08629955 0.3101407 -0.6931845 0.5240998 1.001875   540
#> [19,] -0.10995540 0.3119306 -0.6991762 0.5276430 1.016613   110
#> [20,] -0.08650902 0.3446543 -0.6972476 0.5053386 1.064323   120
#> [21,] -0.01363886 0.3413120 -0.7345508 0.5960454 1.005539   270
#> [22,] -0.02871577 0.3793350 -0.7753366 0.7059076 1.043535   210
#> [23,] -0.06709538 0.2998012 -0.7401313 0.4207540 1.049983   190
#> [24,] -0.07525966 0.3379555 -0.7687588 0.5420502 1.030339   500
#> [25,] -0.10339980 0.3092413 -0.7024801 0.5340110 1.001663   540
#> 
#> $tau
#>             50%        sd        2.5%    97.5%     Rhat n.eff
#>  [1,] 0.2920260 0.3191577 0.012185783 1.234003 1.095561    34
#>  [2,] 0.3402211 0.3217774 0.021879637 1.120819 1.072650    51
#>  [3,] 0.3525900 0.3624417 0.007619496 1.360385 1.127279    37
#>  [4,] 0.4112586 0.3430194 0.076225055 1.389735 1.017568   130
#>  [5,] 0.3703663 0.3360588 0.014504738 1.216818 1.038648   170
#>  [6,] 0.3490794 0.3756886 0.031093984 1.431083 1.026293   120
#>  [7,] 0.3183439 0.3098748 0.014990586 1.118548 1.047194    51
#>  [8,] 0.3867829 0.3362627 0.022810286 1.268442 1.004480   400
#>  [9,] 0.4100766 0.3365592 0.033275692 1.347689 1.059104    80
#> [10,] 0.3951739 0.3178049 0.015699415 1.216536 1.001791   540
#> [11,] 0.3159502 0.3394392 0.016353276 1.306739 1.006790   240
#> [12,] 0.3826033 0.3717465 0.017206442 1.379699 1.038026    55
#> [13,] 0.3918129 0.3797383 0.021273998 1.450184 1.032484   130
#> [14,] 0.3727567 0.3912239 0.029997349 1.432047 1.058440    48
#> [15,] 0.3788205 0.3497103 0.031269266 1.360269 1.018888   540
#> [16,] 0.3594269 0.3783768 0.011561400 1.368894 1.017192   170
#> [17,] 0.2990537 0.3301637 0.024690605 1.314717 1.075949    42
#> [18,] 0.3347645 0.3254784 0.038145186 1.221967 1.025548   160
#> [19,] 0.3003927 0.3291671 0.010689149 1.241862 1.025592   150
#> [20,] 0.3321475 0.4137000 0.040241516 1.309746 1.010720   240
#> [21,] 0.3634524 0.3613881 0.020298995 1.351291 1.032853    89
#> [22,] 0.3759756 0.3547902 0.003858408 1.304093 1.139216    27
#> [23,] 0.1989571 0.2910837 0.007994104 1.075442 1.573563     7
#> [24,] 0.3681049 0.3341118 0.027681658 1.298603 1.002662   450
#> [25,] 0.3321445 0.3237661 0.028431360 1.261068 1.014425   540
#> 
#> $measure
#> [1] "SMD"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] 0 1 1
#> 
# }