Perform sensitivity analysis for aggregate missing outcome participant data

Performs a sensitivity analysis by applying pairwise meta-analysis (PMA) or network meta-analysis (NMA) for a series of different scenarios about the informative missingness parameter,

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin
)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

assumption
Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios
A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length of 5 or larger positive odd integer. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default arguments are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar
A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04

n_chains
Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter
Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin
Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin
Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects PMA:

EM	The estimated summary effect measure (according to the argument measure defined in run_model).

tau	The between-trial standard deviation. This element does not appear in the case of a fixed-effect PMA.

In a random-effects NMA, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore, tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the mean_scenarios vector (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package function to be processed further and provide an end-user-ready output.

In the case of PMA, EM and tau have as many rows as the number of scenarios considered. In the case of NMA, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console. Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to White et al., (2008), Mavridis et al., (2015), Turner et al., (2015) and Spineli (2019) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, and heter_prior from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model). Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475--490. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721--741. doi: 10.1002/sim.6365

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062--2080. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis--part 1: two-stage methods. Stat Med 2008;27(5):711--727. doi: 10.1002/sim.3008

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci. 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (missing-at-random assumption)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 1)
#> The following vector of scenarios was considered by default: c(-2, -1, 0, 1, 2)
#> 
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 578
#> 
#> Initializing model
#> 
#> $EM
#>              mean        sd       2.5%     97.5%     Rhat n.eff
#>  [1,] -0.11759386 0.3697718 -0.8950278 0.6139633 1.003925  2500
#>  [2,] -0.13719703 0.3375567 -0.8202626 0.5047184 1.004349  2700
#>  [3,] -0.13927391 0.3265689 -0.8091553 0.4996222 1.017941   200
#>  [4,] -0.16932442 0.3297035 -0.8415465 0.5257458 1.004960   530
#>  [5,] -0.21270558 0.3697954 -0.9551212 0.5351479 1.005677  2700
#>  [6,] -0.07711719 0.3319128 -0.7395582 0.5878225 1.005578   550
#>  [7,] -0.11653134 0.3668490 -0.8327416 0.6164954 1.012796  2700
#>  [8,] -0.10777515 0.3375969 -0.7678605 0.5964872 1.002931  2700
#>  [9,] -0.15038006 0.3325501 -0.8004208 0.4985778 1.002894   820
#> [10,] -0.16945511 0.3538123 -0.8553994 0.5698165 1.002129  1200
#> [11,] -0.04019550 0.3602070 -0.6920818 0.6724462 1.020785   480
#> [12,] -0.08187754 0.3215122 -0.7242252 0.5672273 1.007351   330
#> [13,] -0.11807809 0.3404251 -0.8077181 0.5474760 1.024565  2700
#> [14,] -0.11989522 0.3887474 -0.9116006 0.6579100 1.006670   410
#> [15,] -0.14201650 0.3666829 -0.9356938 0.5667780 1.026561   630
#> [16,] -0.03555762 0.3459185 -0.7557486 0.6222642 1.005647   880
#> [17,] -0.06550546 0.3394585 -0.7270527 0.5870290 1.022522  1100
#> [18,] -0.08018253 0.3248649 -0.7236674 0.5747286 1.014237   340
#> [19,] -0.08650101 0.3485909 -0.8164155 0.6085059 1.012133  1700
#> [20,] -0.12751436 0.3487594 -0.8174312 0.5305965 1.007665   400
#> [21,]  0.00565248 0.3395068 -0.6931597 0.6985348 1.005223   430
#> [22,] -0.04099815 0.3331173 -0.7348107 0.6421403 1.014076   170
#> [23,] -0.03811316 0.3588734 -0.7294678 0.6652489 1.003509   650
#> [24,] -0.06229953 0.3390764 -0.7356067 0.6098167 1.005289   410
#> [25,] -0.07485776 0.3644357 -0.8390408 0.6579547 1.006136  1000
#> 
#> $tau
#>             50%        sd        2.5%    97.5%     Rhat n.eff
#>  [1,] 0.3660048 0.3628889 0.036240808 1.324672 1.020838   160
#>  [2,] 0.3551076 0.3522604 0.010606515 1.289497 1.151288    31
#>  [3,] 0.3343252 0.3480252 0.018009849 1.381885 1.132615    31
#>  [4,] 0.3559554 0.3230991 0.023681372 1.223435 1.011084   250
#>  [5,] 0.3632938 0.3886656 0.031533746 1.475990 1.011906   440
#>  [6,] 0.3784669 0.3034737 0.050671869 1.178809 1.029606   120
#>  [7,] 0.3607943 0.3766717 0.018828883 1.418781 1.009202   230
#>  [8,] 0.3350266 0.3540792 0.038519800 1.355015 1.019271  2000
#>  [9,] 0.3628724 0.3210903 0.040678133 1.208011 1.003266  2700
#> [10,] 0.3756668 0.3385848 0.017368018 1.261553 1.071890    89
#> [11,] 0.3584105 0.3805340 0.027348404 1.426539 1.029613   150
#> [12,] 0.3398024 0.3458957 0.026869889 1.327572 1.003071   760
#> [13,] 0.3533153 0.3553721 0.031500985 1.316314 1.009233   240
#> [14,] 0.3706416 0.4011895 0.024349334 1.675085 1.062780    64
#> [15,] 0.3776032 0.4005550 0.016334291 1.503765 1.037369   120
#> [16,] 0.3938207 0.3571958 0.033997346 1.381857 1.014729   280
#> [17,] 0.3613297 0.3603264 0.045574951 1.397299 1.055256    41
#> [18,] 0.3407279 0.3519084 0.013329269 1.355399 1.061433    55
#> [19,] 0.3585355 0.3480241 0.021631751 1.328985 1.109746    29
#> [20,] 0.3837548 0.3416869 0.034918927 1.337881 1.020461   280
#> [21,] 0.3479114 0.3542179 0.010966537 1.327438 1.025778   190
#> [22,] 0.2769688 0.3490030 0.003714222 1.303647 1.084334    44
#> [23,] 0.3588856 0.3931033 0.000705896 1.443780 1.204235    35
#> [24,] 0.3678687 0.3109940 0.035025769 1.207455 1.029838   110
#> [25,] 0.3747101 0.3626927 0.044172619 1.434847 1.030620    75
#> 
#> $measure
#> [1] "SMD"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 1
#> 
#> $heter
#> [1] 0 1 1
#> 
# }