Perform sensitivity analysis for missing participant outcome data

Performs a sensitivity analysis by applying pairwise meta-analysis or network meta-analysis for a series of different scenarios about the informative missingness parameter.

Usage

run_sensitivity(
  full,
  assumption,
  mean_scenarios,
  var_misspar,
  n_chains,
  n_iter,
  n_burnin,
  n_thin,
  inits = NULL
)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

assumption

Character string indicating the structure of the informative missingness parameter. Set assumption equal to one of the following two: "HIE-ARM", or "IDE-ARM" (see 'Details'). The default argument is "IDE-ARM". The abbreviations "IDE", and "HIE" stand for identical, and hierarchical, respectively.

mean_scenarios

A vector with numeric values for the mean of the normal distribution of the informative missingness parameter (see 'Details'). The vector should have a length equal to 5 or larger. The missing-at-random (MAR) assumption should be the median of the vector, so that the same number of informative scenarios appear before and after the MAR. The default scenarios are c(-log(3), -log(2), log(0.9999), log(2), log(3)) and c(-2, -1, 0, 1, 2) for binary and continuous outcome data, respectively.

var_misspar

A positive non-zero number for the variance of the normal distribution of the informative missingness parameter. When the measure (defined in run_model) is "OR", "MD", or "SMD" the default argument is 1. When the measure is "ROM", the default argument is 0.04.

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a random-effects pairwise meta-analysis:

EM

The estimated summary effect measure (according to the argument measure defined in run_model).

EM_pred

The predicted summary effect measure (according to the argument measure defined in run_model). This element does not appear in the case of a fixed-effect meta-analysis.

EM_LOR

The estimated summary odd ratio in the logarithmic scale when measure = "RR" or measure = "RD".

EM_LOR_pred

The predicted summary odd ratio in the logarithmic scale when measure = "RR" or measure = "RD". This element does not appear in the case of a fixed-effect meta-analysis.

tau

The between-trial standard deviation. This element does not appear in the case of a fixed-effect meta-analysis.

In a random-effects network meta-analysis, EM refer to all possible pairwise comparisons of interventions in the network. Furthermore, tau is typically assumed to be common for all observed comparisons in the network.

Details

The model runs in JAGS and the progress of the simulation appears on the R console. The number of times run_sensitivity is used appears on the R console as a text in red and it equals the number of scenarios defined as the square of the length of the vector specified in mean_scenarios (see 'Examples'). The output of run_sensitivity is used as an S3 object by other functions of the package to be processed further and provide an end-user-ready output.

In the case of pairwise meta-analysis, EM and tau are estimated as many times as the number of scenarios considered. In the case of network meta-analysis, each possible pairwise comparison is estimated as many times as the number of scenarios considered.

The informative missingness parameter is assumed to differ only across the interventions of the dataset. Therefore, the user can specify the informative missingness parameter to be arm-specific and identical (assumption = "IDE-ARM"), or arm-specific and hierarchical (assumption = "HIE-ARM") (Spineli et al., 2021).

The length of the vector specified in argument mean_scenarios should be equal to or more than 5 (a positive odd integer) to allow for an adequate number of scenarios. It is important that the number corresponding to the MAR assumption is the middle of the numbers in the vector specified in argument mean_scenarios. The MAR assumption constitutes the primary analysis. Under the informative missingness difference of means parameter (relevant for the raw and standardised mean diffenre), the MAR assumption equals 0. Under the informative missingness odds ratio parameter (IMOR; relevant for the odds ratio) and the informative missingness ratio of means (IMRoM; relevant for the ratio of means) parameter, the MAR assumption equals 1; however, both parameters are analysed in the logarithmic scale. We advise using the value 0.999 rather than 1 in mean_scenarios for the IMOR and IMRoM parameters; otherwise, the execution of the function will be stopped and the error 'Invalid parent values' will be printed on the R console.

Currently, there are no empirically-based prior distributions for the informative missingness parameters. The users may refer to Spineli (2019), Mavridis et al. (2015), Turner et al. (2015), and White et al. (2008) to determine mean_scenarios for an informative missingness mechanism and select a proper value for var_misspar.

run_sensitivity inherits the arguments data, measure, model, heter_prior, D, indic, base_risk, and ref from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in the primary analysis (via run_model)–an exception in the assumption argument as it is restricted to only two character strings. Therefore, the user needs first to apply run_model, and then use run_sensitivity (see 'Examples').

The run_sensitivity function also returns the arguments measure, scenarios, D, heter, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

run_sensitivity can be used only when missing participant outcome data have been extracted for at least one trial. Otherwise, the execution of the function will be stopped and an error message will be printed on the R console.

References

Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457–72. doi: 10.1214/ss/1177011136

Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for uncertainty due to missing continuous outcome data in pairwise and network meta-analysis. Stat Med 2015;34(5):721–41. doi: 10.1002/sim.6365

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. Res Synth Methods 2021;12(4):475–90. doi: 10.1002/jrsm.1478

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account for uncertainty due to missing binary outcome data in pairwise meta-analysis. Stat Med 2015;34(12):2062–80. doi: 10.1002/sim.6475

White IR, Higgins JP, Wood AM. Allowing for uncertainty due to missing data in meta-analysis–part 1: two-stage methods. Stat Med 2008;27(5):711–27. doi: 10.1002/sim.3008

See also

Author

Loukia M. Spineli

Examples

data("pma.taylor2004")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_taylor.rds', package = 'rnmamod'))

# \donttest{
# Perform the sensitivity analysis (default arguments)
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_sensitivity(full = res,
                assumption = "IDE-ARM",
                var_misspar = 1,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 5)
#> The default scenarios were considered: c(-2, -1, 0, 1, 2).
#> JAGS generates initial values for the parameters.
#> 1 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 1 until convergence
#> 2 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 2 until convergence
#> 3 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 3 until convergence
#> 4 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 4 until convergence
#> 5 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 5 until convergence
#> 6 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 6 until convergence
#> 7 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 7 until convergence
#> 8 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 8 until convergence
#> 9 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 9 until convergence
#> 10 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 10 until convergence
#> 11 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 11 until convergence
#> 12 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 12 until convergence
#> 13 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 13 until convergence
#> 14 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 14 until convergence
#> 15 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 15 until convergence
#> 16 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 16 until convergence
#> 17 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 17 until convergence
#> 18 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 18 until convergence
#> 19 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 19 until convergence
#> 20 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 20 until convergence
#> 21 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 21 until convergence
#> 22 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 22 until convergence
#> 23 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 23 until convergence
#> 24 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 24 until convergence
#> 25 out of 25 total scenarios
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 28
#>    Total graph size: 348
#> 
#> Initializing model
#> 
#> Updating model for scenario 25 until convergence
#> $EM
#>              mean        sd       2.5%        25%          50%          75%
#>  [1,] -0.12040645 0.2403981 -0.5603110 -0.2736889 -0.125619327  0.035140669
#>  [2,] -0.13084368 0.2512539 -0.6522067 -0.3075912 -0.119149493  0.038361789
#>  [3,] -0.15660748 0.2316688 -0.5865530 -0.3013978 -0.163782289 -0.022066681
#>  [4,] -0.13929102 0.2606617 -0.6231867 -0.2978628 -0.144203871 -0.007715589
#>  [5,] -0.18006862 0.2435978 -0.6487045 -0.3209558 -0.189366583 -0.039786676
#>  [6,] -0.09464421 0.2683596 -0.6231120 -0.2754080 -0.097450927  0.072462461
#>  [7,] -0.13524438 0.2481582 -0.5647975 -0.2836989 -0.153872876  0.009719066
#>  [8,] -0.10422393 0.2484011 -0.5710165 -0.2645562 -0.101463668  0.047295166
#>  [9,] -0.13756518 0.2554983 -0.6225391 -0.2775309 -0.135991914  0.016016727
#> [10,] -0.17648450 0.2726005 -0.7049290 -0.3347477 -0.173958953 -0.031263630
#> [11,] -0.03247413 0.2578893 -0.5243098 -0.2114717 -0.031191679  0.142980939
#> [12,] -0.08051539 0.2485333 -0.5328314 -0.2206604 -0.074418383  0.067366913
#> [13,] -0.11073689 0.2709233 -0.6376907 -0.2848940 -0.109590396  0.048330081
#> [14,] -0.10766200 0.2765759 -0.6267854 -0.2822833 -0.119485738  0.055137036
#> [15,] -0.12335894 0.2502602 -0.6212855 -0.2850822 -0.128734521  0.034149035
#> [16,] -0.02190785 0.2481770 -0.4902504 -0.1821230 -0.018011785  0.130483188
#> [17,] -0.02956139 0.2440552 -0.4957430 -0.1831173 -0.038570704  0.132711895
#> [18,] -0.07478580 0.2506656 -0.5354603 -0.2292411 -0.070747671  0.067871427
#> [19,] -0.07066852 0.2501260 -0.6364362 -0.2150756 -0.064337685  0.080941962
#> [20,] -0.09701460 0.2604603 -0.6643136 -0.2538076 -0.103303453  0.085007411
#> [21,] -0.04021226 0.2384909 -0.5555641 -0.1859412 -0.034750999  0.122097460
#> [22,] -0.01053793 0.2621640 -0.5265575 -0.1764878  0.003314019  0.172635906
#> [23,] -0.04329979 0.2383154 -0.4824517 -0.1864273 -0.029003336  0.089687919
#> [24,] -0.08215108 0.2527328 -0.5638187 -0.2458218 -0.069381330  0.065280597
#> [25,] -0.07341287 0.2276026 -0.5102156 -0.2097185 -0.079644615  0.075540882
#>           97.5%     Rhat n.eff
#>  [1,] 0.3561421 1.001421   600
#>  [2,] 0.3487690 1.003194   600
#>  [3,] 0.3131914 1.025456   600
#>  [4,] 0.3783509 1.002055   600
#>  [5,] 0.2948726 1.009653   230
#>  [6,] 0.4422412 1.004155   350
#>  [7,] 0.3599786 1.007737   460
#>  [8,] 0.3892440 1.014559   130
#>  [9,] 0.3159456 1.008105   350
#> [10,] 0.3019319 1.040807   370
#> [11,] 0.4517071 1.006281   260
#> [12,] 0.3779602 1.001386   600
#> [13,] 0.4183534 1.031986    74
#> [14,] 0.4466652 1.014147   480
#> [15,] 0.3796730 1.015584   130
#> [16,] 0.4987282 1.005031   350
#> [17,] 0.4102389 1.009684   270
#> [18,] 0.4284144 1.017175   170
#> [19,] 0.4272525 1.007757   330
#> [20,] 0.3430726 1.005484   320
#> [21,] 0.3844946 1.013985   240
#> [22,] 0.4653104 1.001246   600
#> [23,] 0.4387606 1.009781   290
#> [24,] 0.3735669 1.004869   380
#> [25,] 0.3462618 1.015675   600
#> 
#> $EM_pred
#>              mean        sd       2.5%        25%         50%          75%
#>  [1,] -0.10354922 0.3774576 -0.7971567 -0.3087643 -0.12528846  0.087505778
#>  [2,] -0.12589837 0.4271295 -0.9071846 -0.3454128 -0.10651267  0.101069941
#>  [3,] -0.16983241 0.3830484 -0.9878552 -0.3640014 -0.17624293  0.021945985
#>  [4,] -0.14662909 0.3817039 -1.0120411 -0.3485534 -0.14031444  0.043464230
#>  [5,] -0.18382731 0.4142448 -1.0003383 -0.3752480 -0.18976221 -0.003845255
#>  [6,] -0.08401466 0.4078439 -0.8242304 -0.2979500 -0.07779700  0.115312312
#>  [7,] -0.14513581 0.4074704 -0.9353675 -0.3339997 -0.16668161  0.057867326
#>  [8,] -0.10454907 0.4050648 -0.9042509 -0.3152519 -0.10752353  0.102622312
#>  [9,] -0.11918137 0.4928485 -0.9647207 -0.3202171 -0.12899750  0.086141638
#> [10,] -0.14837842 0.4212545 -0.8529872 -0.3312908 -0.15876390  0.020319972
#> [11,] -0.03104258 0.4081993 -0.8101354 -0.2488972 -0.02740129  0.181102182
#> [12,] -0.09089241 0.4136173 -0.9195456 -0.2808608 -0.07874226  0.122809461
#> [13,] -0.09837526 0.4316928 -1.0479417 -0.3117039 -0.11105281  0.106880077
#> [14,] -0.10470809 0.4976851 -1.0587363 -0.3239231 -0.12237191  0.074870457
#> [15,] -0.12886403 0.3880417 -0.9100398 -0.3321691 -0.13969328  0.092397003
#> [16,] -0.03459608 0.3728356 -0.7480939 -0.2232545 -0.01734606  0.156875526
#> [17,] -0.04155849 0.3536278 -0.7009118 -0.2248470 -0.05546145  0.141969096
#> [18,] -0.07232762 0.4153530 -0.8342166 -0.2821963 -0.07438316  0.114127150
#> [19,] -0.05632318 0.4279572 -0.8932282 -0.2354637 -0.06334466  0.125087857
#> [20,] -0.11654440 0.4592772 -0.8296064 -0.3036811 -0.11470240  0.104689357
#> [21,] -0.04560525 0.3646915 -0.8883456 -0.2088542 -0.03628311  0.166600091
#> [22,] -0.03559465 0.4138838 -0.8020278 -0.2732689 -0.02200551  0.200065248
#> [23,] -0.04158297 0.3954960 -0.8399953 -0.2253011 -0.03022334  0.126038500
#> [24,] -0.09089937 0.4088232 -0.8914325 -0.2829819 -0.07108194  0.103452522
#> [25,] -0.07316051 0.3689134 -0.8116092 -0.2626117 -0.08114799  0.111154749
#>           97.5%      Rhat n.eff
#>  [1,] 0.7019990 1.0065310   550
#>  [2,] 0.5993380 1.0366865   390
#>  [3,] 0.5754550 1.0306459   600
#>  [4,] 0.7167538 1.0026233   600
#>  [5,] 0.5973863 1.0506766   190
#>  [6,] 0.6166231 1.0046924   600
#>  [7,] 0.6384284 1.0186870   270
#>  [8,] 0.7262983 1.0229040   600
#>  [9,] 0.7678454 1.0458188   200
#> [10,] 0.6188396 1.0311934   270
#> [11,] 0.7459397 0.9985001   600
#> [12,] 0.7676097 1.0175667   600
#> [13,] 0.7657717 1.0156739   140
#> [14,] 0.9987232 1.0367188   550
#> [15,] 0.6125424 1.0072632   230
#> [16,] 0.6486718 1.0106177   600
#> [17,] 0.7221558 1.0011779   600
#> [18,] 0.7901762 1.0096338   240
#> [19,] 0.9153517 1.0003026   600
#> [20,] 0.6221738 1.0773560   600
#> [21,] 0.6311275 1.0025739   600
#> [22,] 0.8847651 1.0108795   600
#> [23,] 0.8145854 1.0172904   390
#> [24,] 0.6424024 1.0152004   600
#> [25,] 0.6145600 1.0112655   600
#> 
#> $tau
#>            mean        sd        2.5%        25%       50%       75%     97.5%
#>  [1,] 0.2235538 0.1954030 0.014813446 0.07612999 0.1705486 0.3028789 0.7781401
#>  [2,] 0.2386254 0.2182597 0.018059647 0.09648567 0.1825280 0.3274375 0.8153199
#>  [3,] 0.2422930 0.2332232 0.020492020 0.09405960 0.1707308 0.3108379 0.8394793
#>  [4,] 0.2521116 0.2221726 0.018992185 0.09562273 0.1890884 0.3398338 0.8329383
#>  [5,] 0.2513868 0.2303881 0.010922882 0.09126404 0.1872720 0.3441144 0.9195018
#>  [6,] 0.2426129 0.2093280 0.017266163 0.09779894 0.1825232 0.3188170 0.7740843
#>  [7,] 0.2299749 0.2186952 0.008123107 0.07691237 0.1668884 0.3184913 0.7847678
#>  [8,] 0.2436524 0.2241645 0.008041883 0.08891906 0.1784471 0.3340168 0.8405054
#>  [9,] 0.2494633 0.2895971 0.011152131 0.09009337 0.1751774 0.3246130 0.8605911
#> [10,] 0.2437270 0.2506043 0.008566591 0.07932822 0.1737269 0.3169708 0.9883990
#> [11,] 0.2490689 0.2064086 0.020567348 0.10698147 0.1986181 0.3332942 0.7868344
#> [12,] 0.2396240 0.2089567 0.018654284 0.08810682 0.1777188 0.3239445 0.7633478
#> [13,] 0.2520577 0.2267029 0.013587222 0.09995854 0.1877756 0.3323845 0.8248848
#> [14,] 0.2623322 0.2654553 0.009563966 0.07892236 0.1856899 0.3606922 0.9782393
#> [15,] 0.2221235 0.2141602 0.011947002 0.07755198 0.1644605 0.3007578 0.7032863
#> [16,] 0.2212497 0.2115768 0.008236245 0.08244452 0.1654097 0.2844905 0.8063939
#> [17,] 0.2011351 0.1974976 0.005885839 0.06820911 0.1491864 0.2604416 0.7187559
#> [18,] 0.2438647 0.2096144 0.018576296 0.09546002 0.1854758 0.3269252 0.7658942
#> [19,] 0.2373383 0.2151952 0.011484691 0.08774793 0.1777226 0.3212383 0.8492440
#> [20,] 0.2210260 0.2802905 0.004973462 0.06330687 0.1584258 0.2935574 0.8317600
#> [21,] 0.2112277 0.1860270 0.011636251 0.08002243 0.1642851 0.2853711 0.6717093
#> [22,] 0.2186791 0.1992936 0.012652059 0.08005889 0.1595971 0.2977125 0.7622800
#> [23,] 0.2203792 0.2047610 0.009473488 0.07768504 0.1631477 0.3035083 0.7515979
#> [24,] 0.2228543 0.2288671 0.008309293 0.07183254 0.1669657 0.2951304 0.8255741
#> [25,] 0.2248829 0.2038103 0.014680379 0.08409449 0.1691970 0.3071836 0.7442226
#>           Rhat n.eff
#>  [1,] 1.043025   380
#>  [2,] 1.007684   600
#>  [3,] 1.039345    56
#>  [4,] 1.002838   600
#>  [5,] 1.041403    57
#>  [6,] 1.019315   120
#>  [7,] 1.043986   100
#>  [8,] 1.095566    38
#>  [9,] 1.013622   210
#> [10,] 1.053754    47
#> [11,] 1.009314   600
#> [12,] 1.027211    86
#> [13,] 1.039704    72
#> [14,] 1.013390   140
#> [15,] 1.017293   600
#> [16,] 1.019192   260
#> [17,] 1.056327    57
#> [18,] 1.000552   600
#> [19,] 1.011908   310
#> [20,] 1.009879   370
#> [21,] 1.018745   320
#> [22,] 1.029174   110
#> [23,] 1.051287    48
#> [24,] 1.025323    84
#> [25,] 1.024655   190
#> 
#> $measure
#> [1] "SMD"
#> 
#> $model
#> [1] "RE"
#> 
#> $scenarios
#> [1] -2 -1  0  1  2
#> 
#> $D
#> [1] 0
#> 
#> $heter
#> [1] -2.9900000  0.2143347  4.0000000
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 5
#> 
#> attr(,"class")
#> [1] "run_sensitivity"
# }