Perform a series of Bayesian pairwise meta-analyses

Performs a Bayesian pairwise meta-analysis for each pairwise comparison with at least two trials in the network.

Usage

run_series_meta(full, n_chains, n_iter, n_burnin, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_chains

Integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM

The summary effect estimate (according to the argument measure defined in run_model) of each observed pairwise comparison with at least two trials in the network.

tau

The between-trial standard deviation for pairwise comparisons with at least two trials, when the random-effects model has been specified.

single

A binary vector that indicates the comparisons in EM with one trial.

Details

run_series_meta inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var_misspar from run_model (now contained in the argument full). This prevents specifying a different Bayesian model from that considered in run_model. Therefore, the user needs first to apply run_model, and then use run_series_meta (see 'Examples').

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_series_meta currently performs a series of pairwise meta-analysis using the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

run_series_meta runs a series of Bayesian pairwise meta-analyses in JAGS. The progress of the simulation appears on the R console. The number of times the function is used is also printed on the console (in red) and is equal to the number of observed pairwise comparisons in the network (see 'Examples'). The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

The output of run_series_meta is not end-user-ready. The series_meta_plot function inherits the output of run_series_meta as an S3 object and processes it further to provide an end-user-ready output.

run_series_meta can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457–72. doi: 10.1214/ss/1177011136

See also

Author

Loukia M. Spineli

Examples

data("nma.dogliotti2014")

# Show the first six trials of the dataset (one-trial-per-row format)
head(nma.dogliotti2014)
#>              study t1 t2 t3   r1   r2 r3  m1  m2 m3   n1   n2 n3
#> 1     BAATAF, 1990  1  7 NA  195  188 NA   0  21 NA  208  212 NA
#> 2     SPINAF, 1992  1  7 NA  186  172 NA  56  81 NA  265  260 NA
#> 3    SPAF-II, 1994  2  7 NA  480  478 NA  23  38 NA  545  555 NA
#> 4      PATAF, 1999  2  7 NA  243   86 NA  54  42 NA  319  131 NA
#> 5 ACTIVE (W), 2006  3  7 NA 2825 3089 NA 410 223 NA 3335 3371 NA
#> 6       JAST, 2006  1  2 NA  338  313 NA  89  96 NA  445  426 NA

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_dogliotti.rds', package = 'rnmamod'))

# \donttest{
# Run separate random-effects pairwise meta-analyses
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_series_meta(full = res,
                n_chains = 3,
                n_iter = 1000,
                n_burnin = 100,
                n_thin = 1)
#> JAGS generates initial values for the parameters.
#> 1 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 20
#>    Unobserved stochastic nodes: 29
#>    Total graph size: 532
#> 
#> Initializing model
#> 
#> 2 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 28
#>    Unobserved stochastic nodes: 37
#>    Total graph size: 699
#> 
#> Initializing model
#> 
#> 3 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> 4 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 16
#>    Unobserved stochastic nodes: 25
#>    Total graph size: 447
#> 
#> Initializing model
#> 
#> 5 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> 6 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> 7 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> 8 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> 9 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> 10 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> 11 out of 11 observed comparisons (and model updating until convergence)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 4
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 195
#> 
#> Initializing model
#> 
#> $EM
#>    t1 t2        mean        sd       2.5%         25%         50%       75%
#> 1   1  7  1.06431223 0.4043830  0.2514131  0.80440128  1.04771176 1.3239966
#> 2   2  7  0.46090893 0.3611715 -0.2395673  0.23454637  0.44448059 0.6925197
#> 3   3  7  0.62666761 1.0466197 -1.7522349  0.19061694  0.63793123 1.0656601
#> 4   1  2  0.36264864 0.3075178 -0.2276485  0.16642980  0.35955055 0.5454429
#> 5   2  3  0.32997991 0.9914709 -1.8505084 -0.09988260  0.33353010 0.7327676
#> 6   6  7 -0.34208005 1.0413139 -2.5237065 -0.83989018 -0.34062911 0.1276220
#> 7   7  8  0.23420289 0.9234513 -1.8044193 -0.12654272  0.24360265 0.5959585
#> 8   2  8  0.80921391 1.0177775 -1.3927139  0.36438043  0.78284264 1.2588806
#> 9   4  5  0.38350801 0.9658629 -1.8305732 -0.04127839  0.37487247 0.8032338
#> 10  4  7 -0.06077283 0.9701563 -2.2506316 -0.46420587 -0.05115666 0.3759651
#> 11  5  7 -0.35368290 1.1234564 -2.7134715 -0.80851645 -0.36618891 0.1335502
#>       97.5%     Rhat n.eff
#> 1  1.864378 1.003657   920
#> 2  1.206581 1.017912   150
#> 3  2.860329 1.004865  1100
#> 4  1.001055 1.017207   160
#> 5  2.389561 1.007871  1200
#> 6  1.959531 1.010995   220
#> 7  2.308140 1.006008  1300
#> 8  3.022745 1.005430   520
#> 9  2.466355 1.002427  1200
#> 10 2.008062 1.003687   630
#> 11 2.132904 1.008430   450
#> 
#> $tau
#>   t1 t2      mean        sd        2.5%        25%       50%       75%    97.5%
#> 1  1  7 0.4817719 0.3259085 0.016650798 0.24284622 0.4288940 0.6662050 1.253482
#> 2  2  7 0.5254615 0.3103862 0.039283722 0.32149698 0.4768967 0.6691229 1.229028
#> 4  1  2 0.2946878 0.2723626 0.007666396 0.09537791 0.2183858 0.4112259 1.005454
#>       Rhat n.eff
#> 1 1.035248   770
#> 2 1.087508    55
#> 4 1.171242    25
#> 
#> $measure
#> [1] "OR"
#> 
#> $model
#> [1] "RE"
#> 
#> $single
#>  [1] 0 0 1 0 1 1 1 1 1 1 1
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> attr(,"class")
#> [1] "run_series_meta"
# }