Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al. (2013) that has been refined (Spineli, 2021a) and extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al. 2021b; Spineli, 2019). This model offers a global evaluation of the plausibility of the consistency assumption in the network.

run_ume(full, n_iter, n_burnin, n_chains, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM	The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o	The deviance contribution of each trial-arm based on the observed outcome.

hat_par	The fitted outcome at each trial-arm.

tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o	The leverage for the observed outcome at each trial-arm.

sign_dev_o	The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment	A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit	An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then use run_ume (see 'Examples').

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm trial of the network (Spineli, 2021a). The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has been also refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the basic parameters of these trials, that is, effect parameters between non-baseline arms and the baseline arm (Spineli, 2021a).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013;33(5):641--56. doi: 10.1177/0272989X12455847

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

Spineli LM. A Revised Framework to Evaluate the Consistency Assumption Globally in a Network of Interventions. Med Decis Making 2021a. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021b;30(4):958--975. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 3045
#> 
#> Initializing model
#> 
#> $EM
#>               mean        sd       2.5%         25%        50%        75%
#> EM[2,1]  0.5960686 0.8347675 -1.0539409  0.08599281  0.5784068  1.0623436
#> EM[3,1]  0.1980291 0.7012713 -1.1679487 -0.24311777  0.1801532  0.6328200
#> EM[4,1]  0.8707993 0.5511713 -0.3294761  0.54562331  0.9023518  1.2410880
#> EM[5,1]  1.9482242 0.7758426  0.4765088  1.47115422  1.9183814  2.4307709
#> EM[4,2] -0.4426829 0.9788584 -2.4367839 -1.02811132 -0.4318534  0.2094253
#> EM[6,2] -1.3437807 1.1556411 -3.6354415 -2.07857899 -1.3360556 -0.5623256
#> EM[4,3]  1.1581781 0.9507915 -0.8832312  0.63609257  1.1780988  1.6982441
#> EM[5,4] -1.0243814 1.3937606 -3.9979942 -1.86329121 -1.0459227 -0.1277621
#>             97.5%     Rhat n.eff
#> EM[2,1] 2.2878379 1.002551   950
#> EM[3,1] 1.6288050 1.000769  2700
#> EM[4,1] 1.8409559 1.013761   180
#> EM[5,1] 3.4738856 1.020261   120
#> EM[4,2] 1.4403666 1.005893   370
#> EM[6,2] 0.8096632 1.004176   640
#> EM[4,3] 3.0111358 1.007616   900
#> EM[5,4] 1.6928937 1.008144   280
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  1.047791184 1.41422794 0.0013260332 0.11148119 0.5280905
#> dev.o[2,1]  0.865410838 1.18601626 0.0008175067 0.09154774 0.4125855
#> dev.o[3,1]  1.067812078 1.47142153 0.0008892917 0.11211683 0.4878154
#> dev.o[4,1]  1.036082735 1.48941562 0.0007316148 0.09899495 0.4604122
#> dev.o[5,1]  1.054932911 1.51368297 0.0010529452 0.11907932 0.4794075
#> dev.o[6,1]  1.091351949 1.43976211 0.0016207299 0.12115241 0.5446234
#> dev.o[7,1]  1.061814050 1.40034433 0.0009107274 0.11904944 0.5030571
#> dev.o[8,1]  0.006072347 0.06791901 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.913965335 1.29879137 0.0008120852 0.10088211 0.4179282
#> dev.o[10,1] 1.113203168 1.46684593 0.0012896225 0.12775116 0.5465508
#> dev.o[11,1] 0.808742500 1.17088322 0.0008619108 0.08202784 0.3694161
#> dev.o[1,2]  0.989608128 1.36684826 0.0008880597 0.10944062 0.4570710
#> dev.o[2,2]  1.155142189 1.55728754 0.0015595013 0.13059169 0.5719796
#> dev.o[3,2]  1.084027952 1.51984455 0.0018489343 0.12307007 0.5155972
#> dev.o[4,2]  0.990276544 1.41272125 0.0010466801 0.10350699 0.4488634
#> dev.o[5,2]  1.049817423 1.44370947 0.0013709118 0.10855408 0.4850150
#> dev.o[6,2]  1.244819906 1.58039727 0.0019240963 0.15016530 0.6472277
#> dev.o[7,2]  0.822117264 1.16397001 0.0007187420 0.07264284 0.3579995
#> dev.o[8,2]  0.014795129 0.15366550 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.046927456 1.47393522 0.0008088997 0.09970632 0.4771998
#> dev.o[10,2] 1.413965840 1.80270808 0.0018169081 0.16410201 0.7156098
#> dev.o[11,2] 1.138962626 1.60517858 0.0011280919 0.11747183 0.5433881
#> dev.o[9,3]  1.029933457 1.37906313 0.0014949699 0.10896863 0.5023613
#> dev.o[10,3] 0.903844197 1.22726781 0.0008543420 0.10399693 0.4169580
#> dev.o[11,3] 0.993657362 1.33363327 0.0013095559 0.10310847 0.4809555
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.436543e+00 4.91048978 1.000900  2700
#> dev.o[2,1]  1.169420e+00 4.20844430 1.000491  2700
#> dev.o[3,1]  1.390010e+00 5.51907966 1.000783  2700
#> dev.o[4,1]  1.342564e+00 5.36549910 1.001727  1600
#> dev.o[5,1]  1.332266e+00 5.52531399 1.001134  2700
#> dev.o[6,1]  1.497338e+00 5.12764884 1.000974  2700
#> dev.o[7,1]  1.460859e+00 5.05964466 1.001305  2300
#> dev.o[8,1]  2.131628e-14 0.01290846 1.047875  2700
#> dev.o[9,1]  1.214489e+00 4.39660495 1.007777   540
#> dev.o[10,1] 1.545683e+00 5.32376229 1.002108  1200
#> dev.o[11,1] 1.070693e+00 3.84692585 1.006299   450
#> dev.o[1,2]  1.371110e+00 4.56961500 1.006118  1400
#> dev.o[2,2]  1.562921e+00 5.64108892 1.000884  2700
#> dev.o[3,2]  1.435416e+00 5.55302781 1.002746   870
#> dev.o[4,2]  1.276880e+00 4.93229641 1.003525  2700
#> dev.o[5,2]  1.392663e+00 5.34748307 1.006028   530
#> dev.o[6,2]  1.772479e+00 5.50153024 1.000878  2700
#> dev.o[7,2]  1.112836e+00 4.26122654 1.001182  2700
#> dev.o[8,2]  2.664535e-14 0.04558705 1.018885  2700
#> dev.o[9,2]  1.422190e+00 5.33384375 1.001173  2700
#> dev.o[10,2] 1.979518e+00 6.68481556 1.002893  1600
#> dev.o[11,2] 1.533016e+00 5.51800572 1.000643  2700
#> dev.o[9,3]  1.429233e+00 4.82147315 1.000945  2700
#> dev.o[10,3] 1.196922e+00 4.47969944 1.003220   720
#> dev.o[11,3] 1.351449e+00 4.59779796 1.001362  2700
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  27.184448 4.73912881 18.6515149 23.753643 26.933279 30.507568
#> hat.par[2,1]   2.758915 1.30645229  0.7297723  1.811037  2.598817  3.521925
#> hat.par[3,1]  10.094793 1.20457228  7.1006834  9.447719 10.349416 10.977051
#> hat.par[4,1]  22.987654 2.52749869 17.5098456 21.341848 23.172805 24.776438
#> hat.par[5,1]   7.985495 2.12239334  4.0006785  6.467564  7.937665  9.446700
#> hat.par[6,1]   3.257923 0.96811272  1.3875373  2.575666  3.256342  3.943786
#> hat.par[7,1]   2.429774 1.23141849  0.6411334  1.523791  2.222870  3.079472
#> hat.par[8,1]  11.997011 0.03302748 11.9935475 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.419470 2.65890240 10.2993514 13.612000 15.458236 17.213264
#> hat.par[10,1]  3.139599 1.19093929  1.1353896  2.290754  3.034286  3.878184
#> hat.par[11,1]  4.503859 1.51857042  1.7993154  3.420878  4.416279  5.506224
#> hat.par[1,2]  38.088585 5.07648508 28.4006690 34.628768 38.100896 41.566781
#> hat.par[2,2]   4.214340 1.67685750  1.5780966  3.018476  3.989099  5.204612
#> hat.par[3,2]   7.960529 1.43113660  4.8545553  6.974891  8.131891  9.028811
#> hat.par[4,2]  16.113061 2.46191061 11.0145812 14.465407 16.163599 17.879679
#> hat.par[5,2]   2.899230 1.47263761  0.6530490  1.781128  2.694258  3.822242
#> hat.par[6,2]   3.770722 0.93742813  1.7747657  3.149698  3.843671  4.474244
#> hat.par[7,2]   2.535601 1.22667555  0.6672922  1.636446  2.327085  3.273311
#> hat.par[8,2]  14.992795 0.07360422 14.9772238 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.768200 2.82590116 11.3451675 14.821311 16.707434 18.713017
#> hat.par[10,2]  3.234922 1.33805561  1.0686117  2.216290  3.144150  4.118679
#> hat.par[11,2]  7.032499 1.59160411  3.9246479  5.942554  7.056583  8.101149
#> hat.par[9,3]  21.822182 2.21623580 17.3562400 20.310289 21.897586 23.423300
#> hat.par[10,3]  8.478028 1.38967400  5.5536134  7.566463  8.554477  9.540038
#> hat.par[11,3] 11.442769 1.64579290  7.7591102 10.411773 11.609285 12.640954
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.637767 1.004905   450
#> hat.par[2,1]   5.779523 1.002801  1200
#> hat.par[3,1]  11.729478 1.012627   320
#> hat.par[4,1]  27.353975 1.012764   160
#> hat.par[5,1]  12.272731 1.005854   370
#> hat.par[6,1]   5.089900 1.003116  2200
#> hat.par[7,1]   5.358888 1.004627   850
#> hat.par[8,1]  12.000000 1.047875  2700
#> hat.par[9,1]  20.536434 1.001742  1500
#> hat.par[10,1]  5.871172 1.012832   230
#> hat.par[11,1]  7.580661 1.018165   270
#> hat.par[1,2]  48.093388 1.003161   740
#> hat.par[2,2]   8.109663 1.001855  1400
#> hat.par[3,2]  10.317706 1.001739  1500
#> hat.par[4,2]  20.705292 1.002451  1000
#> hat.par[5,2]   6.236328 1.004462   550
#> hat.par[6,2]   5.320042 1.004746   990
#> hat.par[7,2]   5.419923 1.001578  1800
#> hat.par[8,2]  15.000000 1.018885  2700
#> hat.par[9,2]  22.406564 1.003990   570
#> hat.par[10,2]  6.166170 1.004349   550
#> hat.par[11,2] 10.102823 1.008320   760
#> hat.par[9,3]  25.782967 1.006838   310
#> hat.par[10,3] 10.820587 1.009169   350
#> hat.par[11,3] 14.074949 1.006715   340
#> 
#> $leverage_o
#>  [1]  1.046215532  0.603196469  1.062279120  1.036059185  1.054887471
#>  [6]  0.711990282  0.914768108 -0.004064329  0.892519319  0.798593830
#> [11]  0.728378522  0.989310630  0.981784599  1.083317744  0.988311284
#> [16]  1.045632928  0.835137909  0.673091706  0.013828145  0.977435027
#> [21]  0.680811020  0.845475276  0.789053498  0.790856068  0.934610359
#> 
#> $sign_dev_o
#>  [1] -1 -1 -1  1  1  1  1 -1 -1  1 -1 -1  1  1 -1  1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>          mean        sd       2.5%       25%       50%      75%    97.5%
#> tau 0.6848891 0.3777458 0.09249689 0.3953515 0.6493151 0.919126 1.506262
#>         Rhat n.eff
#> tau 1.090781   190
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 44.41855 20.47348 23.94507
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "4", fit using jags,
#>  3 chains, each with 1000 iterations (first 100 discarded)
#>  n.sims = 2700 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.596   0.835 -1.054  0.086  0.578  1.062  2.288 1.003   950
#> EM[3,1]         0.198   0.701 -1.168 -0.243  0.180  0.633  1.629 1.001  2700
#> EM[4,1]         0.871   0.551 -0.329  0.546  0.902  1.241  1.841 1.014   180
#> EM[5,1]         1.948   0.776  0.477  1.471  1.918  2.431  3.474 1.020   120
#> EM[4,2]        -0.443   0.979 -2.437 -1.028 -0.432  0.209  1.440 1.006   370
#> EM[6,2]        -1.344   1.156 -3.635 -2.079 -1.336 -0.562  0.810 1.004   640
#> EM[4,3]         1.158   0.951 -0.883  0.636  1.178  1.698  3.011 1.008   900
#> EM[5,4]        -1.024   1.394 -3.998 -1.863 -1.046 -0.128  1.693 1.008   280
#> dev.o[1,1]      1.048   1.414  0.001  0.111  0.528  1.437  4.910 1.001  2700
#> dev.o[2,1]      0.865   1.186  0.001  0.092  0.413  1.169  4.208 1.000  2700
#> dev.o[3,1]      1.068   1.471  0.001  0.112  0.488  1.390  5.519 1.001  2700
#> dev.o[4,1]      1.036   1.489  0.001  0.099  0.460  1.343  5.366 1.002  1600
#> dev.o[5,1]      1.055   1.514  0.001  0.119  0.479  1.332  5.525 1.001  2700
#> dev.o[6,1]      1.091   1.440  0.002  0.121  0.545  1.497  5.128 1.001  2700
#> dev.o[7,1]      1.062   1.400  0.001  0.119  0.503  1.461  5.060 1.001  2300
#> dev.o[8,1]      0.006   0.068  0.000  0.000  0.000  0.000  0.013 1.048  2700
#> dev.o[9,1]      0.914   1.299  0.001  0.101  0.418  1.214  4.397 1.008   540
#> dev.o[10,1]     1.113   1.467  0.001  0.128  0.547  1.546  5.324 1.002  1200
#> dev.o[11,1]     0.809   1.171  0.001  0.082  0.369  1.071  3.847 1.006   450
#> dev.o[1,2]      0.990   1.367  0.001  0.109  0.457  1.371  4.570 1.006  1400
#> dev.o[2,2]      1.155   1.557  0.002  0.131  0.572  1.563  5.641 1.001  2700
#> dev.o[3,2]      1.084   1.520  0.002  0.123  0.516  1.435  5.553 1.003   870
#> dev.o[4,2]      0.990   1.413  0.001  0.104  0.449  1.277  4.932 1.004  2700
#> dev.o[5,2]      1.050   1.444  0.001  0.109  0.485  1.393  5.348 1.006   530
#> dev.o[6,2]      1.245   1.580  0.002  0.150  0.647  1.772  5.502 1.001  2700
#> dev.o[7,2]      0.822   1.164  0.001  0.073  0.358  1.113  4.261 1.001  2700
#> dev.o[8,2]      0.015   0.154  0.000  0.000  0.000  0.000  0.046 1.019  2700
#> dev.o[9,2]      1.047   1.474  0.001  0.100  0.477  1.422  5.334 1.001  2700
#> dev.o[10,2]     1.414   1.803  0.002  0.164  0.716  1.980  6.685 1.003  1600
#> dev.o[11,2]     1.139   1.605  0.001  0.117  0.543  1.533  5.518 1.001  2700
#> dev.o[9,3]      1.030   1.379  0.001  0.109  0.502  1.429  4.821 1.001  2700
#> dev.o[10,3]     0.904   1.227  0.001  0.104  0.417  1.197  4.480 1.003   720
#> dev.o[11,3]     0.994   1.334  0.001  0.103  0.481  1.351  4.598 1.001  2700
#> hat.par[1,1]   27.184   4.739 18.652 23.754 26.933 30.508 36.638 1.005   450
#> hat.par[2,1]    2.759   1.306  0.730  1.811  2.599  3.522  5.780 1.003  1200
#> hat.par[3,1]   10.095   1.205  7.101  9.448 10.349 10.977 11.729 1.013   320
#> hat.par[4,1]   22.988   2.527 17.510 21.342 23.173 24.776 27.354 1.013   160
#> hat.par[5,1]    7.985   2.122  4.001  6.468  7.938  9.447 12.273 1.006   370
#> hat.par[6,1]    3.258   0.968  1.388  2.576  3.256  3.944  5.090 1.003  2200
#> hat.par[7,1]    2.430   1.231  0.641  1.524  2.223  3.079  5.359 1.005   850
#> hat.par[8,1]   11.997   0.033 11.994 12.000 12.000 12.000 12.000 1.048  2700
#> hat.par[9,1]   15.419   2.659 10.299 13.612 15.458 17.213 20.536 1.002  1500
#> hat.par[10,1]   3.140   1.191  1.135  2.291  3.034  3.878  5.871 1.013   230
#> hat.par[11,1]   4.504   1.519  1.799  3.421  4.416  5.506  7.581 1.018   270
#> hat.par[1,2]   38.089   5.076 28.401 34.629 38.101 41.567 48.093 1.003   740
#> hat.par[2,2]    4.214   1.677  1.578  3.018  3.989  5.205  8.110 1.002  1400
#> hat.par[3,2]    7.961   1.431  4.855  6.975  8.132  9.029 10.318 1.002  1500
#> hat.par[4,2]   16.113   2.462 11.015 14.465 16.164 17.880 20.705 1.002  1000
#> hat.par[5,2]    2.899   1.473  0.653  1.781  2.694  3.822  6.236 1.004   550
#> hat.par[6,2]    3.771   0.937  1.775  3.150  3.844  4.474  5.320 1.005   990
#> hat.par[7,2]    2.536   1.227  0.667  1.636  2.327  3.273  5.420 1.002  1800
#> hat.par[8,2]   14.993   0.074 14.977 15.000 15.000 15.000 15.000 1.019  2700
#> hat.par[9,2]   16.768   2.826 11.345 14.821 16.707 18.713 22.407 1.004   570
#> hat.par[10,2]   3.235   1.338  1.069  2.216  3.144  4.119  6.166 1.004   550
#> hat.par[11,2]   7.032   1.592  3.925  5.943  7.057  8.101 10.103 1.008   760
#> hat.par[9,3]   21.822   2.216 17.356 20.310 21.898 23.423 25.783 1.007   310
#> hat.par[10,3]   8.478   1.390  5.554  7.566  8.554  9.540 10.821 1.009   350
#> hat.par[11,3]  11.443   1.646  7.759 10.412 11.609 12.641 14.075 1.007   340
#> m.tau           0.685   0.479  0.062  0.314  0.582  0.954  1.831 1.010   350
#> resdev.o[1]     2.037   1.967  0.056  0.607  1.503  2.821  6.999 1.001  2700
#> resdev.o[2]     2.021   1.900  0.072  0.664  1.467  2.769  6.871 1.001  2700
#> resdev.o[3]     2.152   2.146  0.055  0.664  1.462  2.932  7.921 1.006   680
#> resdev.o[4]     2.026   2.030  0.057  0.565  1.401  2.866  7.422 1.001  2700
#> resdev.o[5]     2.105   2.076  0.057  0.617  1.453  2.906  7.864 1.008   440
#> resdev.o[6]     2.336   1.912  0.096  0.978  1.897  3.195  7.346 1.003  2700
#> resdev.o[7]     1.884   1.796  0.054  0.588  1.336  2.611  6.761 1.003   800
#> resdev.o[8]     0.021   0.199  0.000  0.000  0.000  0.000  0.063 1.012  2700
#> resdev.o[9]     2.991   2.364  0.202  1.244  2.422  4.099  8.993 1.003   760
#> resdev.o[10]    3.431   2.459  0.305  1.595  2.901  4.673  9.540 1.002  1500
#> resdev.o[11]    2.941   2.349  0.212  1.257  2.384  3.971  8.810 1.007   300
#> tau             0.685   0.378  0.092  0.395  0.649  0.919  1.506 1.091   190
#> totresdev.o    23.945   6.874 12.238 19.150 23.348 27.907 39.650 1.004   540
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $m_tau
#>            mean        sd       2.5%      25%       50%       75%    97.5%
#> m.tau 0.6854763 0.4793458 0.06155042 0.313925 0.5822261 0.9539289 1.830828
#>           Rhat n.eff
#> m.tau 1.009623   350
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
# }