Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al., (2013a) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli, 2019; Spineli et al., 2021). This model offers a global evaluation of the plausibility of the consistency assumption in the network (Dias et al. (2013b)).

Usage,

run_ume(full, n_iter, n_burnin, n_chains, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

EM	The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o	The deviance contribution of each trial-arm based on the observed outcome.

hat_par	The fitted outcome at each trial-arm.

tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o	The leverage for the observed outcome at each trial-arm.

sign_dev_o	The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment	A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit	An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var_misspar from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then usenrun_ume (see 'Examples').

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al., (2013a) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al., (2013a) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has been also refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the basic parameters of these trials, that is, effect parameters between non-baseline arms and the baseline arm (Dias et al., 2013b; Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. submitted 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013a;33(5):641--56. doi: 10.1177/0272989X12455847

Dias S, Sutton AJ, Ades AE, Welton NJ. Evidence synthesis for decision making 2: a generalized linear modeling framework for pairwise and network meta-analysis of randomized controlled trials. Med Decis Making 2013b;33(5):607--617. doi: 10.1177/0272989X12458724

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 3045
#> 
#> Initializing model
#> 
#> $EM
#>               mean        sd       2.5%        25%        50%         75%
#> EM[2,1]  0.5786762 0.8094871 -1.1263875  0.1675792  0.5680382  1.00290699
#> EM[3,1]  0.2558978 0.6828836 -1.0211616 -0.1621946  0.2203778  0.62379349
#> EM[4,1]  0.8034261 0.5787301 -0.4119036  0.4647182  0.8387378  1.15355576
#> EM[5,1]  1.9683994 0.8068848  0.3998492  1.4255265  1.9531398  2.47459844
#> EM[4,2] -0.5079962 0.9210405 -2.4076038 -1.0545542 -0.5172771  0.03305231
#> EM[6,2] -1.2855224 1.1539427 -3.8611454 -1.9085741 -1.2421811 -0.57820793
#> EM[4,3]  1.2007963 0.9241550 -0.7116659  0.7149614  1.2298682  1.69732965
#> EM[5,4] -0.7422757 1.3761759 -3.3434012 -1.7454511 -0.6833021  0.22450867
#>             97.5%     Rhat n.eff
#> EM[2,1] 2.2830141 1.036819  1400
#> EM[3,1] 1.7033841 1.020857   160
#> EM[4,1] 1.9741030 1.036596   170
#> EM[5,1] 3.5549693 1.048663    58
#> EM[4,2] 1.3712262 1.030090   160
#> EM[6,2] 0.9565648 1.011329   350
#> EM[4,3] 3.0590291 1.007234   430
#> EM[5,4] 1.8890804 1.033099   110
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  1.039288339 1.50176483 0.0007830222 0.10150011 0.4632443
#> dev.o[2,1]  0.902262408 1.25023025 0.0011827801 0.08892626 0.4166665
#> dev.o[3,1]  1.018004819 1.33194197 0.0008165420 0.10493004 0.4960851
#> dev.o[4,1]  1.036488792 1.39187488 0.0008113444 0.11478319 0.4946131
#> dev.o[5,1]  1.045673937 1.48490308 0.0011635071 0.09642179 0.4559480
#> dev.o[6,1]  1.106294812 1.40092527 0.0010987213 0.12618607 0.5736710
#> dev.o[7,1]  0.967594281 1.30137592 0.0009223628 0.11843324 0.4777844
#> dev.o[8,1]  0.005634759 0.09635580 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.940766279 1.27885024 0.0006418108 0.09510433 0.4362069
#> dev.o[10,1] 1.049135638 1.32445109 0.0010347049 0.12887123 0.5390411
#> dev.o[11,1] 0.781225111 1.02464735 0.0010795238 0.09690137 0.3958773
#> dev.o[1,2]  1.004119634 1.38985448 0.0008079059 0.10032464 0.4800671
#> dev.o[2,2]  1.196649988 1.60971964 0.0008727009 0.12323796 0.5628581
#> dev.o[3,2]  1.100106906 1.44314996 0.0019929636 0.11969051 0.5432649
#> dev.o[4,2]  1.028861800 1.42510436 0.0008130067 0.10710349 0.4815055
#> dev.o[5,2]  1.081968785 1.57553536 0.0007268241 0.11016585 0.4941316
#> dev.o[6,2]  1.290405594 1.65332609 0.0013065919 0.14836174 0.6640884
#> dev.o[7,2]  0.812395047 1.19348796 0.0012029437 0.08141467 0.3352691
#> dev.o[8,2]  0.006506969 0.07713282 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.042023108 1.51694471 0.0007096002 0.09831862 0.4508691
#> dev.o[10,2] 1.415118059 1.73747256 0.0016202117 0.17315738 0.7758828
#> dev.o[11,2] 1.143463548 1.46139416 0.0014138257 0.11645224 0.5787517
#> dev.o[9,3]  1.182665510 1.55801114 0.0012892021 0.11703703 0.5810425
#> dev.o[10,3] 0.834046083 1.18189005 0.0005051931 0.07479436 0.3618870
#> dev.o[11,3] 1.126914722 1.60069055 0.0007543536 0.09940471 0.5056258
#>                      75%       97.5%     Rhat n.eff
#> dev.o[1,1]  1.355449e+00 5.165308180 1.002266  1100
#> dev.o[2,1]  1.205548e+00 4.421093367 1.002601  2700
#> dev.o[3,1]  1.434195e+00 4.658970752 1.012236   170
#> dev.o[4,1]  1.368420e+00 5.078037065 1.001338  2200
#> dev.o[5,1]  1.419020e+00 5.136769316 1.001027  2700
#> dev.o[6,1]  1.528924e+00 4.984478383 1.005397   600
#> dev.o[7,1]  1.310564e+00 4.476334269 1.002072  1200
#> dev.o[8,1]  3.463896e-14 0.009014163 1.195102  2600
#> dev.o[9,1]  1.274768e+00 4.759187557 1.000876  2700
#> dev.o[10,1] 1.477097e+00 4.877065811 1.004981   440
#> dev.o[11,1] 1.111445e+00 3.588961490 1.013131   160
#> dev.o[1,2]  1.352975e+00 5.141281973 1.001657  2200
#> dev.o[2,2]  1.638590e+00 5.713486048 1.000551  2700
#> dev.o[3,2]  1.489319e+00 5.161297029 1.001713  1600
#> dev.o[4,2]  1.370655e+00 5.190639508 1.003770   790
#> dev.o[5,2]  1.352753e+00 5.542499905 1.000541  2700
#> dev.o[6,2]  1.819309e+00 6.077085322 1.006946   310
#> dev.o[7,2]  1.040344e+00 4.186497083 1.000818  2700
#> dev.o[8,2]  7.993606e-14 0.022594353 1.003571  2700
#> dev.o[9,2]  1.352707e+00 5.510033012 1.001881  2700
#> dev.o[10,2] 2.012843e+00 6.327540673 1.008558   260
#> dev.o[11,2] 1.624719e+00 5.181288254 1.005425   500
#> dev.o[9,3]  1.635858e+00 5.473420620 1.001342  2700
#> dev.o[10,3] 1.090854e+00 4.236178186 1.001793  1500
#> dev.o[11,3] 1.502026e+00 5.774095619 1.002687   890
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  27.275638 4.70029809 18.4812245 23.998876 27.197124 30.390185
#> hat.par[2,1]   2.742689 1.34053183  0.6629610  1.758273  2.551355  3.578931
#> hat.par[3,1]   9.894803 1.24658595  6.9962101  9.117823 10.060646 10.877976
#> hat.par[4,1]  23.163445 2.49482432 17.7930883 21.499247 23.317975 24.986513
#> hat.par[5,1]   7.929928 2.11189067  4.0424971  6.451601  7.851599  9.317449
#> hat.par[6,1]   3.256132 0.97700652  1.4171494  2.552389  3.234011  3.965685
#> hat.par[7,1]   2.470870 1.20096429  0.7151836  1.592941  2.261509  3.162392
#> hat.par[8,1]  11.997275 0.04484649 11.9954938 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.260604 2.71042946  9.7346896 13.458774 15.383847 17.163295
#> hat.par[10,1]  3.296956 1.27058949  1.2299597  2.362595  3.147241  4.092001
#> hat.par[11,1]  4.518344 1.49686042  1.9017853  3.385152  4.465885  5.553320
#> hat.par[1,2]  37.736421 5.09090669 28.3501610 34.180886 37.606182 41.154700
#> hat.par[2,2]   4.244040 1.73360237  1.5630108  2.959638  4.040042  5.286633
#> hat.par[3,2]   8.194150 1.38686110  5.0041186  7.310015  8.403116  9.265818
#> hat.par[4,2]  15.849272 2.52235998 10.8395176 14.123413 15.837410 17.670031
#> hat.par[5,2]   3.040115 1.54719474  0.6687718  1.887448  2.843418  3.968605
#> hat.par[6,2]   3.772099 0.95570824  1.7110875  3.147299  3.843042  4.490436
#> hat.par[7,2]   2.501647 1.21786394  0.6169419  1.603152  2.328308  3.225109
#> hat.par[8,2]  14.996796 0.03753444 14.9887071 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.998549 2.74597996 11.6475123 15.161495 16.950924 18.847440
#> hat.par[10,2]  3.270846 1.29840523  1.0348162  2.310264  3.186020  4.125765
#> hat.par[11,2]  6.828622 1.47866751  4.0126741  5.768068  6.841943  7.852533
#> hat.par[9,3]  21.473974 2.26274912 16.8572176 19.950441 21.602041 23.080981
#> hat.par[10,3]  8.493613 1.33389262  5.6220692  7.637043  8.614370  9.459080
#> hat.par[11,3] 11.630014 1.62662339  8.2644282 10.575798 11.718030 12.844903
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.784597 1.005169   660
#> hat.par[2,1]   5.837367 1.039477    94
#> hat.par[3,1]  11.660191 1.015561   150
#> hat.par[4,1]  27.612007 1.014096   150
#> hat.par[5,1]  12.205983 1.003783   610
#> hat.par[6,1]   5.102878 1.003086   760
#> hat.par[7,1]   5.225594 1.004459  1100
#> hat.par[8,1]  12.000000 1.195102  2600
#> hat.par[9,1]  20.299728 1.009142   230
#> hat.par[10,1]  6.147654 1.049361    47
#> hat.par[11,1]  7.505167 1.023393   140
#> hat.par[1,2]  48.054436 1.000757  2700
#> hat.par[2,2]   8.214199 1.003617   900
#> hat.par[3,2]  10.292788 1.011528   380
#> hat.par[4,2]  20.563551 1.013120   160
#> hat.par[5,2]   6.583518 1.009257   240
#> hat.par[6,2]   5.386803 1.013188   170
#> hat.par[7,2]   5.321651 1.009655   230
#> hat.par[8,2]  15.000000 1.003571  2700
#> hat.par[9,2]  22.431059 1.012843   200
#> hat.par[10,2]  6.048023 1.021264   110
#> hat.par[11,2]  9.719387 1.016080   130
#> hat.par[9,3]  25.525578 1.007603   280
#> hat.par[10,3] 10.765570 1.012925   190
#> hat.par[11,3] 14.454548 1.028403    83
#> 
#> $leverage_o
#>  [1]  1.035775393  0.650240752  1.011543779  1.032328090  1.044612093
#>  [6]  0.725146012  0.842094081 -0.005822368  0.932480586  0.842653799
#> [11]  0.696290370  1.001476687  1.036695453  1.082335405  1.025386199
#> [16]  1.081321945  0.879206854  0.680712180 -0.002665839  0.924649207
#> [21]  0.642670786  0.715207598  0.785634344  0.727507182  1.005399377
#> 
#> $sign_dev_o
#>  [1] -1 -1  1 -1  1  1  1 -1 -1  1 -1  1  1 -1  1 -1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>          mean        sd      2.5%       25%       50%       75%    97.5%
#> tau 0.6274438 0.3967543 0.0601406 0.2877601 0.5746103 0.8785762 1.476585
#>         Rhat n.eff
#> tau 1.185131    18
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 44.55049 20.39288 24.15761
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "40", fit using jags,
#>  3 chains, each with 1000 iterations (first 100 discarded)
#>  n.sims = 2700 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.579   0.809 -1.126  0.168  0.568  1.003  2.283 1.037  1400
#> EM[3,1]         0.256   0.683 -1.021 -0.162  0.220  0.624  1.703 1.021   160
#> EM[4,1]         0.803   0.579 -0.412  0.465  0.839  1.154  1.974 1.037   170
#> EM[5,1]         1.968   0.807  0.400  1.426  1.953  2.475  3.555 1.049    58
#> EM[4,2]        -0.508   0.921 -2.408 -1.055 -0.517  0.033  1.371 1.030   160
#> EM[6,2]        -1.286   1.154 -3.861 -1.909 -1.242 -0.578  0.957 1.011   350
#> EM[4,3]         1.201   0.924 -0.712  0.715  1.230  1.697  3.059 1.007   430
#> EM[5,4]        -0.742   1.376 -3.343 -1.745 -0.683  0.225  1.889 1.033   110
#> dev.o[1,1]      1.039   1.502  0.001  0.102  0.463  1.355  5.165 1.002  1100
#> dev.o[2,1]      0.902   1.250  0.001  0.089  0.417  1.206  4.421 1.003  2700
#> dev.o[3,1]      1.018   1.332  0.001  0.105  0.496  1.434  4.659 1.012   170
#> dev.o[4,1]      1.036   1.392  0.001  0.115  0.495  1.368  5.078 1.001  2200
#> dev.o[5,1]      1.046   1.485  0.001  0.096  0.456  1.419  5.137 1.001  2700
#> dev.o[6,1]      1.106   1.401  0.001  0.126  0.574  1.529  4.984 1.005   600
#> dev.o[7,1]      0.968   1.301  0.001  0.118  0.478  1.311  4.476 1.002  1200
#> dev.o[8,1]      0.006   0.096  0.000  0.000  0.000  0.000  0.009 1.195  2600
#> dev.o[9,1]      0.941   1.279  0.001  0.095  0.436  1.275  4.759 1.001  2700
#> dev.o[10,1]     1.049   1.324  0.001  0.129  0.539  1.477  4.877 1.005   440
#> dev.o[11,1]     0.781   1.025  0.001  0.097  0.396  1.111  3.589 1.013   160
#> dev.o[1,2]      1.004   1.390  0.001  0.100  0.480  1.353  5.141 1.002  2200
#> dev.o[2,2]      1.197   1.610  0.001  0.123  0.563  1.639  5.713 1.001  2700
#> dev.o[3,2]      1.100   1.443  0.002  0.120  0.543  1.489  5.161 1.002  1600
#> dev.o[4,2]      1.029   1.425  0.001  0.107  0.482  1.371  5.191 1.004   790
#> dev.o[5,2]      1.082   1.576  0.001  0.110  0.494  1.353  5.543 1.001  2700
#> dev.o[6,2]      1.290   1.653  0.001  0.148  0.664  1.819  6.077 1.007   310
#> dev.o[7,2]      0.812   1.193  0.001  0.081  0.335  1.040  4.187 1.001  2700
#> dev.o[8,2]      0.007   0.077  0.000  0.000  0.000  0.000  0.023 1.004  2700
#> dev.o[9,2]      1.042   1.517  0.001  0.098  0.451  1.353  5.510 1.002  2700
#> dev.o[10,2]     1.415   1.737  0.002  0.173  0.776  2.013  6.328 1.009   260
#> dev.o[11,2]     1.143   1.461  0.001  0.116  0.579  1.625  5.181 1.005   500
#> dev.o[9,3]      1.183   1.558  0.001  0.117  0.581  1.636  5.473 1.001  2700
#> dev.o[10,3]     0.834   1.182  0.001  0.075  0.362  1.091  4.236 1.002  1500
#> dev.o[11,3]     1.127   1.601  0.001  0.099  0.506  1.502  5.774 1.003   890
#> hat.par[1,1]   27.276   4.700 18.481 23.999 27.197 30.390 36.785 1.005   660
#> hat.par[2,1]    2.743   1.341  0.663  1.758  2.551  3.579  5.837 1.039    94
#> hat.par[3,1]    9.895   1.247  6.996  9.118 10.061 10.878 11.660 1.016   150
#> hat.par[4,1]   23.163   2.495 17.793 21.499 23.318 24.987 27.612 1.014   150
#> hat.par[5,1]    7.930   2.112  4.042  6.452  7.852  9.317 12.206 1.004   610
#> hat.par[6,1]    3.256   0.977  1.417  2.552  3.234  3.966  5.103 1.003   760
#> hat.par[7,1]    2.471   1.201  0.715  1.593  2.262  3.162  5.226 1.004  1100
#> hat.par[8,1]   11.997   0.045 11.995 12.000 12.000 12.000 12.000 1.195  2600
#> hat.par[9,1]   15.261   2.710  9.735 13.459 15.384 17.163 20.300 1.009   230
#> hat.par[10,1]   3.297   1.271  1.230  2.363  3.147  4.092  6.148 1.049    47
#> hat.par[11,1]   4.518   1.497  1.902  3.385  4.466  5.553  7.505 1.023   140
#> hat.par[1,2]   37.736   5.091 28.350 34.181 37.606 41.155 48.054 1.001  2700
#> hat.par[2,2]    4.244   1.734  1.563  2.960  4.040  5.287  8.214 1.004   900
#> hat.par[3,2]    8.194   1.387  5.004  7.310  8.403  9.266 10.293 1.012   380
#> hat.par[4,2]   15.849   2.522 10.840 14.123 15.837 17.670 20.564 1.013   160
#> hat.par[5,2]    3.040   1.547  0.669  1.887  2.843  3.969  6.584 1.009   240
#> hat.par[6,2]    3.772   0.956  1.711  3.147  3.843  4.490  5.387 1.013   170
#> hat.par[7,2]    2.502   1.218  0.617  1.603  2.328  3.225  5.322 1.010   230
#> hat.par[8,2]   14.997   0.038 14.989 15.000 15.000 15.000 15.000 1.004  2700
#> hat.par[9,2]   16.999   2.746 11.648 15.161 16.951 18.847 22.431 1.013   200
#> hat.par[10,2]   3.271   1.298  1.035  2.310  3.186  4.126  6.048 1.021   110
#> hat.par[11,2]   6.829   1.479  4.013  5.768  6.842  7.853  9.719 1.016   130
#> hat.par[9,3]   21.474   2.263 16.857 19.950 21.602 23.081 25.526 1.008   280
#> hat.par[10,3]   8.494   1.334  5.622  7.637  8.614  9.459 10.766 1.013   190
#> hat.par[11,3]  11.630   1.627  8.264 10.576 11.718 12.845 14.455 1.028    83
#> m.tau           0.706   0.493  0.056  0.326  0.604  0.993  1.908 1.035    84
#> resdev.o[1]     2.043   2.069  0.050  0.616  1.435  2.782  7.439 1.001  2700
#> resdev.o[2]     2.099   1.956  0.058  0.649  1.598  2.929  7.007 1.001  2700
#> resdev.o[3]     2.118   1.966  0.059  0.677  1.573  2.988  7.493 1.011   190
#> resdev.o[4]     2.065   1.948  0.053  0.626  1.451  2.917  7.228 1.005   700
#> resdev.o[5]     2.128   2.157  0.058  0.607  1.414  2.928  8.036 1.001  2700
#> resdev.o[6]     2.397   1.981  0.115  0.973  1.962  3.250  7.448 1.015   240
#> resdev.o[7]     1.780   1.801  0.050  0.562  1.265  2.385  6.211 1.007   500
#> resdev.o[8]     0.012   0.136  0.000  0.000  0.000  0.000  0.030 1.066  2700
#> resdev.o[9]     3.165   2.481  0.238  1.316  2.535  4.345  9.482 1.002  2700
#> resdev.o[10]    3.298   2.327  0.308  1.520  2.834  4.525  8.925 1.018   120
#> resdev.o[11]    3.052   2.344  0.259  1.344  2.487  4.116  9.263 1.001  2000
#> tau             0.627   0.397  0.060  0.288  0.575  0.879  1.477 1.185    18
#> totresdev.o    24.158   6.589 12.953 19.589 23.650 28.086 39.164 1.001  2700
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $m_tau
#>            mean        sd       2.5%       25%       50%       75%    97.5%
#> m.tau 0.7057927 0.4925447 0.05595595 0.3256758 0.6035027 0.9930314 1.907963
#>           Rhat n.eff
#> m.tau 1.034841    84
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
# }