Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al. (2013) that has been refined (Spineli, 2021a) and extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al. 2021b; Spineli, 2019). This model offers a global evaluation of the plausibility of the consistency assumption in the network.

run_ume(full, n_iter, n_burnin, n_chains, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM	The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o	The deviance contribution of each trial-arm based on the observed outcome.

hat_par	The fitted outcome at each trial-arm.

tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o	The leverage for the observed outcome at each trial-arm.

sign_dev_o	The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment	A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit	An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then use run_ume (see 'Examples').

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm trial of the network (Spineli, 2021a). The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has been also refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the basic parameters of these trials, that is, effect parameters between non-baseline arms and the baseline arm (Spineli, 2021a).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013;33(5):641--56. doi: 10.1177/0272989X12455847

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

Spineli LM. A Revised Framework to Evaluate the Consistency Assumption Globally in a Network of Interventions. Med Decis Making 2021a. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021b;30(4):958--975. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 3045
#> 
#> Initializing model
#> 
#> $EM
#>               mean        sd       2.5%        25%         50%        75%
#> EM[2,1]  0.6372301 0.8592002 -1.1811915  0.2023541  0.63261455  1.0599322
#> EM[3,1]  0.0708324 0.7382579 -1.4031113 -0.3795196  0.06499456  0.5032018
#> EM[4,1]  0.8581501 0.6000277 -0.4109521  0.5076435  0.90234372  1.2860954
#> EM[5,1]  2.0217810 0.7695262  0.4634034  1.5443419  2.02618953  2.5465689
#> EM[4,2] -0.3783125 0.9676859 -2.4182758 -0.9151041 -0.36823684  0.2180710
#> EM[6,2] -1.1408944 1.1186308 -3.6542093 -1.7564700 -1.08126496 -0.4760136
#> EM[4,3]  1.2040689 1.0626112 -1.0322335  0.6255183  1.18527338  1.7954209
#> EM[5,4] -1.0693795 1.5559008 -4.2166359 -1.8901060 -1.10122683 -0.1612050
#>             97.5%     Rhat n.eff
#> EM[2,1] 2.5324477 1.010299   490
#> EM[3,1] 1.5773257 1.010325   410
#> EM[4,1] 1.8684463 1.027271    91
#> EM[5,1] 3.4169415 1.025052    87
#> EM[4,2] 1.4860246 1.007734   770
#> EM[6,2] 0.9633231 1.006528   630
#> EM[4,3] 3.3186929 1.001538  1800
#> EM[5,4] 1.9306517 1.002740   870
#> 
#> $dev_o
#>                   mean         sd         2.5%        25%       50%
#> dev.o[1,1]  0.97493340 1.34885469 0.0010179202 0.10713668 0.4400531
#> dev.o[2,1]  0.97306384 1.35095211 0.0011939542 0.09288456 0.4508589
#> dev.o[3,1]  1.23349935 1.76359549 0.0009905946 0.10313194 0.5061941
#> dev.o[4,1]  0.96311977 1.29886364 0.0009516776 0.09698722 0.4440019
#> dev.o[5,1]  1.02595538 1.39117460 0.0012999701 0.11984118 0.4907867
#> dev.o[6,1]  1.15623719 1.49951797 0.0010386132 0.12655598 0.5870590
#> dev.o[7,1]  0.99858722 1.26441347 0.0015979769 0.12000807 0.5030681
#> dev.o[8,1]  0.00515026 0.09789809 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.91318518 1.29774066 0.0006801297 0.09119616 0.4236681
#> dev.o[10,1] 1.15065700 1.39945793 0.0013135385 0.13014688 0.6534244
#> dev.o[11,1] 0.75364191 1.12201528 0.0006372803 0.06875495 0.3284457
#> dev.o[1,2]  0.98761596 1.35023594 0.0007116607 0.10007349 0.4473911
#> dev.o[2,2]  1.27217129 1.61531683 0.0008867245 0.16405481 0.6289914
#> dev.o[3,2]  1.05195500 1.46422837 0.0010007564 0.11617106 0.5038351
#> dev.o[4,2]  0.96770103 1.35554419 0.0010282012 0.09067798 0.4169408
#> dev.o[5,2]  0.97336828 1.35892775 0.0007339350 0.09882142 0.4455048
#> dev.o[6,2]  1.26472712 1.59047210 0.0021404909 0.16565449 0.6842824
#> dev.o[7,2]  0.84405296 1.14760659 0.0012099506 0.09144452 0.4022081
#> dev.o[8,2]  0.01062289 0.12548524 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.00442965 1.42420450 0.0010585671 0.10870574 0.4600344
#> dev.o[10,2] 1.35933793 1.69908517 0.0014680998 0.17216984 0.7345099
#> dev.o[11,2] 1.05762173 1.35469829 0.0014741640 0.12694108 0.5542962
#> dev.o[9,3]  1.02168643 1.47401897 0.0013148135 0.10327229 0.4620144
#> dev.o[10,3] 0.82830571 1.23240649 0.0007637126 0.07713641 0.3382053
#> dev.o[11,3] 1.03415866 1.36661442 0.0012977634 0.09989348 0.4851376
#>                      75%       97.5%     Rhat n.eff
#> dev.o[1,1]  1.332988e+00 4.489713933 1.001408  2100
#> dev.o[2,1]  1.312756e+00 4.914989777 1.003055  1600
#> dev.o[3,1]  1.601660e+00 6.546029836 1.000597  2700
#> dev.o[4,1]  1.326613e+00 4.539908080 1.001956  2700
#> dev.o[5,1]  1.367074e+00 5.057595623 1.001708  1600
#> dev.o[6,1]  1.632131e+00 5.305686222 1.009654   280
#> dev.o[7,1]  1.400875e+00 4.526161705 1.004131   580
#> dev.o[8,1]  4.529710e-14 0.006636119 1.251100  1600
#> dev.o[9,1]  1.205510e+00 4.523704157 1.003263  2600
#> dev.o[10,1] 1.632266e+00 5.004710606 1.009723   240
#> dev.o[11,1] 1.004280e+00 3.936587450 1.003160  2700
#> dev.o[1,2]  1.322596e+00 4.739376810 1.001467  2000
#> dev.o[2,2]  1.770474e+00 5.770114727 1.004402  1600
#> dev.o[3,2]  1.373632e+00 5.084540775 1.001265  2400
#> dev.o[4,2]  1.293201e+00 4.830450046 1.000454  2700
#> dev.o[5,2]  1.298990e+00 4.947276506 1.008924   270
#> dev.o[6,2]  1.780104e+00 5.541733728 1.005480   400
#> dev.o[7,2]  1.175065e+00 4.114386297 1.001279  2400
#> dev.o[8,2]  9.325873e-14 0.012912191 1.039060  2700
#> dev.o[9,2]  1.335875e+00 4.820693571 1.000911  2700
#> dev.o[10,2] 1.890988e+00 6.216653580 1.003312   730
#> dev.o[11,2] 1.477355e+00 4.801612411 1.001494  2700
#> dev.o[9,3]  1.309882e+00 5.231653111 1.003884   580
#> dev.o[10,3] 1.084582e+00 4.152583237 1.003944   720
#> dev.o[11,3] 1.428203e+00 4.939538174 1.004282   520
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  26.733023 4.49284031 18.6575147 23.496374 26.644970 29.696773
#> hat.par[2,1]   2.931269 1.36396691  0.8265646  1.913095  2.722751  3.747392
#> hat.par[3,1]  10.071990 1.23776328  7.1896586  9.380599 10.258739 10.998298
#> hat.par[4,1]  22.661939 2.47487344 17.4704297 20.933540 22.867914 24.399453
#> hat.par[5,1]   7.803134 2.08157988  4.0667387  6.255995  7.719144  9.235530
#> hat.par[6,1]   3.208542 0.97730168  1.3562099  2.488104  3.230849  3.935951
#> hat.par[7,1]   2.442965 1.20606391  0.6974870  1.544161  2.231818  3.124486
#> hat.par[8,1]  11.997520 0.04526719 11.9966824 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.560412 2.63582202 10.4740971 13.724891 15.582605 17.322460
#> hat.par[10,1]  3.199622 1.27996954  1.2114487  2.235194  3.032902  4.030973
#> hat.par[11,1]  4.436953 1.49703298  1.9963746  3.346306  4.280104  5.393901
#> hat.par[1,2]  38.020141 5.08128261 28.7031118 34.458347 37.934506 41.291620
#> hat.par[2,2]   4.052832 1.66259727  1.4993717  2.844727  3.827001  4.976924
#> hat.par[3,2]   7.880392 1.44080081  4.8424969  6.916500  8.037385  8.979737
#> hat.par[4,2]  16.181292 2.43189884 11.1522919 14.620122 16.272199 17.833961
#> hat.par[5,2]   3.207369 1.51008239  0.7800296  2.076325  3.036461  4.171381
#> hat.par[6,2]   3.775863 0.94461872  1.7480775  3.163953  3.862968  4.480612
#> hat.par[7,2]   2.589926 1.21760208  0.6630832  1.654205  2.444062  3.358205
#> hat.par[8,2]  14.994817 0.06038518 14.9935453 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.501650 2.82061282 11.1300182 14.493851 16.538485 18.434569
#> hat.par[10,2]  3.180859 1.31733009  0.9973097  2.184501  3.112367  4.059334
#> hat.par[11,2]  6.940736 1.47144454  4.1645684  5.887336  6.924894  7.978698
#> hat.par[9,3]  21.909207 2.23385819 17.1981675 20.471162 22.005158 23.451444
#> hat.par[10,3]  8.580500 1.33953502  5.6556920  7.762027  8.679350  9.549441
#> hat.par[11,3] 11.688572 1.53773651  8.3782147 10.676475 11.782240 12.817293
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  35.896079 1.009141   230
#> hat.par[2,1]   6.095145 1.014176   150
#> hat.par[3,1]  11.818315 1.006784   480
#> hat.par[4,1]  27.006743 1.001126  2700
#> hat.par[5,1]  11.946683 1.007829   270
#> hat.par[6,1]   5.000653 1.007405   300
#> hat.par[7,1]   5.377761 1.006871   310
#> hat.par[8,1]  12.000000 1.251100  1600
#> hat.par[9,1]  20.751690 1.003048  1400
#> hat.par[10,1]  6.060923 1.023250    94
#> hat.par[11,1]  7.737830 1.013780   170
#> hat.par[1,2]  48.601096 1.006112   350
#> hat.par[2,2]   7.912450 1.004342   550
#> hat.par[3,2]  10.169922 1.005671   380
#> hat.par[4,2]  20.711421 1.000709  2700
#> hat.par[5,2]   6.460412 1.008682   400
#> hat.par[6,2]   5.325129 1.010245   340
#> hat.par[7,2]   5.294272 1.006735   330
#> hat.par[8,2]  15.000000 1.039060  2700
#> hat.par[9,2]  21.996859 1.007195   300
#> hat.par[10,2]  5.971229 1.007656   400
#> hat.par[11,2]  9.766170 1.002577  2700
#> hat.par[9,3]  25.919292 1.009246   340
#> hat.par[10,3] 10.878987 1.006822   480
#> hat.par[11,3] 14.252507 1.021190   110
#> 
#> $leverage_o
#>  [1]  0.971604688  0.592485105  1.230328399  0.945712576  1.017548078
#>  [6]  0.726083298  0.858665487 -0.007699946  0.874936334  0.880232233
#> [11]  0.692816207  0.987600572  1.015396518  1.045503058  0.962641098
#> [16]  0.956560110  0.849367834  0.665313437  0.007110697  0.974776553
#> [21]  0.683656586  0.706586622  0.814104869  0.754415017  0.888230592
#> 
#> $sign_dev_o
#>  [1]  1 -1 -1  1  1  1  1 -1 -1  1 -1 -1  1  1 -1 -1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>          mean        sd       2.5%       25%       50%       75%    97.5%
#> tau 0.6624023 0.4687415 0.04064863 0.2658847 0.6047081 0.9569402 1.757694
#>         Rhat n.eff
#> tau 1.185855    19
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 43.91976 20.09398 23.82579
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "32", fit using jags,
#>  3 chains, each with 1000 iterations (first 100 discarded)
#>  n.sims = 2700 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.637   0.859 -1.181  0.202  0.633  1.060  2.532 1.010   490
#> EM[3,1]         0.071   0.738 -1.403 -0.380  0.065  0.503  1.577 1.010   410
#> EM[4,1]         0.858   0.600 -0.411  0.508  0.902  1.286  1.868 1.027    91
#> EM[5,1]         2.022   0.770  0.463  1.544  2.026  2.547  3.417 1.025    87
#> EM[4,2]        -0.378   0.968 -2.418 -0.915 -0.368  0.218  1.486 1.008   770
#> EM[6,2]        -1.141   1.119 -3.654 -1.756 -1.081 -0.476  0.963 1.007   630
#> EM[4,3]         1.204   1.063 -1.032  0.626  1.185  1.795  3.319 1.002  1800
#> EM[5,4]        -1.069   1.556 -4.217 -1.890 -1.101 -0.161  1.931 1.003   870
#> dev.o[1,1]      0.975   1.349  0.001  0.107  0.440  1.333  4.490 1.001  2100
#> dev.o[2,1]      0.973   1.351  0.001  0.093  0.451  1.313  4.915 1.003  1600
#> dev.o[3,1]      1.233   1.764  0.001  0.103  0.506  1.602  6.546 1.001  2700
#> dev.o[4,1]      0.963   1.299  0.001  0.097  0.444  1.327  4.540 1.002  2700
#> dev.o[5,1]      1.026   1.391  0.001  0.120  0.491  1.367  5.058 1.002  1600
#> dev.o[6,1]      1.156   1.500  0.001  0.127  0.587  1.632  5.306 1.010   280
#> dev.o[7,1]      0.999   1.264  0.002  0.120  0.503  1.401  4.526 1.004   580
#> dev.o[8,1]      0.005   0.098  0.000  0.000  0.000  0.000  0.007 1.251  1600
#> dev.o[9,1]      0.913   1.298  0.001  0.091  0.424  1.206  4.524 1.003  2600
#> dev.o[10,1]     1.151   1.399  0.001  0.130  0.653  1.632  5.005 1.010   240
#> dev.o[11,1]     0.754   1.122  0.001  0.069  0.328  1.004  3.937 1.003  2700
#> dev.o[1,2]      0.988   1.350  0.001  0.100  0.447  1.323  4.739 1.001  2000
#> dev.o[2,2]      1.272   1.615  0.001  0.164  0.629  1.770  5.770 1.004  1600
#> dev.o[3,2]      1.052   1.464  0.001  0.116  0.504  1.374  5.085 1.001  2400
#> dev.o[4,2]      0.968   1.356  0.001  0.091  0.417  1.293  4.830 1.000  2700
#> dev.o[5,2]      0.973   1.359  0.001  0.099  0.446  1.299  4.947 1.009   270
#> dev.o[6,2]      1.265   1.590  0.002  0.166  0.684  1.780  5.542 1.005   400
#> dev.o[7,2]      0.844   1.148  0.001  0.091  0.402  1.175  4.115 1.001  2400
#> dev.o[8,2]      0.011   0.125  0.000  0.000  0.000  0.000  0.013 1.039  2700
#> dev.o[9,2]      1.004   1.424  0.001  0.109  0.460  1.336  4.821 1.001  2700
#> dev.o[10,2]     1.359   1.699  0.001  0.172  0.735  1.891  6.217 1.003   730
#> dev.o[11,2]     1.058   1.355  0.001  0.127  0.554  1.477  4.802 1.001  2700
#> dev.o[9,3]      1.022   1.474  0.001  0.103  0.462  1.310  5.232 1.004   580
#> dev.o[10,3]     0.828   1.232  0.001  0.077  0.338  1.085  4.153 1.004   720
#> dev.o[11,3]     1.034   1.367  0.001  0.100  0.485  1.428  4.940 1.004   520
#> hat.par[1,1]   26.733   4.493 18.658 23.496 26.645 29.697 35.896 1.009   230
#> hat.par[2,1]    2.931   1.364  0.827  1.913  2.723  3.747  6.095 1.014   150
#> hat.par[3,1]   10.072   1.238  7.190  9.381 10.259 10.998 11.818 1.007   480
#> hat.par[4,1]   22.662   2.475 17.470 20.934 22.868 24.399 27.007 1.001  2700
#> hat.par[5,1]    7.803   2.082  4.067  6.256  7.719  9.236 11.947 1.008   270
#> hat.par[6,1]    3.209   0.977  1.356  2.488  3.231  3.936  5.001 1.007   300
#> hat.par[7,1]    2.443   1.206  0.697  1.544  2.232  3.124  5.378 1.007   310
#> hat.par[8,1]   11.998   0.045 11.997 12.000 12.000 12.000 12.000 1.251  1600
#> hat.par[9,1]   15.560   2.636 10.474 13.725 15.583 17.322 20.752 1.003  1400
#> hat.par[10,1]   3.200   1.280  1.211  2.235  3.033  4.031  6.061 1.023    94
#> hat.par[11,1]   4.437   1.497  1.996  3.346  4.280  5.394  7.738 1.014   170
#> hat.par[1,2]   38.020   5.081 28.703 34.458 37.935 41.292 48.601 1.006   350
#> hat.par[2,2]    4.053   1.663  1.499  2.845  3.827  4.977  7.912 1.004   550
#> hat.par[3,2]    7.880   1.441  4.843  6.917  8.037  8.980 10.170 1.006   380
#> hat.par[4,2]   16.181   2.432 11.152 14.620 16.272 17.834 20.711 1.001  2700
#> hat.par[5,2]    3.207   1.510  0.780  2.076  3.036  4.171  6.460 1.009   400
#> hat.par[6,2]    3.776   0.945  1.748  3.164  3.863  4.481  5.325 1.010   340
#> hat.par[7,2]    2.590   1.218  0.663  1.654  2.444  3.358  5.294 1.007   330
#> hat.par[8,2]   14.995   0.060 14.994 15.000 15.000 15.000 15.000 1.039  2700
#> hat.par[9,2]   16.502   2.821 11.130 14.494 16.538 18.435 21.997 1.007   300
#> hat.par[10,2]   3.181   1.317  0.997  2.185  3.112  4.059  5.971 1.008   400
#> hat.par[11,2]   6.941   1.471  4.165  5.887  6.925  7.979  9.766 1.003  2700
#> hat.par[9,3]   21.909   2.234 17.198 20.471 22.005 23.451 25.919 1.009   340
#> hat.par[10,3]   8.581   1.340  5.656  7.762  8.679  9.549 10.879 1.007   480
#> hat.par[11,3]  11.689   1.538  8.378 10.676 11.782 12.817 14.253 1.021   110
#> m.tau           0.765   0.520  0.023  0.354  0.675  1.091  1.931 1.139    33
#> resdev.o[1]     1.963   1.900  0.047  0.589  1.361  2.785  7.220 1.000  2700
#> resdev.o[2]     2.245   1.936  0.074  0.763  1.718  3.245  7.148 1.010   440
#> resdev.o[3]     2.285   2.287  0.046  0.611  1.529  3.227  8.605 1.007   900
#> resdev.o[4]     1.931   1.874  0.050  0.561  1.364  2.738  6.878 1.001  2700
#> resdev.o[5]     1.999   1.937  0.054  0.625  1.391  2.778  7.104 1.009   260
#> resdev.o[6]     2.421   1.969  0.100  1.014  1.992  3.350  7.262 1.040    73
#> resdev.o[7]     1.843   1.677  0.061  0.618  1.388  2.596  6.137 1.012   290
#> resdev.o[8]     0.016   0.200  0.000  0.000  0.000  0.000  0.030 1.120  2000
#> resdev.o[9]     2.939   2.435  0.225  1.225  2.251  3.971  9.148 1.006   380
#> resdev.o[10]    3.338   2.410  0.283  1.574  2.833  4.552  9.400 1.012   200
#> resdev.o[11]    2.845   2.167  0.185  1.222  2.344  3.894  8.360 1.001  2000
#> tau             0.662   0.469  0.041  0.266  0.605  0.957  1.758 1.186    19
#> totresdev.o    23.826   6.595 12.555 19.154 23.424 28.077 38.261 1.010   450
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $m_tau
#>            mean        sd       2.5%       25%       50%      75%    97.5%
#> m.tau 0.7648348 0.5203487 0.02315624 0.3543708 0.6752405 1.090681 1.930659
#>           Rhat n.eff
#> m.tau 1.139463    33
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
# }