Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al., (2013a) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli, 2019; Spineli et al., 2021). This model offers a global evaluation of the plausibility of the consistency assumption in the network (Dias et al. (2013b)).

run_ume(full, n_iter, n_burnin, n_chains, n_thin)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

n_iter
Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin
Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains
Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin
Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

EM	The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o	The deviance contribution of each trial-arm based on the observed outcome.

hat_par	The fitted outcome at each trial-arm.

tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o	The leverage for the observed outcome at each trial-arm.

sign_dev_o	The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment	A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit	An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var_misspar from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then usenrun_ume (see 'Examples').

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al., (2013a) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al., (2013a) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has been also refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the basic parameters of these trials, that is, effect parameters between non-baseline arms and the baseline arm (Dias et al., 2013b; Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. submitted 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013a;33(5):641--56. doi: 10.1177/0272989X12455847

Dias S, Sutton AJ, Ades AE, Welton NJ. Evidence synthesis for decision making 2: a generalized linear modeling framework for pairwise and network meta-analysis of randomized controlled trials. Med Decis Making 2013b;33(5):607--617. doi: 10.1177/0272989X12458724

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 3045
#> 
#> Initializing model
#> 
#> $EM
#>               mean        sd       2.5%        25%        50%         75%
#> EM[2,1]  0.5975273 0.9705746 -1.3221059  0.1012922  0.5975437  1.09031663
#> EM[3,1]  0.1996594 0.7982822 -1.4117272 -0.2528291  0.1753141  0.63870030
#> EM[4,1]  0.7600531 0.5986451 -0.5255653  0.4134701  0.7927516  1.15293771
#> EM[5,1]  1.9067900 0.8394889  0.1953770  1.3885140  1.9354329  2.43359690
#> EM[4,2] -0.4570747 1.0916751 -2.7304937 -1.0488697 -0.4328043  0.17536525
#> EM[6,2] -1.3530553 1.2046424 -4.0060997 -2.0743010 -1.2660482 -0.57661309
#> EM[4,3]  1.1218848 0.9229205 -0.7542208  0.5933562  1.1169794  1.68318339
#> EM[5,4] -0.9621557 1.4994429 -4.2442487 -1.8113958 -0.8984412  0.06058895
#>             97.5%     Rhat n.eff
#> EM[2,1] 2.5829581 1.018877  2700
#> EM[3,1] 1.9066425 1.021833   150
#> EM[4,1] 1.8697773 1.014378   810
#> EM[5,1] 3.4979892 1.005948   760
#> EM[4,2] 1.7071371 1.029582   190
#> EM[6,2] 0.9039845 1.012308   280
#> EM[4,3] 2.9541452 1.004176   660
#> EM[5,4] 1.7329562 1.020834   100
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  0.977615066 1.37723115 0.0009769839 0.09162188 0.4354780
#> dev.o[2,1]  0.908402492 1.24209481 0.0013131504 0.10779206 0.4460630
#> dev.o[3,1]  1.115475285 1.54163590 0.0015043641 0.10966295 0.5089935
#> dev.o[4,1]  0.920031212 1.29621637 0.0008781533 0.08990253 0.4095407
#> dev.o[5,1]  0.981729683 1.35534124 0.0011859200 0.09242010 0.4552672
#> dev.o[6,1]  1.075274308 1.44014971 0.0006366311 0.11903466 0.5404669
#> dev.o[7,1]  0.995516528 1.35439494 0.0010276178 0.10619086 0.4536406
#> dev.o[8,1]  0.005437734 0.08325524 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.868456807 1.21865972 0.0009117441 0.08855401 0.3982315
#> dev.o[10,1] 1.037814547 1.35606400 0.0011100500 0.12311163 0.5395824
#> dev.o[11,1] 0.788676233 1.06251942 0.0012206852 0.08133757 0.3710500
#> dev.o[1,2]  0.940010126 1.37962212 0.0008597297 0.09379791 0.4119893
#> dev.o[2,2]  1.179386959 1.57329490 0.0013837832 0.13375563 0.5544966
#> dev.o[3,2]  1.012469227 1.42503233 0.0010716606 0.10142772 0.4609799
#> dev.o[4,2]  0.986181931 1.40234422 0.0011912193 0.09764721 0.4395820
#> dev.o[5,2]  1.008465839 1.41003910 0.0011169669 0.10220252 0.4668301
#> dev.o[6,2]  1.173592131 1.47871411 0.0015744742 0.14085668 0.6027151
#> dev.o[7,2]  0.828254571 1.13735839 0.0007109183 0.07651319 0.3674258
#> dev.o[8,2]  0.008746821 0.11295357 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.062706339 1.55310517 0.0011211314 0.10799651 0.4754153
#> dev.o[10,2] 1.308881663 1.60323430 0.0020831124 0.16242849 0.7012913
#> dev.o[11,2] 1.090216654 1.42353733 0.0014869405 0.12539449 0.5460893
#> dev.o[9,3]  1.185613478 1.67851208 0.0018075701 0.13132835 0.5511990
#> dev.o[10,3] 0.865747760 1.27922324 0.0005769798 0.07804697 0.3679097
#> dev.o[11,3] 0.983356866 1.34852518 0.0013665447 0.09722605 0.4470108
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.333983e+00 4.88120840 1.000773  2700
#> dev.o[2,1]  1.197997e+00 4.41128692 1.002306  1200
#> dev.o[3,1]  1.499702e+00 5.40242634 1.003407  1800
#> dev.o[4,1]  1.212590e+00 4.65918210 1.001241  2500
#> dev.o[5,1]  1.344023e+00 5.05862145 1.001786  1500
#> dev.o[6,1]  1.465541e+00 5.12904752 1.002594   930
#> dev.o[7,1]  1.341796e+00 4.82360303 1.005618   430
#> dev.o[8,1]  1.332268e-14 0.01079661 1.184847  2700
#> dev.o[9,1]  1.158568e+00 4.35053904 1.000901  2700
#> dev.o[10,1] 1.387341e+00 4.73532886 1.005080   430
#> dev.o[11,1] 1.086454e+00 3.73960110 1.004055  1100
#> dev.o[1,2]  1.204912e+00 4.80798542 1.000662  2700
#> dev.o[2,2]  1.570665e+00 5.71820332 1.000779  2700
#> dev.o[3,2]  1.297635e+00 4.83444379 1.001317  2300
#> dev.o[4,2]  1.281286e+00 4.87797560 1.001102  2700
#> dev.o[5,2]  1.362988e+00 4.93080780 1.002884  2700
#> dev.o[6,2]  1.666091e+00 5.36923884 1.002185  1200
#> dev.o[7,2]  1.140497e+00 4.08263653 1.003328   700
#> dev.o[8,2]  2.664535e-14 0.02009773 1.039818  2700
#> dev.o[9,2]  1.366791e+00 5.19722784 1.000815  2700
#> dev.o[10,2] 1.867405e+00 5.61159152 1.003560   640
#> dev.o[11,2] 1.513495e+00 4.89250757 1.001287  2400
#> dev.o[9,3]  1.532863e+00 6.11828514 1.002308  2700
#> dev.o[10,3] 1.124233e+00 4.35026667 1.002822  2100
#> dev.o[11,3] 1.314917e+00 4.85694952 1.003197   730
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  26.743275 4.52645351 18.4874085 23.558549 26.604785 29.518190
#> hat.par[2,1]   2.699449 1.37041896  0.6699701  1.632683  2.503279  3.554945
#> hat.par[3,1]  10.025665 1.23503811  7.1603230  9.299945 10.214373 10.974744
#> hat.par[4,1]  22.796699 2.40430444 17.7689483 21.237586 22.931967 24.482523
#> hat.par[5,1]   8.057503 2.05907531  4.2279366  6.577615  7.983094  9.465649
#> hat.par[6,1]   3.295861 0.98311039  1.3872525  2.595878  3.303023  4.021544
#> hat.par[7,1]   2.566533 1.26903177  0.6585362  1.617796  2.369571  3.295515
#> hat.par[8,1]  11.997351 0.03921738 11.9946029 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.464278 2.58905823 10.5571702 13.652258 15.409160 17.228498
#> hat.par[10,1]  3.367401 1.30715020  1.2748308  2.402825  3.217750  4.198142
#> hat.par[11,1]  4.397400 1.50956123  1.7120551  3.298337  4.325477  5.424192
#> hat.par[1,2]  37.998994 4.94853389 28.7131525 34.659662 37.835092 41.273185
#> hat.par[2,2]   4.224649 1.69129447  1.5174758  3.014993  4.007809  5.223130
#> hat.par[3,2]   8.012840 1.38404314  4.9708183  7.107268  8.173437  9.055147
#> hat.par[4,2]  16.022895 2.46723333 11.0892938 14.379080 16.071352 17.745262
#> hat.par[5,2]   2.973766 1.48034594  0.6966118  1.859878  2.758363  3.918404
#> hat.par[6,2]   3.687209 0.96610789  1.6313849  3.048341  3.770497  4.416281
#> hat.par[7,2]   2.417133 1.21898415  0.5586036  1.516411  2.236115  3.142357
#> hat.par[8,2]  14.995730 0.05396585 14.9899545 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.860990 2.81904240 11.4733651 14.920640 16.886430 18.718997
#> hat.par[10,2]  3.156464 1.28576742  0.8827813  2.217186  3.058833  4.016805
#> hat.par[11,2]  7.025340 1.54963786  4.1668456  5.916981  6.994862  8.140765
#> hat.par[9,3]  21.675328 2.37815380 16.5073599 20.140891 21.793622 23.458533
#> hat.par[10,3]  8.494144 1.36894692  5.5402664  7.586550  8.654372  9.502623
#> hat.par[11,3] 11.488491 1.60517636  7.9844717 10.481179 11.616462 12.646728
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.544894 1.000834  2700
#> hat.par[2,1]   5.822338 1.069490    39
#> hat.par[3,1]  11.737521 1.009159   230
#> hat.par[4,1]  27.099052 1.009608   220
#> hat.par[5,1]  12.195766 1.001651  1700
#> hat.par[6,1]   5.126706 1.006580   440
#> hat.par[7,1]   5.576764 1.012111   250
#> hat.par[8,1]  12.000000 1.184847  2700
#> hat.par[9,1]  20.594385 1.001538  1800
#> hat.par[10,1]  6.284590 1.019732   190
#> hat.par[11,1]  7.546964 1.002501  1000
#> hat.par[1,2]  48.142733 1.000716  2700
#> hat.par[2,2]   8.060272 1.003111   750
#> hat.par[3,2]  10.198310 1.004531   490
#> hat.par[4,2]  20.634762 1.008392   250
#> hat.par[5,2]   6.230022 1.008430   250
#> hat.par[6,2]   5.298805 1.009262   420
#> hat.par[7,2]   5.183105 1.005265  1200
#> hat.par[8,2]  15.000000 1.039818  2700
#> hat.par[9,2]  22.426094 1.000630  2700
#> hat.par[10,2]  5.788417 1.020914   120
#> hat.par[11,2]  9.983180 1.003818   750
#> hat.par[9,3]  25.733286 1.003527   650
#> hat.par[10,3] 10.782473 1.003100   750
#> hat.par[11,3] 14.178608 1.003435   670
#> 
#> $leverage_o
#>  [1]  0.974537694  0.682707715  1.115077336  0.913699178  0.981016677
#>  [6]  0.732804395  0.912962362 -0.006431571  0.842191178  0.871845298
#> [11]  0.738170583  0.940010087  1.010748494  1.012393533  0.986101480
#> [16]  1.008185981  0.850193818  0.735498405  0.003184400  0.975426877
#> [21]  0.658475835  0.792422674  0.883302664  0.759425451  0.911205146
#> 
#> $sign_dev_o
#>  [1]  1 -1 -1  1 -1  1  1 -1 -1  1 -1  1  1 -1 -1  1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>         mean        sd      2.5%       25%       50%     75%    97.5%     Rhat
#> tau 0.793174 0.4457836 0.1082022 0.4465373 0.7167924 1.09523 1.775154 1.126478
#>     n.eff
#> tau    27
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 43.59322 20.28516 23.30806
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "9", fit using jags,
#>  3 chains, each with 1000 iterations (first 100 discarded)
#>  n.sims = 2700 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.598   0.971 -1.322  0.101  0.598  1.090  2.583 1.019  2700
#> EM[3,1]         0.200   0.798 -1.412 -0.253  0.175  0.639  1.907 1.022   150
#> EM[4,1]         0.760   0.599 -0.526  0.413  0.793  1.153  1.870 1.014   810
#> EM[5,1]         1.907   0.839  0.195  1.389  1.935  2.434  3.498 1.006   760
#> EM[4,2]        -0.457   1.092 -2.730 -1.049 -0.433  0.175  1.707 1.030   190
#> EM[6,2]        -1.353   1.205 -4.006 -2.074 -1.266 -0.577  0.904 1.012   280
#> EM[4,3]         1.122   0.923 -0.754  0.593  1.117  1.683  2.954 1.004   660
#> EM[5,4]        -0.962   1.499 -4.244 -1.811 -0.898  0.061  1.733 1.021   100
#> dev.o[1,1]      0.978   1.377  0.001  0.092  0.435  1.334  4.881 1.001  2700
#> dev.o[2,1]      0.908   1.242  0.001  0.108  0.446  1.198  4.411 1.002  1200
#> dev.o[3,1]      1.115   1.542  0.002  0.110  0.509  1.500  5.402 1.003  1800
#> dev.o[4,1]      0.920   1.296  0.001  0.090  0.410  1.213  4.659 1.001  2500
#> dev.o[5,1]      0.982   1.355  0.001  0.092  0.455  1.344  5.059 1.002  1500
#> dev.o[6,1]      1.075   1.440  0.001  0.119  0.540  1.466  5.129 1.003   930
#> dev.o[7,1]      0.996   1.354  0.001  0.106  0.454  1.342  4.824 1.006   430
#> dev.o[8,1]      0.005   0.083  0.000  0.000  0.000  0.000  0.011 1.185  2700
#> dev.o[9,1]      0.868   1.219  0.001  0.089  0.398  1.159  4.351 1.001  2700
#> dev.o[10,1]     1.038   1.356  0.001  0.123  0.540  1.387  4.735 1.005   430
#> dev.o[11,1]     0.789   1.063  0.001  0.081  0.371  1.086  3.740 1.004  1100
#> dev.o[1,2]      0.940   1.380  0.001  0.094  0.412  1.205  4.808 1.001  2700
#> dev.o[2,2]      1.179   1.573  0.001  0.134  0.554  1.571  5.718 1.001  2700
#> dev.o[3,2]      1.012   1.425  0.001  0.101  0.461  1.298  4.834 1.001  2300
#> dev.o[4,2]      0.986   1.402  0.001  0.098  0.440  1.281  4.878 1.001  2700
#> dev.o[5,2]      1.008   1.410  0.001  0.102  0.467  1.363  4.931 1.003  2700
#> dev.o[6,2]      1.174   1.479  0.002  0.141  0.603  1.666  5.369 1.002  1200
#> dev.o[7,2]      0.828   1.137  0.001  0.077  0.367  1.140  4.083 1.003   700
#> dev.o[8,2]      0.009   0.113  0.000  0.000  0.000  0.000  0.020 1.040  2700
#> dev.o[9,2]      1.063   1.553  0.001  0.108  0.475  1.367  5.197 1.001  2700
#> dev.o[10,2]     1.309   1.603  0.002  0.162  0.701  1.867  5.612 1.004   640
#> dev.o[11,2]     1.090   1.424  0.001  0.125  0.546  1.513  4.893 1.001  2400
#> dev.o[9,3]      1.186   1.679  0.002  0.131  0.551  1.533  6.118 1.002  2700
#> dev.o[10,3]     0.866   1.279  0.001  0.078  0.368  1.124  4.350 1.003  2100
#> dev.o[11,3]     0.983   1.349  0.001  0.097  0.447  1.315  4.857 1.003   730
#> hat.par[1,1]   26.743   4.526 18.487 23.559 26.605 29.518 36.545 1.001  2700
#> hat.par[2,1]    2.699   1.370  0.670  1.633  2.503  3.555  5.822 1.069    39
#> hat.par[3,1]   10.026   1.235  7.160  9.300 10.214 10.975 11.738 1.009   230
#> hat.par[4,1]   22.797   2.404 17.769 21.238 22.932 24.483 27.099 1.010   220
#> hat.par[5,1]    8.058   2.059  4.228  6.578  7.983  9.466 12.196 1.002  1700
#> hat.par[6,1]    3.296   0.983  1.387  2.596  3.303  4.022  5.127 1.007   440
#> hat.par[7,1]    2.567   1.269  0.659  1.618  2.370  3.296  5.577 1.012   250
#> hat.par[8,1]   11.997   0.039 11.995 12.000 12.000 12.000 12.000 1.185  2700
#> hat.par[9,1]   15.464   2.589 10.557 13.652 15.409 17.228 20.594 1.002  1800
#> hat.par[10,1]   3.367   1.307  1.275  2.403  3.218  4.198  6.285 1.020   190
#> hat.par[11,1]   4.397   1.510  1.712  3.298  4.325  5.424  7.547 1.003  1000
#> hat.par[1,2]   37.999   4.949 28.713 34.660 37.835 41.273 48.143 1.001  2700
#> hat.par[2,2]    4.225   1.691  1.517  3.015  4.008  5.223  8.060 1.003   750
#> hat.par[3,2]    8.013   1.384  4.971  7.107  8.173  9.055 10.198 1.005   490
#> hat.par[4,2]   16.023   2.467 11.089 14.379 16.071 17.745 20.635 1.008   250
#> hat.par[5,2]    2.974   1.480  0.697  1.860  2.758  3.918  6.230 1.008   250
#> hat.par[6,2]    3.687   0.966  1.631  3.048  3.770  4.416  5.299 1.009   420
#> hat.par[7,2]    2.417   1.219  0.559  1.516  2.236  3.142  5.183 1.005  1200
#> hat.par[8,2]   14.996   0.054 14.990 15.000 15.000 15.000 15.000 1.040  2700
#> hat.par[9,2]   16.861   2.819 11.473 14.921 16.886 18.719 22.426 1.001  2700
#> hat.par[10,2]   3.156   1.286  0.883  2.217  3.059  4.017  5.788 1.021   120
#> hat.par[11,2]   7.025   1.550  4.167  5.917  6.995  8.141  9.983 1.004   750
#> hat.par[9,3]   21.675   2.378 16.507 20.141 21.794 23.459 25.733 1.004   650
#> hat.par[10,3]   8.494   1.369  5.540  7.587  8.654  9.503 10.782 1.003   750
#> hat.par[11,3]  11.488   1.605  7.985 10.481 11.616 12.647 14.179 1.003   670
#> m.tau           0.696   0.465  0.059  0.326  0.616  0.987  1.750 1.023   130
#> resdev.o[1]     1.918   1.928  0.047  0.549  1.303  2.645  6.951 1.005  2600
#> resdev.o[2]     2.088   1.908  0.065  0.721  1.557  2.869  7.154 1.003   830
#> resdev.o[3]     2.128   2.136  0.050  0.616  1.456  2.991  7.786 1.003  1200
#> resdev.o[4]     1.906   1.908  0.055  0.551  1.264  2.735  6.865 1.002  1200
#> resdev.o[5]     1.990   1.951  0.056  0.587  1.425  2.714  7.318 1.001  2700
#> resdev.o[6]     2.249   1.875  0.093  0.875  1.816  3.076  6.935 1.010   220
#> resdev.o[7]     1.824   1.784  0.059  0.560  1.310  2.513  6.729 1.008   290
#> resdev.o[8]     0.014   0.166  0.000  0.000  0.000  0.000  0.035 1.021  2700
#> resdev.o[9]     3.117   2.574  0.271  1.263  2.411  4.200  9.807 1.001  2700
#> resdev.o[10]    3.212   2.310  0.265  1.495  2.774  4.359  8.752 1.014   170
#> resdev.o[11]    2.862   2.190  0.247  1.206  2.325  3.969  8.144 1.003  2700
#> tau             0.793   0.446  0.108  0.447  0.717  1.095  1.775 1.126    27
#> totresdev.o    23.308   6.599 12.313 18.628 22.630 27.454 37.866 1.013   160
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $m_tau
#>            mean        sd       2.5%       25%       50%       75%   97.5%
#> m.tau 0.6957161 0.4647121 0.05870822 0.3264985 0.6163363 0.9867809 1.74975
#>           Rhat n.eff
#> m.tau 1.022688   130
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
# }