Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al., (2013a) extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli, 2019; Spineli et al., 2021). This model offers a global evaluation of the plausibility of the consistency assumption in the network (Dias et al. (2013b)).

Usage,

run_ume(full, n_iter, n_burnin, n_chains, n_thin)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

EM	The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o	The deviance contribution of each trial-arm based on the observed outcome.

hat_par	The fitted outcome at each trial-arm.

tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o	The leverage for the observed outcome at each trial-arm.

sign_dev_o	The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment	A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit	An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var_misspar from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then usenrun_ume (see 'Examples').

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al., (2013a) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al., (2013a) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has been also refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the basic parameters of these trials, that is, effect parameters between non-baseline arms and the baseline arm (Dias et al., 2013b; Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. submitted 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013a;33(5):641--56. doi: 10.1177/0272989X12455847

Dias S, Sutton AJ, Ades AE, Welton NJ. Evidence synthesis for decision making 2: a generalized linear modeling framework for pairwise and network meta-analysis of randomized controlled trials. Med Decis Making 2013b;33(5):607--617. doi: 10.1177/0272989X12458724

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472.

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 3045
#> 
#> Initializing model
#> 
#> $EM
#>               mean        sd       2.5%        25%        50%        75%
#> EM[2,1]  0.6034594 0.8175479 -1.0745053  0.1709546  0.6111703  1.0078241
#> EM[3,1]  0.2881048 0.7118144 -1.1727887 -0.1778562  0.3269031  0.8248702
#> EM[4,1]  0.8040565 0.5584079 -0.3517217  0.4962128  0.7517016  1.1657617
#> EM[5,1]  1.8260934 0.7700901  0.3862864  1.2590302  1.7599645  2.3499193
#> EM[4,2] -0.2826230 0.9551659 -2.2830235 -0.8520880 -0.2233916  0.4318422
#> EM[6,2] -1.2043882 1.2178672 -3.7833472 -1.9782534 -1.1488705 -0.3632719
#> EM[4,3]  1.1803193 0.9478788 -0.8208444  0.6453746  1.2015698  1.6905968
#> EM[5,4] -0.6441800 1.2152248 -3.4617733 -1.3175540 -0.5002686  0.1611753
#>             97.5%     Rhat n.eff
#> EM[2,1] 2.4275560 1.009770  2600
#> EM[3,1] 1.6135786 1.039849    65
#> EM[4,1] 1.9094475 1.019692   170
#> EM[5,1] 3.4544426 1.032016    80
#> EM[4,2] 1.4677794 1.049169    45
#> EM[6,2] 0.7964821 1.026464    89
#> EM[4,3] 3.0492247 1.002714   880
#> EM[5,4] 1.4101420 1.177277    18
#> 
#> $dev_o
#>                    mean        sd         2.5%        25%       50%
#> dev.o[1,1]  0.974026306 1.3993733 0.0009318850 0.09610119 0.4429699
#> dev.o[2,1]  0.758981701 1.1203142 0.0005444060 0.06854645 0.3212130
#> dev.o[3,1]  0.906091258 1.3478354 0.0009761593 0.08987745 0.4056838
#> dev.o[4,1]  1.067337699 1.5587814 0.0009263392 0.10358910 0.4758936
#> dev.o[5,1]  1.157691005 1.5763824 0.0010634467 0.11787620 0.5364119
#> dev.o[6,1]  1.068341093 1.4192968 0.0012285470 0.12156679 0.5172376
#> dev.o[7,1]  1.001817772 1.2660430 0.0011767045 0.12503678 0.5000889
#> dev.o[8,1]  0.009327816 0.1011663 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.816221678 1.1496696 0.0008403551 0.07933696 0.3662028
#> dev.o[10,1] 1.015247489 1.3531871 0.0009712057 0.10920940 0.4974121
#> dev.o[11,1] 0.877567585 1.1492728 0.0007186650 0.08349884 0.4455349
#> dev.o[1,2]  0.988032880 1.4135779 0.0007723593 0.09610197 0.4348864
#> dev.o[2,2]  1.046889137 1.4459688 0.0008903441 0.10371169 0.4798151
#> dev.o[3,2]  1.045566397 1.4561322 0.0014104491 0.11328006 0.5034799
#> dev.o[4,2]  1.099815734 1.5221152 0.0010740860 0.11394027 0.5111551
#> dev.o[5,2]  1.172291027 1.5873998 0.0012272524 0.11513910 0.5351602
#> dev.o[6,2]  1.228676906 1.5285661 0.0011785990 0.15403158 0.6392973
#> dev.o[7,2]  0.777864248 1.0757373 0.0006622328 0.08475736 0.3662723
#> dev.o[8,2]  0.011649640 0.1520274 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.469215847 2.0168731 0.0014360147 0.14878432 0.6561640
#> dev.o[10,2] 1.409146667 1.7552270 0.0015011541 0.16159505 0.7374695
#> dev.o[11,2] 1.056976046 1.3975378 0.0012640630 0.12421286 0.5100876
#> dev.o[9,3]  1.461344895 1.9563601 0.0020241198 0.13147010 0.6272159
#> dev.o[10,3] 0.956752630 1.2856937 0.0010180361 0.09308751 0.4382405
#> dev.o[11,3] 0.975724340 1.4265436 0.0013325320 0.08892859 0.4235329
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.283443e+00 4.73110169 1.003303   700
#> dev.o[2,1]  1.011282e+00 3.89359088 1.019152   110
#> dev.o[3,1]  1.124339e+00 4.77844807 1.008080   260
#> dev.o[4,1]  1.403180e+00 5.26743708 1.004002   610
#> dev.o[5,1]  1.548449e+00 5.77512381 1.006395   430
#> dev.o[6,1]  1.449283e+00 5.15396781 1.002968   830
#> dev.o[7,1]  1.413940e+00 4.50406883 1.005511   400
#> dev.o[8,1]  4.529710e-14 0.02067527 1.145895   560
#> dev.o[9,1]  1.079861e+00 4.03418020 1.000656  2700
#> dev.o[10,1] 1.387534e+00 5.11001741 1.000697  2700
#> dev.o[11,1] 1.233806e+00 3.95005079 1.008241   270
#> dev.o[1,2]  1.329183e+00 4.96106400 1.001598  2700
#> dev.o[2,2]  1.427994e+00 5.21274722 1.002677   900
#> dev.o[3,2]  1.385766e+00 5.18609803 1.000604  2700
#> dev.o[4,2]  1.417875e+00 5.52618606 1.006471   360
#> dev.o[5,2]  1.601257e+00 5.80816196 1.008871   270
#> dev.o[6,2]  1.805548e+00 5.44524596 1.000791  2700
#> dev.o[7,2]  1.047970e+00 3.93764295 1.002144  1200
#> dev.o[8,2]  6.661338e-14 0.03034825 1.168843  1900
#> dev.o[9,2]  1.947413e+00 7.18168515 1.031981    68
#> dev.o[10,2] 2.013820e+00 6.43408553 1.010785   220
#> dev.o[11,2] 1.416198e+00 4.92380105 1.001868  2700
#> dev.o[9,3]  2.048262e+00 7.02422707 1.019063   120
#> dev.o[10,3] 1.294796e+00 4.76828646 1.004678   470
#> dev.o[11,3] 1.309913e+00 5.03415408 1.002775  1500
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  26.875793 4.53719503 18.6072646 23.785040 26.649456 29.879543
#> hat.par[2,1]   2.713649 1.23841841  0.8195944  1.813294  2.524402  3.429760
#> hat.par[3,1]   9.851837 1.19868261  7.0679465  9.143452 10.029803 10.761548
#> hat.par[4,1]  22.478397 2.57865958 17.2569627 20.815105 22.605189 24.307355
#> hat.par[5,1]   7.812849 2.19475818  3.6935419  6.262422  7.800291  9.359006
#> hat.par[6,1]   3.268611 0.96082230  1.3860498  2.601287  3.283855  3.978807
#> hat.par[7,1]   2.541870 1.27899411  0.7156245  1.564446  2.312652  3.288116
#> hat.par[8,1]  11.995441 0.04889795 11.9896668 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.212260 2.53641891 10.3439689 13.472602 15.126691 16.891666
#> hat.par[10,1]  3.193153 1.15667836  1.1929896  2.344194  3.092290  3.943321
#> hat.par[11,1]  4.756133 1.54100405  1.9970666  3.684653  4.658177  5.825611
#> hat.par[1,2]  38.220299 5.08227952 28.6020006 34.838092 38.065673 41.612730
#> hat.par[2,2]   4.297822 1.62891488  1.6305740  3.131968  4.136231  5.258660
#> hat.par[3,2]   8.181902 1.36196839  5.1347620  7.342421  8.390715  9.215356
#> hat.par[4,2]  16.453904 2.52878473 11.2030007 14.748107 16.505560 18.286748
#> hat.par[5,2]   3.165110 1.66954401  0.6891395  1.904865  2.931683  4.177919
#> hat.par[6,2]   3.776638 0.92735497  1.8397826  3.126321  3.865585  4.505099
#> hat.par[7,2]   2.389983 1.18545572  0.6095106  1.466787  2.206055  3.135642
#> hat.par[8,2]  14.994361 0.07125527 14.9848335 15.000000 15.000000 15.000000
#> hat.par[9,2]  17.506978 3.11059740 11.7475711 15.266306 17.448727 19.630205
#> hat.par[10,2]  3.258473 1.30641151  1.0158187  2.289527  3.145038  4.141361
#> hat.par[11,2]  7.008522 1.50879825  4.1184340  5.989168  7.012955  8.081015
#> hat.par[9,3]  21.401748 2.60045311 15.9814802 19.609241 21.708488 23.317157
#> hat.par[10,3]  8.355509 1.41139348  5.3795169  7.408508  8.486589  9.406768
#> hat.par[11,3] 11.201747 1.68546482  7.7969964 10.108450 11.276294 12.450321
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.322981 1.017579   130
#> hat.par[2,1]   5.610888 1.012144   300
#> hat.par[3,1]  11.634797 1.033201    67
#> hat.par[4,1]  27.202867 1.060936    38
#> hat.par[5,1]  12.113339 1.016750   200
#> hat.par[6,1]   5.020698 1.008842   320
#> hat.par[7,1]   5.484569 1.017940   120
#> hat.par[8,1]  12.000000 1.145895   560
#> hat.par[9,1]  20.308163 1.008416   270
#> hat.par[10,1]  5.619325 1.002451  1000
#> hat.par[11,1]  7.861735 1.063193    39
#> hat.par[1,2]  48.123095 1.011126   200
#> hat.par[2,2]   7.956103 1.007537   400
#> hat.par[3,2]  10.264426 1.053418    57
#> hat.par[4,2]  21.013640 1.068422    34
#> hat.par[5,2]   6.957304 1.043182    67
#> hat.par[6,2]   5.316570 1.006435   640
#> hat.par[7,2]   5.078864 1.006810   410
#> hat.par[8,2]  15.000000 1.168843  1900
#> hat.par[9,2]  23.506071 1.060922    41
#> hat.par[10,2]  6.083673 1.012829   170
#> hat.par[11,2]  9.799422 1.002079  1300
#> hat.par[9,3]  25.780861 1.066839    66
#> hat.par[10,3] 10.756012 1.018593   150
#> hat.par[11,3] 14.176749 1.042675    53
#> 
#> $leverage_o
#>  [1] 0.973307736 0.524776017 0.893411635 1.026216082 1.150094865 0.699559716
#>  [7] 0.909131624 0.004497312 0.810723203 0.740237446 0.700680348 0.986195362
#> [13] 0.909721405 1.029995392 1.067915776 1.161561421 0.812457608 0.696292663
#> [19] 0.008913431 1.201924580 0.650331981 0.748942789 1.027433717 0.786916012
#> [25] 0.963702667
#> 
#> $sign_dev_o
#>  [1]  1 -1  1  1  1  1  1 -1 -1  1 -1 -1  1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>          mean     sd       2.5%       25%       50%       75%    97.5%     Rhat
#> tau 0.6530688 0.4007 0.00246017 0.3366081 0.6874658 0.9086464 1.537732 1.380215
#>     n.eff
#> tau    15
#> 
#> $model_assessment
#>        DIC       pD     dev
#> 1 44.83754 20.48494 24.3526
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "13", fit using jags,
#>  3 chains, each with 1000 iterations (first 100 discarded)
#>  n.sims = 2700 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.603   0.818 -1.075  0.171  0.611  1.008  2.428 1.010  2600
#> EM[3,1]         0.288   0.712 -1.173 -0.178  0.327  0.825  1.614 1.040    65
#> EM[4,1]         0.804   0.558 -0.352  0.496  0.752  1.166  1.909 1.020   170
#> EM[5,1]         1.826   0.770  0.386  1.259  1.760  2.350  3.454 1.032    80
#> EM[4,2]        -0.283   0.955 -2.283 -0.852 -0.223  0.432  1.468 1.049    45
#> EM[6,2]        -1.204   1.218 -3.783 -1.978 -1.149 -0.363  0.796 1.026    89
#> EM[4,3]         1.180   0.948 -0.821  0.645  1.202  1.691  3.049 1.003   880
#> EM[5,4]        -0.644   1.215 -3.462 -1.318 -0.500  0.161  1.410 1.177    18
#> dev.o[1,1]      0.974   1.399  0.001  0.096  0.443  1.283  4.731 1.003   700
#> dev.o[2,1]      0.759   1.120  0.001  0.069  0.321  1.011  3.894 1.019   110
#> dev.o[3,1]      0.906   1.348  0.001  0.090  0.406  1.124  4.778 1.008   260
#> dev.o[4,1]      1.067   1.559  0.001  0.104  0.476  1.403  5.267 1.004   610
#> dev.o[5,1]      1.158   1.576  0.001  0.118  0.536  1.548  5.775 1.006   430
#> dev.o[6,1]      1.068   1.419  0.001  0.122  0.517  1.449  5.154 1.003   830
#> dev.o[7,1]      1.002   1.266  0.001  0.125  0.500  1.414  4.504 1.006   400
#> dev.o[8,1]      0.009   0.101  0.000  0.000  0.000  0.000  0.021 1.146   560
#> dev.o[9,1]      0.816   1.150  0.001  0.079  0.366  1.080  4.034 1.001  2700
#> dev.o[10,1]     1.015   1.353  0.001  0.109  0.497  1.388  5.110 1.001  2700
#> dev.o[11,1]     0.878   1.149  0.001  0.083  0.446  1.234  3.950 1.008   270
#> dev.o[1,2]      0.988   1.414  0.001  0.096  0.435  1.329  4.961 1.002  2700
#> dev.o[2,2]      1.047   1.446  0.001  0.104  0.480  1.428  5.213 1.003   900
#> dev.o[3,2]      1.046   1.456  0.001  0.113  0.503  1.386  5.186 1.001  2700
#> dev.o[4,2]      1.100   1.522  0.001  0.114  0.511  1.418  5.526 1.006   360
#> dev.o[5,2]      1.172   1.587  0.001  0.115  0.535  1.601  5.808 1.009   270
#> dev.o[6,2]      1.229   1.529  0.001  0.154  0.639  1.806  5.445 1.001  2700
#> dev.o[7,2]      0.778   1.076  0.001  0.085  0.366  1.048  3.938 1.002  1200
#> dev.o[8,2]      0.012   0.152  0.000  0.000  0.000  0.000  0.030 1.169  1900
#> dev.o[9,2]      1.469   2.017  0.001  0.149  0.656  1.947  7.182 1.032    68
#> dev.o[10,2]     1.409   1.755  0.002  0.162  0.737  2.014  6.434 1.011   220
#> dev.o[11,2]     1.057   1.398  0.001  0.124  0.510  1.416  4.924 1.002  2700
#> dev.o[9,3]      1.461   1.956  0.002  0.131  0.627  2.048  7.024 1.019   120
#> dev.o[10,3]     0.957   1.286  0.001  0.093  0.438  1.295  4.768 1.005   470
#> dev.o[11,3]     0.976   1.427  0.001  0.089  0.424  1.310  5.034 1.003  1500
#> hat.par[1,1]   26.876   4.537 18.607 23.785 26.649 29.880 36.323 1.018   130
#> hat.par[2,1]    2.714   1.238  0.820  1.813  2.524  3.430  5.611 1.012   300
#> hat.par[3,1]    9.852   1.199  7.068  9.143 10.030 10.762 11.635 1.033    67
#> hat.par[4,1]   22.478   2.579 17.257 20.815 22.605 24.307 27.203 1.061    38
#> hat.par[5,1]    7.813   2.195  3.694  6.262  7.800  9.359 12.113 1.017   200
#> hat.par[6,1]    3.269   0.961  1.386  2.601  3.284  3.979  5.021 1.009   320
#> hat.par[7,1]    2.542   1.279  0.716  1.564  2.313  3.288  5.485 1.018   120
#> hat.par[8,1]   11.995   0.049 11.990 12.000 12.000 12.000 12.000 1.146   560
#> hat.par[9,1]   15.212   2.536 10.344 13.473 15.127 16.892 20.308 1.008   270
#> hat.par[10,1]   3.193   1.157  1.193  2.344  3.092  3.943  5.619 1.002  1000
#> hat.par[11,1]   4.756   1.541  1.997  3.685  4.658  5.826  7.862 1.063    39
#> hat.par[1,2]   38.220   5.082 28.602 34.838 38.066 41.613 48.123 1.011   200
#> hat.par[2,2]    4.298   1.629  1.631  3.132  4.136  5.259  7.956 1.008   400
#> hat.par[3,2]    8.182   1.362  5.135  7.342  8.391  9.215 10.264 1.053    57
#> hat.par[4,2]   16.454   2.529 11.203 14.748 16.506 18.287 21.014 1.068    34
#> hat.par[5,2]    3.165   1.670  0.689  1.905  2.932  4.178  6.957 1.043    67
#> hat.par[6,2]    3.777   0.927  1.840  3.126  3.866  4.505  5.317 1.006   640
#> hat.par[7,2]    2.390   1.185  0.610  1.467  2.206  3.136  5.079 1.007   410
#> hat.par[8,2]   14.994   0.071 14.985 15.000 15.000 15.000 15.000 1.169  1900
#> hat.par[9,2]   17.507   3.111 11.748 15.266 17.449 19.630 23.506 1.061    41
#> hat.par[10,2]   3.258   1.306  1.016  2.290  3.145  4.141  6.084 1.013   170
#> hat.par[11,2]   7.009   1.509  4.118  5.989  7.013  8.081  9.799 1.002  1300
#> hat.par[9,3]   21.402   2.600 15.982 19.609 21.708 23.317 25.781 1.067    66
#> hat.par[10,3]   8.356   1.411  5.380  7.409  8.487  9.407 10.756 1.019   150
#> hat.par[11,3]  11.202   1.685  7.797 10.108 11.276 12.450 14.177 1.043    53
#> m.tau           0.704   0.463  0.031  0.357  0.631  0.992  1.780 1.049   120
#> resdev.o[1]     1.962   1.991  0.043  0.564  1.385  2.727  7.325 1.011   240
#> resdev.o[2]     1.806   1.815  0.051  0.526  1.249  2.501  6.734 1.030    73
#> resdev.o[3]     1.952   1.996  0.052  0.592  1.326  2.664  7.525 1.003  1200
#> resdev.o[4]     2.167   2.173  0.045  0.614  1.510  3.052  7.757 1.012   170
#> resdev.o[5]     2.330   2.205  0.054  0.653  1.618  3.397  8.018 1.010   210
#> resdev.o[6]     2.297   1.873  0.104  0.979  1.852  3.122  7.345 1.004   690
#> resdev.o[7]     1.780   1.689  0.054  0.554  1.279  2.452  6.452 1.014   150
#> resdev.o[8]     0.021   0.220  0.000  0.000  0.000  0.000  0.059 1.072  1300
#> resdev.o[9]     3.747   3.141  0.260  1.412  2.752  5.090 11.678 1.067    35
#> resdev.o[10]    3.381   2.444  0.299  1.549  2.799  4.694  9.237 1.015   140
#> resdev.o[11]    2.910   2.218  0.222  1.319  2.368  4.002  8.352 1.003  2700
#> tau             0.653   0.401  0.002  0.337  0.687  0.909  1.538 1.380    15
#> totresdev.o    24.353   7.079 12.641 19.203 23.621 28.871 39.773 1.019   130
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $m_tau
#>            mean        sd       2.5%       25%       50%       75%    97.5%
#> m.tau 0.7040411 0.4630815 0.03114374 0.3566375 0.6308681 0.9924293 1.779528
#>           Rhat n.eff
#> m.tau 1.048803   120
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
# }