Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al. (2013) that has been refined (Spineli, 2021) and extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al. 2021; Spineli, 2019). This model offers a global evaluation of the plausibility of the consistency assumption in the network.

Usage

run_ume(full, n_iter, n_burnin, n_chains, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM

The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o

The deviance contribution of each trial-arm based on the observed outcome.

hat_par

The fitted outcome at each trial-arm.

tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o

The leverage for the observed outcome at each trial-arm.

sign_dev_o

The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment

A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit

An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then use run_ume (see 'Examples').

The run_ume function also returns the arguments data, model, measure, assumption, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm or multi-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_ume currently considers the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has also been refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the underlying trial-specific effect size of comparisons with the baseline arm of the multi-arm trial (Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console. The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013;33(5):641–56. doi: 10.1177/0272989X12455847

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457–72. doi: 10.1214/ss/1177011136

Spineli LM. A Revised Framework to Evaluate the Consistency Assumption Globally in a Network of Interventions. Med Decis Making 2021. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958–75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> JAGS generates initial values for the parameters.
#> Running the model ...
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 1291
#> 
#> Initializing model
#> 
#> ... Updating the model until convergence
#> $EM
#>                mean        sd       2.5%        25%         50%        75%
#> EM[2,1]  0.49642480 0.6905131 -0.9567945  0.1786713  0.45862595  0.7681942
#> EM[3,1] -0.02367709 0.6120869 -1.1167940 -0.4266807 -0.07119421  0.3469384
#> EM[4,1]  0.74073006 0.4749441 -0.4955669  0.5672850  0.78358244  0.9643396
#> EM[5,1]  1.47803115 0.8324419  0.3287901  0.7598876  1.39756862  2.1507544
#> EM[4,2] -0.49716691 0.8366196 -2.2793140 -0.8713242 -0.61012707 -0.1085123
#> EM[6,2] -1.23526632 1.0633944 -3.7190981 -1.7734224 -1.10207371 -0.6490275
#> EM[4,3]  1.15855939 0.9894310 -0.8740707  0.5749576  1.15485960  1.7664568
#> EM[5,4] -1.09284436 1.2380836 -3.7247265 -1.6657975 -1.33274012 -0.3161956
#>            97.5%     Rhat n.eff
#> EM[2,1] 2.137820 1.106463   200
#> EM[3,1] 1.296626 1.204116    16
#> EM[4,1] 1.579046 1.109636   150
#> EM[5,1] 3.167902 1.775213     6
#> EM[4,2] 1.353754 1.136974    47
#> EM[6,2] 0.756402 1.122374    41
#> EM[4,3] 3.038140 1.028043   110
#> EM[5,4] 1.361460 1.285247    14
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  0.935117781 1.28850298 0.0008564073 0.09700999 0.4431592
#> dev.o[2,1]  1.036621072 1.38265438 0.0014493326 0.12007700 0.5327132
#> dev.o[3,1]  0.930239767 1.20301698 0.0011577014 0.11126026 0.4782858
#> dev.o[4,1]  0.930246758 1.32883578 0.0010810394 0.09136053 0.4185878
#> dev.o[5,1]  1.042182331 1.52904408 0.0015491846 0.11230286 0.4740713
#> dev.o[6,1]  1.090447371 1.43079879 0.0020057633 0.13267308 0.5413477
#> dev.o[7,1]  0.968893639 1.28036468 0.0015491884 0.10170487 0.4448654
#> dev.o[8,1]  0.006290771 0.07668972 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.976257934 1.34611731 0.0009298877 0.10470269 0.4685238
#> dev.o[10,1] 0.932062031 1.23085220 0.0010961397 0.10225219 0.4620421
#> dev.o[11,1] 0.978393835 1.21427773 0.0010617699 0.12808495 0.5206264
#> dev.o[1,2]  0.989668349 1.41648950 0.0006260779 0.09413614 0.4407477
#> dev.o[2,2]  1.400275015 1.74949157 0.0014988809 0.17828619 0.7648413
#> dev.o[3,2]  0.994631067 1.42801167 0.0011497626 0.09353783 0.4359848
#> dev.o[4,2]  0.957328581 1.25875829 0.0010201433 0.09689188 0.4716471
#> dev.o[5,2]  1.029739491 1.60272404 0.0007622272 0.09229198 0.4335656
#> dev.o[6,2]  1.214356864 1.42917168 0.0018847781 0.17282622 0.7275810
#> dev.o[7,2]  0.761704981 1.11077943 0.0011434278 0.07269318 0.3431934
#> dev.o[8,2]  0.010362713 0.11503776 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  0.988898849 1.45537459 0.0011451667 0.10443924 0.4538463
#> dev.o[10,2] 1.974983800 2.13377151 0.0024669520 0.35475554 1.2703733
#> dev.o[11,2] 1.007060667 1.32812291 0.0011810753 0.10710766 0.4825511
#> dev.o[9,3]  1.183041984 1.47854729 0.0017208060 0.15283648 0.6383246
#> dev.o[10,3] 1.291332561 1.63085507 0.0015249684 0.15443253 0.6727814
#> dev.o[11,3] 0.908817154 1.15888516 0.0010348276 0.10939039 0.4579764
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.239979e+00 4.52999932 1.004237   540
#> dev.o[2,1]  1.430706e+00 4.92716872 1.013045   160
#> dev.o[3,1]  1.245374e+00 4.42792709 1.030933    75
#> dev.o[4,1]  1.270511e+00 4.42279043 1.002533   990
#> dev.o[5,1]  1.351520e+00 5.34519707 1.008881   240
#> dev.o[6,1]  1.488175e+00 5.14841485 1.002457  1000
#> dev.o[7,1]  1.338314e+00 4.61934741 1.003948   580
#> dev.o[8,1]  9.592327e-14 0.02189270 1.193889  1600
#> dev.o[9,1]  1.285428e+00 4.93488992 1.005791   380
#> dev.o[10,1] 1.271270e+00 4.28434382 1.058144    41
#> dev.o[11,1] 1.390307e+00 4.18514899 1.149897    18
#> dev.o[1,2]  1.331880e+00 4.88840119 1.002623   940
#> dev.o[2,2]  1.972224e+00 6.39416726 1.020359   110
#> dev.o[3,2]  1.340487e+00 4.84551985 1.021969   110
#> dev.o[4,2]  1.313479e+00 4.48735466 1.001055  3000
#> dev.o[5,2]  1.273664e+00 5.72533080 1.023830    89
#> dev.o[6,2]  1.750424e+00 5.04324123 1.006645   490
#> dev.o[7,2]  9.731499e-01 3.92598277 1.001491  2000
#> dev.o[8,2]  3.530509e-13 0.05402013 1.120512  1400
#> dev.o[9,2]  1.250386e+00 5.08766718 1.011035   190
#> dev.o[10,2] 2.844125e+00 7.64206992 1.088581    30
#> dev.o[11,2] 1.410709e+00 4.78315173 1.038671    59
#> dev.o[9,3]  1.621025e+00 5.37320583 1.002352  3000
#> dev.o[10,3] 1.809426e+00 5.77945121 1.113709    22
#> dev.o[11,3] 1.291898e+00 4.13286019 1.009541   310
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  27.365683 4.47946855 18.8599456 24.214509 27.214251 30.366827
#> hat.par[2,1]   3.066296 1.36748722  0.8943778  2.055825  2.908327  3.901797
#> hat.par[3,1]  10.071285 1.14873364  7.3121140  9.403677 10.320028 10.946771
#> hat.par[4,1]  23.153083 2.37997147 18.1420865 21.609393 23.323352 24.841959
#> hat.par[5,1]   7.945519 2.10794220  4.2163678  6.424249  7.833740  9.358734
#> hat.par[6,1]   3.219977 0.93960830  1.3912850  2.557921  3.233287  3.870667
#> hat.par[7,1]   2.451741 1.18033489  0.6810645  1.582444  2.285698  3.138829
#> hat.par[8,1]  11.996914 0.03658284 11.9890586 12.000000 12.000000 12.000000
#> hat.par[9,1]  16.211212 2.52209212 11.0426175 14.538971 16.286277 17.885399
#> hat.par[10,1]  3.466742 1.28913620  1.3697320  2.537859  3.351008  4.268252
#> hat.par[11,1]  5.113332 1.48947493  2.3666952  4.009097  5.085279  6.149442
#> hat.par[1,2]  37.365987 4.98830853 28.0133007 33.954410 37.285377 40.689708
#> hat.par[2,2]   3.831022 1.56905571  1.3709651  2.681728  3.618227  4.737592
#> hat.par[3,2]   7.907729 1.40414611  4.8419278  7.012769  8.103431  8.937729
#> hat.par[4,2]  15.791691 2.44432306 10.9355737 14.083700 15.866585 17.586582
#> hat.par[5,2]   3.118186 1.49310448  0.6005184  2.022093  2.965191  4.078892
#> hat.par[6,2]   3.808221 0.90326721  1.8224236  3.234641  3.930111  4.487295
#> hat.par[7,2]   2.464259 1.17430482  0.6362858  1.589575  2.305671  3.167304
#> hat.par[8,2]  14.994927 0.05508356 14.9730142 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.542572 2.78790834 11.1834860 14.547256 16.606268 18.429939
#> hat.par[10,2]  3.670438 1.40166362  1.1145078  2.634970  3.643964  4.616213
#> hat.par[11,2]  6.899478 1.38904084  4.1550258  5.932756  6.929307  7.860611
#> hat.par[9,3]  21.307579 2.14011481 16.9376721 19.940511 21.320023 22.714639
#> hat.par[10,3]  7.897752 1.50250815  4.8983961  6.845119  7.924962  9.041010
#> hat.par[11,3] 10.888061 1.69700491  7.6153829  9.666119 10.911821 12.151708
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.380676 1.025891    94
#> hat.par[2,1]   6.179124 1.071143    36
#> hat.par[3,1]  11.655385 1.139809    25
#> hat.par[4,1]  27.291619 1.113250    24
#> hat.par[5,1]  12.280872 1.019459   130
#> hat.par[6,1]   5.049689 1.011901   220
#> hat.par[7,1]   5.334995 1.003541   960
#> hat.par[8,1]  12.000000 1.193889  1600
#> hat.par[9,1]  21.207408 1.122000    23
#> hat.par[10,1]  6.464539 1.074240    37
#> hat.par[11,1]  7.999582 1.327241    10
#> hat.par[1,2]  47.549733 1.003274   740
#> hat.par[2,2]   7.434218 1.063951    36
#> hat.par[3,2]  10.153205 1.070615    39
#> hat.par[4,2]  20.362227 1.087169    28
#> hat.par[5,2]   6.291657 1.091428    39
#> hat.par[6,2]   5.262013 1.029838   110
#> hat.par[7,2]   5.195535 1.001960  3000
#> hat.par[8,2]  15.000000 1.120512  1400
#> hat.par[9,2]  22.009871 1.042923    59
#> hat.par[10,2]  6.464374 1.154894    18
#> hat.par[11,2]  9.568994 1.053479    62
#> hat.par[9,3]  25.438338 1.034963   360
#> hat.par[10,3] 10.543249 1.365643     9
#> hat.par[11,3] 13.924038 1.429172     8
#> 
#> $leverage_o
#>  [1]  0.928944852  0.551142727  0.927131173  0.926598657  1.041540667
#>  [6]  0.672333884  0.833602449 -0.003400990  0.798067831  0.815273533
#> [11]  0.606023465  0.974319720  0.998289931  0.990777458  0.950697096
#> [16]  1.024198943  0.762263947  0.647999225  0.006642804  0.954214583
#> [21]  0.708079931  0.628492620  0.698776949  0.817234044  0.905200966
#> 
#> $sign_dev_o
#>  [1] -1 -1 -1 -1  1  1  1 -1 -1  1 -1  1  1  1  1 -1 -1 -1 -1 -1 -1  1  1  1  1
#> 
#> $tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.45724699  0.45556736  0.02123328  0.12416876  0.21564509  0.71594362 
#>       97.5%        Rhat       n.eff 
#>  1.58101425  1.21457671 15.00000000 
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.75183153  0.51745987  0.05055095  0.33755831  0.66086965  1.07577866 
#>       97.5%        Rhat       n.eff 
#>  1.95052346  1.07199460 39.00000000 
#> 
#> $model_assessment
#>       DIC       pD      dev
#> 1 43.7034 19.16445 24.53896
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "15", fit using jags,
#>  3 chains, each with 1000 iterations (first 0 discarded)
#>  n.sims = 3000 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.496   0.691 -0.957  0.179  0.459  0.768  2.138 1.106   200
#> EM[3,1]        -0.024   0.612 -1.117 -0.427 -0.071  0.347  1.297 1.204    16
#> EM[4,1]         0.741   0.475 -0.496  0.567  0.784  0.964  1.579 1.110   150
#> EM[5,1]         1.478   0.832  0.329  0.760  1.398  2.151  3.168 1.775     6
#> EM[4,2]        -0.497   0.837 -2.279 -0.871 -0.610 -0.109  1.354 1.137    47
#> EM[6,2]        -1.235   1.063 -3.719 -1.773 -1.102 -0.649  0.756 1.122    41
#> EM[4,3]         1.159   0.989 -0.874  0.575  1.155  1.766  3.038 1.028   110
#> EM[5,4]        -1.093   1.238 -3.725 -1.666 -1.333 -0.316  1.361 1.285    14
#> dev.o[1,1]      0.935   1.289  0.001  0.097  0.443  1.240  4.530 1.004   540
#> dev.o[2,1]      1.037   1.383  0.001  0.120  0.533  1.431  4.927 1.013   160
#> dev.o[3,1]      0.930   1.203  0.001  0.111  0.478  1.245  4.428 1.031    75
#> dev.o[4,1]      0.930   1.329  0.001  0.091  0.419  1.271  4.423 1.003   990
#> dev.o[5,1]      1.042   1.529  0.002  0.112  0.474  1.352  5.345 1.009   240
#> dev.o[6,1]      1.090   1.431  0.002  0.133  0.541  1.488  5.148 1.002  1000
#> dev.o[7,1]      0.969   1.280  0.002  0.102  0.445  1.338  4.619 1.004   580
#> dev.o[8,1]      0.006   0.077  0.000  0.000  0.000  0.000  0.022 1.194  1600
#> dev.o[9,1]      0.976   1.346  0.001  0.105  0.469  1.285  4.935 1.006   380
#> dev.o[10,1]     0.932   1.231  0.001  0.102  0.462  1.271  4.284 1.058    41
#> dev.o[11,1]     0.978   1.214  0.001  0.128  0.521  1.390  4.185 1.150    18
#> dev.o[1,2]      0.990   1.416  0.001  0.094  0.441  1.332  4.888 1.003   940
#> dev.o[2,2]      1.400   1.749  0.001  0.178  0.765  1.972  6.394 1.020   110
#> dev.o[3,2]      0.995   1.428  0.001  0.094  0.436  1.340  4.846 1.022   110
#> dev.o[4,2]      0.957   1.259  0.001  0.097  0.472  1.313  4.487 1.001  3000
#> dev.o[5,2]      1.030   1.603  0.001  0.092  0.434  1.274  5.725 1.024    89
#> dev.o[6,2]      1.214   1.429  0.002  0.173  0.728  1.750  5.043 1.007   490
#> dev.o[7,2]      0.762   1.111  0.001  0.073  0.343  0.973  3.926 1.001  2000
#> dev.o[8,2]      0.010   0.115  0.000  0.000  0.000  0.000  0.054 1.121  1400
#> dev.o[9,2]      0.989   1.455  0.001  0.104  0.454  1.250  5.088 1.011   190
#> dev.o[10,2]     1.975   2.134  0.002  0.355  1.270  2.844  7.642 1.089    30
#> dev.o[11,2]     1.007   1.328  0.001  0.107  0.483  1.411  4.783 1.039    59
#> dev.o[9,3]      1.183   1.479  0.002  0.153  0.638  1.621  5.373 1.002  3000
#> dev.o[10,3]     1.291   1.631  0.002  0.154  0.673  1.809  5.779 1.114    22
#> dev.o[11,3]     0.909   1.159  0.001  0.109  0.458  1.292  4.133 1.010   310
#> hat.par[1,1]   27.366   4.479 18.860 24.215 27.214 30.367 36.381 1.026    94
#> hat.par[2,1]    3.066   1.367  0.894  2.056  2.908  3.902  6.179 1.071    36
#> hat.par[3,1]   10.071   1.149  7.312  9.404 10.320 10.947 11.655 1.140    25
#> hat.par[4,1]   23.153   2.380 18.142 21.609 23.323 24.842 27.292 1.113    24
#> hat.par[5,1]    7.946   2.108  4.216  6.424  7.834  9.359 12.281 1.019   130
#> hat.par[6,1]    3.220   0.940  1.391  2.558  3.233  3.871  5.050 1.012   220
#> hat.par[7,1]    2.452   1.180  0.681  1.582  2.286  3.139  5.335 1.004   960
#> hat.par[8,1]   11.997   0.037 11.989 12.000 12.000 12.000 12.000 1.194  1600
#> hat.par[9,1]   16.211   2.522 11.043 14.539 16.286 17.885 21.207 1.122    23
#> hat.par[10,1]   3.467   1.289  1.370  2.538  3.351  4.268  6.465 1.074    37
#> hat.par[11,1]   5.113   1.489  2.367  4.009  5.085  6.149  8.000 1.327    10
#> hat.par[1,2]   37.366   4.988 28.013 33.954 37.285 40.690 47.550 1.003   740
#> hat.par[2,2]    3.831   1.569  1.371  2.682  3.618  4.738  7.434 1.064    36
#> hat.par[3,2]    7.908   1.404  4.842  7.013  8.103  8.938 10.153 1.071    39
#> hat.par[4,2]   15.792   2.444 10.936 14.084 15.867 17.587 20.362 1.087    28
#> hat.par[5,2]    3.118   1.493  0.601  2.022  2.965  4.079  6.292 1.091    39
#> hat.par[6,2]    3.808   0.903  1.822  3.235  3.930  4.487  5.262 1.030   110
#> hat.par[7,2]    2.464   1.174  0.636  1.590  2.306  3.167  5.196 1.002  3000
#> hat.par[8,2]   14.995   0.055 14.973 15.000 15.000 15.000 15.000 1.121  1400
#> hat.par[9,2]   16.543   2.788 11.183 14.547 16.606 18.430 22.010 1.043    59
#> hat.par[10,2]   3.670   1.402  1.115  2.635  3.644  4.616  6.464 1.155    18
#> hat.par[11,2]   6.899   1.389  4.155  5.933  6.929  7.861  9.569 1.053    62
#> hat.par[9,3]   21.308   2.140 16.938 19.941 21.320 22.715 25.438 1.035   360
#> hat.par[10,3]   7.898   1.503  4.898  6.845  7.925  9.041 10.543 1.366     9
#> hat.par[11,3]  10.888   1.697  7.615  9.666 10.912 12.152 13.924 1.429     8
#> m.tau           0.752   0.517  0.051  0.338  0.661  1.076  1.951 1.072    39
#> resdev.o[1]     1.925   1.906  0.052  0.578  1.348  2.653  7.168 1.005   430
#> resdev.o[2]     2.437   1.956  0.125  1.041  2.010  3.282  7.604 1.133    24
#> resdev.o[3]     1.925   1.872  0.052  0.549  1.352  2.742  6.653 1.088    29
#> resdev.o[4]     1.888   1.831  0.063  0.626  1.347  2.581  6.627 1.009  3000
#> resdev.o[5]     2.072   2.285  0.042  0.544  1.320  2.720  8.751 1.054    44
#> resdev.o[6]     2.305   1.761  0.167  1.226  1.902  2.924  6.684 1.075    84
#> resdev.o[7]     1.731   1.723  0.052  0.551  1.206  2.340  6.335 1.010   220
#> resdev.o[8]     0.017   0.186  0.000  0.000  0.000  0.000  0.085 1.152  1400
#> resdev.o[9]     3.148   2.433  0.322  1.433  2.466  4.199  9.840 1.031  1100
#> resdev.o[10]    4.198   2.634  0.483  2.231  3.754  5.628 10.248 1.162    17
#> resdev.o[11]    2.894   1.880  0.336  1.668  2.536  3.684  7.775 1.100    46
#> tau             0.457   0.456  0.021  0.124  0.216  0.716  1.581 1.215    15
#> totresdev.o    24.539   6.149 13.724 20.436 24.048 28.121 38.024 1.008   740
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.75183153  0.51745987  0.05055095  0.33755831  0.66086965  1.07577866 
#>       97.5%        Rhat       n.eff 
#>  1.95052346  1.07199460 39.00000000 
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
#> attr(,"class")
#> [1] "run_ume"
# }