Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al. (2013) that has been refined (Spineli, 2021) and extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al. 2021; Spineli, 2019). This model offers a global evaluation of the plausibility of the consistency assumption in the network.

Usage

run_ume(full, n_iter, n_burnin, n_chains, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM

The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o

The deviance contribution of each trial-arm based on the observed outcome.

hat_par

The fitted outcome at each trial-arm.

tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o

The leverage for the observed outcome at each trial-arm.

sign_dev_o

The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment

A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit

An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in

ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then use run_ume (see 'Examples').

The run_ume function also returns the arguments data, model, measure, assumption, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm or multi-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_ume currently considers the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has also been refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the underlying trial-specific effect size of comparisons with the baseline arm of the multi-arm trial (Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console. The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013;33(5):641--56. doi: 10.1177/0272989X12455847

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Spineli LM. A Revised Framework to Evaluate the Consistency Assumption Globally in a Network of Interventions. Med Decis Making 2021. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> JAGS generates initial values for the parameters.
#> Running the model ...
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 1291
#> 
#> Initializing model
#> 
#> ... Updating the model until convergence
#> $EM
#>               mean        sd       2.5%        25%        50%         75%
#> EM[2,1]  0.5681292 0.8912131 -1.3203372  0.1176278  0.5741752  1.05236808
#> EM[3,1]  0.1715804 0.7260854 -1.2770450 -0.2670411  0.1369518  0.61971457
#> EM[4,1]  0.7528986 0.5718437 -0.4818844  0.4086407  0.7784579  1.15694168
#> EM[5,1]  1.8711790 0.8303822  0.2082119  1.3467923  1.9049764  2.41737624
#> EM[4,2] -0.4463600 0.9955184 -2.5102008 -1.0099103 -0.4473231  0.16807986
#> EM[6,2] -1.2752044 1.1559992 -3.7018401 -1.9010846 -1.2324002 -0.60932553
#> EM[4,3]  1.1348160 1.0534356 -1.0638717  0.5272045  1.1181986  1.72285976
#> EM[5,4] -0.9007546 1.4615159 -3.9141208 -1.8003520 -0.8424791  0.04003111
#>            97.5%     Rhat n.eff
#> EM[2,1] 2.449376 1.016460   230
#> EM[3,1] 1.704471 1.000798  3000
#> EM[4,1] 1.765843 1.025656    89
#> EM[5,1] 3.400341 1.008515   670
#> EM[4,2] 1.499241 1.002813   880
#> EM[6,2] 1.012703 1.001111  3000
#> EM[4,3] 3.324289 1.011875   320
#> EM[5,4] 1.855078 1.049140    51
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  0.970963571 1.36148728 0.0010111172 0.09386610 0.4460961
#> dev.o[2,1]  0.917329137 1.24604622 0.0013529294 0.09021059 0.4356120
#> dev.o[3,1]  1.103704146 1.52652981 0.0009674224 0.11482700 0.5237140
#> dev.o[4,1]  0.948581438 1.36111445 0.0006594315 0.08673430 0.4237909
#> dev.o[5,1]  1.029419304 1.45898368 0.0008789751 0.10191027 0.4550973
#> dev.o[6,1]  1.081482610 1.40614158 0.0012629258 0.12109131 0.5394558
#> dev.o[7,1]  0.973741432 1.30107663 0.0010777738 0.10930357 0.4571097
#> dev.o[8,1]  0.005346873 0.05799602 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.915892187 1.39149799 0.0007771566 0.09017262 0.3870513
#> dev.o[10,1] 0.963967309 1.20983803 0.0010663618 0.11133591 0.5019317
#> dev.o[11,1] 0.930890943 1.43253385 0.0007036155 0.08196349 0.4111237
#> dev.o[1,2]  0.987033691 1.46831189 0.0007417604 0.09932607 0.4340799
#> dev.o[2,2]  1.141183759 1.49078481 0.0013080503 0.12572520 0.5608204
#> dev.o[3,2]  1.036591110 1.47268945 0.0008486308 0.10128498 0.4614623
#> dev.o[4,2]  1.068996277 1.53193502 0.0014546151 0.09995544 0.4739631
#> dev.o[5,2]  1.029377530 1.44751665 0.0010388246 0.09963814 0.4474852
#> dev.o[6,2]  1.131849491 1.46403866 0.0009463409 0.14018186 0.5721071
#> dev.o[7,2]  0.760219913 1.05690852 0.0009615915 0.07745369 0.3365414
#> dev.o[8,2]  0.008929242 0.08914316 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.046793470 1.51913824 0.0010046834 0.10694817 0.4814010
#> dev.o[10,2] 1.391634448 1.73034570 0.0013100903 0.18094357 0.7289814
#> dev.o[11,2] 1.129136652 1.49950939 0.0008867486 0.11456361 0.5257653
#> dev.o[9,3]  1.240317506 1.73336995 0.0014739687 0.11253051 0.5622008
#> dev.o[10,3] 0.856577405 1.21580612 0.0010979446 0.08837889 0.3812082
#> dev.o[11,3] 1.081476645 1.44399550 0.0015586247 0.12736416 0.5266357
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.243112e+00 5.04451324 1.008496   720
#> dev.o[2,1]  1.267057e+00 4.23240503 1.000788  3000
#> dev.o[3,1]  1.417584e+00 5.72052298 1.005279   420
#> dev.o[4,1]  1.247564e+00 4.88889763 1.004556   500
#> dev.o[5,1]  1.362659e+00 5.07489728 1.002212  1200
#> dev.o[6,1]  1.513443e+00 5.03583520 1.002605   950
#> dev.o[7,1]  1.344400e+00 4.57147440 1.000841  3000
#> dev.o[8,1]  2.131628e-14 0.01197497 1.009160  3000
#> dev.o[9,1]  1.200898e+00 4.65062103 1.001288  2500
#> dev.o[10,1] 1.363772e+00 4.34039116 1.001153  3000
#> dev.o[11,1] 1.203937e+00 4.86214600 1.002340  1300
#> dev.o[1,2]  1.296360e+00 5.23276194 1.000690  3000
#> dev.o[2,2]  1.565150e+00 5.45205614 1.002187  1300
#> dev.o[3,2]  1.398969e+00 5.14756462 1.001756  1600
#> dev.o[4,2]  1.391335e+00 5.26498302 1.001240  2700
#> dev.o[5,2]  1.364684e+00 5.11306911 1.004512   780
#> dev.o[6,2]  1.594826e+00 5.31235936 1.000821  3000
#> dev.o[7,2]  1.017476e+00 3.77265758 1.001774  3000
#> dev.o[8,2]  3.996803e-14 0.02572654 1.094847  2100
#> dev.o[9,2]  1.384587e+00 5.16795390 1.000561  3000
#> dev.o[10,2] 1.970629e+00 6.05530800 1.000994  3000
#> dev.o[11,2] 1.600020e+00 5.12083972 1.001571  1900
#> dev.o[9,3]  1.660940e+00 5.98406220 1.012072   190
#> dev.o[10,3] 1.117273e+00 4.33290156 1.003818   600
#> dev.o[11,3] 1.456546e+00 5.16437178 1.001546  1900
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  27.096999 4.55850427 18.5583387 23.916743 26.873529 30.036949
#> hat.par[2,1]   2.765541 1.34879274  0.6375523  1.759074  2.582536  3.594605
#> hat.par[3,1]  10.011834 1.23685570  7.0885398  9.274469 10.247520 10.975565
#> hat.par[4,1]  23.005496 2.41076327 17.9136832 21.440126 23.102573 24.719263
#> hat.par[5,1]   7.857640 2.08740103  4.0470057  6.386321  7.807589  9.275931
#> hat.par[6,1]   3.270669 0.97303099  1.4064387  2.552636  3.280861  4.001999
#> hat.par[7,1]   2.490043 1.21350140  0.6929613  1.586406  2.311003  3.184244
#> hat.par[8,1]  11.997361 0.02844125 11.9940140 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.630130 2.61969205 10.5834493 13.872279 15.600276 17.404821
#> hat.par[10,1]  3.363480 1.25464639  1.3309647  2.422807  3.221254  4.141476
#> hat.par[11,1]  4.653607 1.62605731  1.9485598  3.510408  4.495649  5.664815
#> hat.par[1,2]  37.860969 5.03290337 27.8663080 34.394716 37.894111 41.193323
#> hat.par[2,2]   4.196059 1.65043438  1.5557809  2.986912  3.977811  5.178795
#> hat.par[3,2]   8.063744 1.38766291  4.9521273  7.193159  8.206144  9.105972
#> hat.par[4,2]  16.181946 2.53852625 11.1822900 14.482939 16.201885 17.956516
#> hat.par[5,2]   3.031828 1.51187283  0.6683662  1.922380  2.792466  3.964113
#> hat.par[6,2]   3.683265 0.94485714  1.6524792  3.031520  3.771094  4.393343
#> hat.par[7,2]   2.483608 1.16236154  0.6548759  1.622135  2.341155  3.186961
#> hat.par[8,2]  14.995601 0.04355646 14.9871422 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.693908 2.84060697 11.2641435 14.753277 16.697490 18.639685
#> hat.par[10,2]  3.220690 1.31717975  0.9450829  2.245069  3.123782  4.103729
#> hat.par[11,2]  6.966097 1.55324640  4.0422347  5.844454  7.006330  8.067815
#> hat.par[9,3]  21.574689 2.39057933 16.5671824 20.045627 21.803618 23.253723
#> hat.par[10,3]  8.439279 1.34688779  5.5589419  7.585440  8.540986  9.433256
#> hat.par[11,3] 11.477103 1.70199398  7.7186714 10.406705 11.693299 12.718340
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.785851 1.023498    92
#> hat.par[2,1]   5.672698 1.002734   920
#> hat.par[3,1]  11.771486 1.030281    89
#> hat.par[4,1]  27.429474 1.004553   760
#> hat.par[5,1]  12.099977 1.003862  3000
#> hat.par[6,1]   5.061812 1.011858   200
#> hat.par[7,1]   5.358772 1.005199  1700
#> hat.par[8,1]  12.000000 1.009160  3000
#> hat.par[9,1]  20.824272 1.007680   650
#> hat.par[10,1]  6.211958 1.000843  3000
#> hat.par[11,1]  8.115920 1.010288   490
#> hat.par[1,2]  47.802737 1.011347   190
#> hat.par[2,2]   7.952980 1.001774  1600
#> hat.par[3,2]  10.220351 1.009042   240
#> hat.par[4,2]  21.042772 1.001505  2000
#> hat.par[5,2]   6.457335 1.001532  3000
#> hat.par[6,2]   5.299355 1.001920  1400
#> hat.par[7,2]   5.099728 1.003425   690
#> hat.par[8,2]  15.000000 1.094847  2100
#> hat.par[9,2]  22.157005 1.011682   180
#> hat.par[10,2]  5.955262 1.005615   540
#> hat.par[11,2]  9.953841 1.002109  1400
#> hat.par[9,3]  25.873896 1.019505   140
#> hat.par[10,3] 10.751522 1.008300   580
#> hat.par[11,3] 14.229942 1.002993  1600
#> 
#> $leverage_o
#>  [1]  0.970527112  0.650903920  1.103619855  0.948576767  1.025028874
#>  [6]  0.714721185  0.857645223 -0.006581762  0.867552196  0.795867508
#> [11]  0.797514917  0.986299246  0.959282822  1.034711926  1.063899715
#> [16]  1.028969766  0.812252574  0.637358474  0.003704423  0.990083556
#> [21]  0.673802956  0.794514909  0.892164888  0.726783188  1.012716919
#> 
#> $sign_dev_o
#>  [1] -1 -1 -1 -1  1  1  1 -1 -1  1 -1  1  1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>         mean           sd         2.5%          25%          50%          75% 
#>   0.70391650   0.43155849   0.05070578   0.38544236   0.67683734   0.96698359 
#>        97.5%         Rhat        n.eff 
#>   1.67475714   1.01288096 160.00000000 
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.80404912  0.53249428  0.05171884  0.38760970  0.71620686  1.13182895 
#>       97.5%        Rhat       n.eff 
#>  2.02388401  1.06189177 61.00000000 
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 44.09336 20.34192 23.75144
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "9", fit using jags,
#>  3 chains, each with 1000 iterations (first 0 discarded)
#>  n.sims = 3000 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.568   0.891 -1.320  0.118  0.574  1.052  2.449 1.016   230
#> EM[3,1]         0.172   0.726 -1.277 -0.267  0.137  0.620  1.704 1.001  3000
#> EM[4,1]         0.753   0.572 -0.482  0.409  0.778  1.157  1.766 1.026    89
#> EM[5,1]         1.871   0.830  0.208  1.347  1.905  2.417  3.400 1.009   670
#> EM[4,2]        -0.446   0.996 -2.510 -1.010 -0.447  0.168  1.499 1.003   880
#> EM[6,2]        -1.275   1.156 -3.702 -1.901 -1.232 -0.609  1.013 1.001  3000
#> EM[4,3]         1.135   1.053 -1.064  0.527  1.118  1.723  3.324 1.012   320
#> EM[5,4]        -0.901   1.462 -3.914 -1.800 -0.842  0.040  1.855 1.049    51
#> dev.o[1,1]      0.971   1.361  0.001  0.094  0.446  1.243  5.045 1.008   720
#> dev.o[2,1]      0.917   1.246  0.001  0.090  0.436  1.267  4.232 1.001  3000
#> dev.o[3,1]      1.104   1.527  0.001  0.115  0.524  1.418  5.721 1.005   420
#> dev.o[4,1]      0.949   1.361  0.001  0.087  0.424  1.248  4.889 1.005   500
#> dev.o[5,1]      1.029   1.459  0.001  0.102  0.455  1.363  5.075 1.002  1200
#> dev.o[6,1]      1.081   1.406  0.001  0.121  0.539  1.513  5.036 1.003   950
#> dev.o[7,1]      0.974   1.301  0.001  0.109  0.457  1.344  4.571 1.001  3000
#> dev.o[8,1]      0.005   0.058  0.000  0.000  0.000  0.000  0.012 1.009  3000
#> dev.o[9,1]      0.916   1.391  0.001  0.090  0.387  1.201  4.651 1.001  2500
#> dev.o[10,1]     0.964   1.210  0.001  0.111  0.502  1.364  4.340 1.001  3000
#> dev.o[11,1]     0.931   1.433  0.001  0.082  0.411  1.204  4.862 1.002  1300
#> dev.o[1,2]      0.987   1.468  0.001  0.099  0.434  1.296  5.233 1.001  3000
#> dev.o[2,2]      1.141   1.491  0.001  0.126  0.561  1.565  5.452 1.002  1300
#> dev.o[3,2]      1.037   1.473  0.001  0.101  0.461  1.399  5.148 1.002  1600
#> dev.o[4,2]      1.069   1.532  0.001  0.100  0.474  1.391  5.265 1.001  2700
#> dev.o[5,2]      1.029   1.448  0.001  0.100  0.447  1.365  5.113 1.005   780
#> dev.o[6,2]      1.132   1.464  0.001  0.140  0.572  1.595  5.312 1.001  3000
#> dev.o[7,2]      0.760   1.057  0.001  0.077  0.337  1.017  3.773 1.002  3000
#> dev.o[8,2]      0.009   0.089  0.000  0.000  0.000  0.000  0.026 1.095  2100
#> dev.o[9,2]      1.047   1.519  0.001  0.107  0.481  1.385  5.168 1.001  3000
#> dev.o[10,2]     1.392   1.730  0.001  0.181  0.729  1.971  6.055 1.001  3000
#> dev.o[11,2]     1.129   1.500  0.001  0.115  0.526  1.600  5.121 1.002  1900
#> dev.o[9,3]      1.240   1.733  0.001  0.113  0.562  1.661  5.984 1.012   190
#> dev.o[10,3]     0.857   1.216  0.001  0.088  0.381  1.117  4.333 1.004   600
#> dev.o[11,3]     1.081   1.444  0.002  0.127  0.527  1.457  5.164 1.002  1900
#> hat.par[1,1]   27.097   4.559 18.558 23.917 26.874 30.037 36.786 1.023    92
#> hat.par[2,1]    2.766   1.349  0.638  1.759  2.583  3.595  5.673 1.003   920
#> hat.par[3,1]   10.012   1.237  7.089  9.274 10.248 10.976 11.771 1.030    89
#> hat.par[4,1]   23.005   2.411 17.914 21.440 23.103 24.719 27.429 1.005   760
#> hat.par[5,1]    7.858   2.087  4.047  6.386  7.808  9.276 12.100 1.004  3000
#> hat.par[6,1]    3.271   0.973  1.406  2.553  3.281  4.002  5.062 1.012   200
#> hat.par[7,1]    2.490   1.214  0.693  1.586  2.311  3.184  5.359 1.005  1700
#> hat.par[8,1]   11.997   0.028 11.994 12.000 12.000 12.000 12.000 1.009  3000
#> hat.par[9,1]   15.630   2.620 10.583 13.872 15.600 17.405 20.824 1.008   650
#> hat.par[10,1]   3.363   1.255  1.331  2.423  3.221  4.141  6.212 1.001  3000
#> hat.par[11,1]   4.654   1.626  1.949  3.510  4.496  5.665  8.116 1.010   490
#> hat.par[1,2]   37.861   5.033 27.866 34.395 37.894 41.193 47.803 1.011   190
#> hat.par[2,2]    4.196   1.650  1.556  2.987  3.978  5.179  7.953 1.002  1600
#> hat.par[3,2]    8.064   1.388  4.952  7.193  8.206  9.106 10.220 1.009   240
#> hat.par[4,2]   16.182   2.539 11.182 14.483 16.202 17.957 21.043 1.002  2000
#> hat.par[5,2]    3.032   1.512  0.668  1.922  2.792  3.964  6.457 1.002  3000
#> hat.par[6,2]    3.683   0.945  1.652  3.032  3.771  4.393  5.299 1.002  1400
#> hat.par[7,2]    2.484   1.162  0.655  1.622  2.341  3.187  5.100 1.003   690
#> hat.par[8,2]   14.996   0.044 14.987 15.000 15.000 15.000 15.000 1.095  2100
#> hat.par[9,2]   16.694   2.841 11.264 14.753 16.697 18.640 22.157 1.012   180
#> hat.par[10,2]   3.221   1.317  0.945  2.245  3.124  4.104  5.955 1.006   540
#> hat.par[11,2]   6.966   1.553  4.042  5.844  7.006  8.068  9.954 1.002  1400
#> hat.par[9,3]   21.575   2.391 16.567 20.046 21.804 23.254 25.874 1.020   140
#> hat.par[10,3]   8.439   1.347  5.559  7.585  8.541  9.433 10.752 1.008   580
#> hat.par[11,3]  11.477   1.702  7.719 10.407 11.693 12.718 14.230 1.003  1600
#> m.tau           0.804   0.532  0.052  0.388  0.716  1.132  2.024 1.062    61
#> resdev.o[1]     1.958   2.013  0.048  0.576  1.313  2.656  7.481 1.001  3000
#> resdev.o[2]     2.059   1.873  0.051  0.673  1.592  2.877  7.008 1.001  2600
#> resdev.o[3]     2.140   2.115  0.060  0.636  1.510  2.920  8.152 1.009   290
#> resdev.o[4]     2.018   2.105  0.052  0.562  1.377  2.731  7.869 1.007   290
#> resdev.o[5]     2.059   2.069  0.047  0.558  1.399  2.866  7.429 1.005   400
#> resdev.o[6]     2.213   1.847  0.101  0.876  1.817  2.992  7.024 1.005   410
#> resdev.o[7]     1.734   1.666  0.053  0.545  1.241  2.466  6.268 1.001  3000
#> resdev.o[8]     0.014   0.132  0.000  0.000  0.000  0.000  0.048 1.042  3000
#> resdev.o[9]     3.203   2.628  0.213  1.276  2.499  4.476  9.810 1.006   360
#> resdev.o[10]    3.212   2.290  0.298  1.533  2.721  4.382  9.193 1.002  1100
#> resdev.o[11]    3.142   2.447  0.246  1.376  2.568  4.189  9.446 1.006   460
#> tau             0.704   0.432  0.051  0.385  0.677  0.967  1.675 1.013   160
#> totresdev.o    23.751   6.812 12.479 18.847 23.020 28.029 38.904 1.003   710
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.80404912  0.53249428  0.05171884  0.38760970  0.71620686  1.13182895 
#>       97.5%        Rhat       n.eff 
#>  2.02388401  1.06189177 61.00000000 
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
#> attr(,"class")
#> [1] "run_ume"
# }