Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al. (2013) that has been refined (Spineli, 2021) and extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al. 2021; Spineli, 2019). This model offers a global evaluation of the plausibility of the consistency assumption in the network.

Usage

run_ume(full, n_iter, n_burnin, n_chains, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM

The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o

The deviance contribution of each trial-arm based on the observed outcome.

hat_par

The fitted outcome at each trial-arm.

tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o

The leverage for the observed outcome at each trial-arm.

sign_dev_o

The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment

A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit

An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then use run_ume (see 'Examples').

The run_ume function also returns the arguments data, model, measure, assumption, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm or multi-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_ume currently considers the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has also been refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the underlying trial-specific effect size of comparisons with the baseline arm of the multi-arm trial (Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console. The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013;33(5):641–56. doi: 10.1177/0272989X12455847

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457–72. doi: 10.1214/ss/1177011136

Spineli LM. A Revised Framework to Evaluate the Consistency Assumption Globally in a Network of Interventions. Med Decis Making 2021. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958–75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> JAGS generates initial values for the parameters.
#> Running the model ...
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 1291
#> 
#> Initializing model
#> 
#> ... Updating the model until convergence
#> $EM
#>               mean        sd       2.5%        25%        50%         75%
#> EM[2,1]  0.5947222 0.8430483 -1.2005171  0.1575535  0.6004243  1.05460962
#> EM[3,1]  0.1591737 0.6939274 -1.2870231 -0.2418464  0.1722985  0.61293914
#> EM[4,1]  0.7434191 0.5222778 -0.3670261  0.4626603  0.7599240  1.05245281
#> EM[5,1]  1.8308135 0.7776631  0.4469792  1.3403337  1.7779057  2.25997082
#> EM[4,2] -0.4524609 0.9170062 -2.3390133 -0.9695892 -0.4729023  0.09366015
#> EM[6,2] -1.4627077 1.1348636 -3.6967621 -2.1593898 -1.4037489 -0.76798556
#> EM[4,3]  1.1116045 0.9349771 -0.9611786  0.5989397  1.1876997  1.63650496
#> EM[5,4] -1.0862288 1.4764196 -3.8890788 -2.0634799 -1.0827566 -0.16439937
#>             97.5%     Rhat n.eff
#> EM[2,1] 2.3449609 1.025794  3000
#> EM[3,1] 1.5272344 1.018473   660
#> EM[4,1] 1.7357237 1.037343   290
#> EM[5,1] 3.5211586 1.028526   100
#> EM[4,2] 1.4544861 1.019178   700
#> EM[6,2] 0.7292316 1.006155   860
#> EM[4,3] 2.9574815 1.003563   880
#> EM[5,4] 1.8527278 1.032855   680
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  0.985983787 1.40490586 0.0010046933 0.11020435 0.4673353
#> dev.o[2,1]  0.919791470 1.26141256 0.0011401185 0.09635134 0.4458006
#> dev.o[3,1]  1.243734559 1.71873099 0.0012670737 0.11642395 0.5507678
#> dev.o[4,1]  0.902589174 1.25667548 0.0011319355 0.08760769 0.4048065
#> dev.o[5,1]  1.006698811 1.44346175 0.0008604505 0.10938977 0.4621310
#> dev.o[6,1]  1.070421380 1.44093882 0.0011437083 0.10125210 0.4994910
#> dev.o[7,1]  0.921456282 1.16639970 0.0011177070 0.11742322 0.4725905
#> dev.o[8,1]  0.005482831 0.07736295 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  0.822162593 1.15774276 0.0011721758 0.07967904 0.3703845
#> dev.o[10,1] 0.995261000 1.29888238 0.0013469038 0.11526749 0.4967196
#> dev.o[11,1] 0.820181867 1.19515108 0.0005195380 0.07581729 0.3518669
#> dev.o[1,2]  0.971189005 1.36998635 0.0010034590 0.09472016 0.4497191
#> dev.o[2,2]  1.177886105 1.60599406 0.0009268151 0.11555453 0.5408863
#> dev.o[3,2]  1.004605032 1.45101083 0.0010593475 0.09977836 0.4502062
#> dev.o[4,2]  0.980890916 1.39242521 0.0009036894 0.09561195 0.4340786
#> dev.o[5,2]  1.090904310 1.47893409 0.0013522680 0.10934646 0.4971036
#> dev.o[6,2]  1.163874002 1.44612473 0.0011913232 0.12765310 0.5996290
#> dev.o[7,2]  0.735312682 1.03914465 0.0006481353 0.07482311 0.3313387
#> dev.o[8,2]  0.010169872 0.12656844 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  1.012799013 1.39004952 0.0016683229 0.11845876 0.4733460
#> dev.o[10,2] 1.446373021 1.73177053 0.0026215358 0.19383249 0.7968175
#> dev.o[11,2] 1.244043607 1.59050585 0.0015822987 0.15195041 0.6509902
#> dev.o[9,3]  1.249472738 1.59887141 0.0015970499 0.14759694 0.6381660
#> dev.o[10,3] 0.879844676 1.25155523 0.0009062425 0.08849844 0.4069015
#> dev.o[11,3] 1.014732848 1.50559039 0.0008608670 0.09901235 0.4282766
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.285832e+00 4.98045379 1.003114  1100
#> dev.o[2,1]  1.214172e+00 4.59699544 1.001551  1900
#> dev.o[3,1]  1.684661e+00 6.12757302 1.004924   450
#> dev.o[4,1]  1.224398e+00 4.48964382 1.000894  3000
#> dev.o[5,1]  1.365406e+00 4.74224106 1.003645  3000
#> dev.o[6,1]  1.478914e+00 5.05822200 1.004120   550
#> dev.o[7,1]  1.259554e+00 4.34814218 1.003017   800
#> dev.o[8,1]  4.263256e-14 0.01407587 1.076300  3000
#> dev.o[9,1]  1.125181e+00 4.37895483 1.000747  3000
#> dev.o[10,1] 1.333162e+00 4.80449417 1.016040   180
#> dev.o[11,1] 1.053907e+00 4.37750960 1.009376   230
#> dev.o[1,2]  1.279290e+00 4.82290628 1.002172  1200
#> dev.o[2,2]  1.639259e+00 5.90247930 1.000714  3000
#> dev.o[3,2]  1.282902e+00 5.10187280 1.005282   440
#> dev.o[4,2]  1.308308e+00 5.05717334 1.006832   480
#> dev.o[5,2]  1.508699e+00 5.37195910 1.005915  1000
#> dev.o[6,2]  1.703769e+00 5.14886605 1.003091   770
#> dev.o[7,2]  9.735489e-01 3.78266333 1.000971  3000
#> dev.o[8,2]  1.365574e-13 0.03017531 1.084359  3000
#> dev.o[9,2]  1.357082e+00 5.12471205 1.002709   910
#> dev.o[10,2] 2.144792e+00 6.15383554 1.009327   240
#> dev.o[11,2] 1.720726e+00 5.66070651 1.009214   290
#> dev.o[9,3]  1.733396e+00 5.70532388 1.015835   160
#> dev.o[10,3] 1.148878e+00 4.34282061 1.003916   590
#> dev.o[11,3] 1.289410e+00 5.29948842 1.001818  3000
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  27.084147 4.55952549 18.5294994 23.846531 26.923828 30.170731
#> hat.par[2,1]   2.881072 1.32939458  0.7791472  1.912607  2.713528  3.649450
#> hat.par[3,1]  10.157241 1.23850288  7.1039028  9.499212 10.395638 11.116004
#> hat.par[4,1]  22.965355 2.36809611 17.9821955 21.437424 23.073463 24.637432
#> hat.par[5,1]   8.228573 2.08322009  4.5143358  6.716998  8.168844  9.625220
#> hat.par[6,1]   3.268687 0.96380353  1.4078866  2.558003  3.325323  3.960825
#> hat.par[7,1]   2.545305 1.22578640  0.7568332  1.622818  2.330505  3.292534
#> hat.par[8,1]  11.997320 0.03724392 11.9929641 12.000000 12.000000 12.000000
#> hat.par[9,1]  15.579530 2.49663250 10.8964371 13.862498 15.540176 17.299153
#> hat.par[10,1]  3.295808 1.22988979  1.2781704  2.409504  3.169625  4.039878
#> hat.par[11,1]  4.664264 1.50088021  1.9333652  3.663153  4.583564  5.604451
#> hat.par[1,2]  38.074015 5.03276389 28.6408994 34.564537 38.003453 41.411313
#> hat.par[2,2]   4.168838 1.63817871  1.4829176  2.923775  3.975995  5.212129
#> hat.par[3,2]   7.851603 1.41159676  4.7555186  6.936488  7.972689  8.905724
#> hat.par[4,2]  16.088300 2.45719348 11.0565474 14.491114 16.195306 17.847879
#> hat.par[5,2]   2.837039 1.48162642  0.6385541  1.715516  2.602804  3.738770
#> hat.par[6,2]   3.753185 0.91279125  1.8180245  3.125764  3.819324  4.468994
#> hat.par[7,2]   2.452636 1.14937047  0.6386105  1.587643  2.304123  3.149827
#> hat.par[8,2]  14.995045 0.06024151 14.9849199 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.899899 2.74053658 11.5128012 14.965564 16.981483 18.814750
#> hat.par[10,2]  3.282876 1.31510854  1.0042255  2.319033  3.199081  4.212683
#> hat.par[11,2]  6.821531 1.56366274  3.8836049  5.740580  6.789153  7.888193
#> hat.par[9,3]  21.304858 2.24564732 16.7556198 19.830633 21.406683 22.853312
#> hat.par[10,3]  8.397038 1.35561028  5.4860084  7.530189  8.497482  9.372634
#> hat.par[11,3] 11.442663 1.63033725  7.9861655 10.404622 11.569351 12.587880
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.347957 1.000973  3000
#> hat.par[2,1]   5.991596 1.016786   170
#> hat.par[3,1]  11.774320 1.007522  1000
#> hat.par[4,1]  27.196344 1.001338  2400
#> hat.par[5,1]  12.338856 1.000509  3000
#> hat.par[6,1]   5.027113 1.010496   210
#> hat.par[7,1]   5.337914 1.011301   190
#> hat.par[8,1]  12.000000 1.076300  3000
#> hat.par[9,1]  20.567990 1.007296   330
#> hat.par[10,1]  6.050334 1.024371    86
#> hat.par[11,1]  7.912445 1.015552   750
#> hat.par[1,2]  48.226456 1.002233  1200
#> hat.par[2,2]   7.823680 1.002778  1000
#> hat.par[3,2]  10.196624 1.005455  2300
#> hat.par[4,2]  20.519165 1.003512   670
#> hat.par[5,2]   6.290378 1.009785   320
#> hat.par[6,2]   5.272194 1.011627   230
#> hat.par[7,2]   5.103067 1.012555   230
#> hat.par[8,2]  15.000000 1.084359  3000
#> hat.par[9,2]  22.167800 1.004334   520
#> hat.par[10,2]  5.989117 1.037229    73
#> hat.par[11,2]  9.901348 1.020325   110
#> hat.par[9,3]  25.462852 1.024125    97
#> hat.par[10,3] 10.733275 1.001403  2200
#> hat.par[11,3] 14.311746 1.015586   150
#> 
#> $leverage_o
#>  [1]  0.985655246  0.575542703  1.228236532  0.902403918  0.995473428
#>  [6]  0.701714463  0.830219835 -0.006215128  0.781263695  0.788079501
#> [11]  0.682552727  0.970981289  0.982845000  0.994710699  0.979692783
#> [16]  1.079835640  0.773237438  0.626952701  0.006214386  0.917458542
#> [21]  0.660572562  0.810647453  0.763715079  0.730488506  0.955714700
#> 
#> $sign_dev_o
#>  [1] -1 -1 -1  1 -1  1  1 -1 -1  1 -1 -1  1  1 -1  1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.63923086  0.42951255  0.06153243  0.28004386  0.59011883  0.89141283 
#>       97.5%        Rhat       n.eff 
#>  1.64156031  1.18046201 16.00000000 
#> 
#> $m_tau
#>         mean           sd         2.5%          25%          50%          75% 
#> 6.823073e-01 5.101420e-01 5.061281e-04 2.753894e-01 6.061996e-01 9.971981e-01 
#>        97.5%         Rhat        n.eff 
#> 1.881547e+00 1.310647e+00 1.800000e+01 
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 43.39386 19.71799 23.67586
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "11", fit using jags,
#>  3 chains, each with 1000 iterations (first 0 discarded)
#>  n.sims = 3000 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.595   0.843 -1.201  0.158  0.600  1.055  2.345 1.026  3000
#> EM[3,1]         0.159   0.694 -1.287 -0.242  0.172  0.613  1.527 1.018   660
#> EM[4,1]         0.743   0.522 -0.367  0.463  0.760  1.052  1.736 1.037   290
#> EM[5,1]         1.831   0.778  0.447  1.340  1.778  2.260  3.521 1.029   100
#> EM[4,2]        -0.452   0.917 -2.339 -0.970 -0.473  0.094  1.454 1.019   700
#> EM[6,2]        -1.463   1.135 -3.697 -2.159 -1.404 -0.768  0.729 1.006   860
#> EM[4,3]         1.112   0.935 -0.961  0.599  1.188  1.637  2.957 1.004   880
#> EM[5,4]        -1.086   1.476 -3.889 -2.063 -1.083 -0.164  1.853 1.033   680
#> dev.o[1,1]      0.986   1.405  0.001  0.110  0.467  1.286  4.980 1.003  1100
#> dev.o[2,1]      0.920   1.261  0.001  0.096  0.446  1.214  4.597 1.002  1900
#> dev.o[3,1]      1.244   1.719  0.001  0.116  0.551  1.685  6.128 1.005   450
#> dev.o[4,1]      0.903   1.257  0.001  0.088  0.405  1.224  4.490 1.001  3000
#> dev.o[5,1]      1.007   1.443  0.001  0.109  0.462  1.365  4.742 1.004  3000
#> dev.o[6,1]      1.070   1.441  0.001  0.101  0.499  1.479  5.058 1.004   550
#> dev.o[7,1]      0.921   1.166  0.001  0.117  0.473  1.260  4.348 1.003   800
#> dev.o[8,1]      0.005   0.077  0.000  0.000  0.000  0.000  0.014 1.076  3000
#> dev.o[9,1]      0.822   1.158  0.001  0.080  0.370  1.125  4.379 1.001  3000
#> dev.o[10,1]     0.995   1.299  0.001  0.115  0.497  1.333  4.804 1.016   180
#> dev.o[11,1]     0.820   1.195  0.001  0.076  0.352  1.054  4.378 1.009   230
#> dev.o[1,2]      0.971   1.370  0.001  0.095  0.450  1.279  4.823 1.002  1200
#> dev.o[2,2]      1.178   1.606  0.001  0.116  0.541  1.639  5.902 1.001  3000
#> dev.o[3,2]      1.005   1.451  0.001  0.100  0.450  1.283  5.102 1.005   440
#> dev.o[4,2]      0.981   1.392  0.001  0.096  0.434  1.308  5.057 1.007   480
#> dev.o[5,2]      1.091   1.479  0.001  0.109  0.497  1.509  5.372 1.006  1000
#> dev.o[6,2]      1.164   1.446  0.001  0.128  0.600  1.704  5.149 1.003   770
#> dev.o[7,2]      0.735   1.039  0.001  0.075  0.331  0.974  3.783 1.001  3000
#> dev.o[8,2]      0.010   0.127  0.000  0.000  0.000  0.000  0.030 1.084  3000
#> dev.o[9,2]      1.013   1.390  0.002  0.118  0.473  1.357  5.125 1.003   910
#> dev.o[10,2]     1.446   1.732  0.003  0.194  0.797  2.145  6.154 1.009   240
#> dev.o[11,2]     1.244   1.591  0.002  0.152  0.651  1.721  5.661 1.009   290
#> dev.o[9,3]      1.249   1.599  0.002  0.148  0.638  1.733  5.705 1.016   160
#> dev.o[10,3]     0.880   1.252  0.001  0.088  0.407  1.149  4.343 1.004   590
#> dev.o[11,3]     1.015   1.506  0.001  0.099  0.428  1.289  5.299 1.002  3000
#> hat.par[1,1]   27.084   4.560 18.529 23.847 26.924 30.171 36.348 1.001  3000
#> hat.par[2,1]    2.881   1.329  0.779  1.913  2.714  3.649  5.992 1.017   170
#> hat.par[3,1]   10.157   1.239  7.104  9.499 10.396 11.116 11.774 1.008  1000
#> hat.par[4,1]   22.965   2.368 17.982 21.437 23.073 24.637 27.196 1.001  2400
#> hat.par[5,1]    8.229   2.083  4.514  6.717  8.169  9.625 12.339 1.001  3000
#> hat.par[6,1]    3.269   0.964  1.408  2.558  3.325  3.961  5.027 1.010   210
#> hat.par[7,1]    2.545   1.226  0.757  1.623  2.331  3.293  5.338 1.011   190
#> hat.par[8,1]   11.997   0.037 11.993 12.000 12.000 12.000 12.000 1.076  3000
#> hat.par[9,1]   15.580   2.497 10.896 13.862 15.540 17.299 20.568 1.007   330
#> hat.par[10,1]   3.296   1.230  1.278  2.410  3.170  4.040  6.050 1.024    86
#> hat.par[11,1]   4.664   1.501  1.933  3.663  4.584  5.604  7.912 1.016   750
#> hat.par[1,2]   38.074   5.033 28.641 34.565 38.003 41.411 48.226 1.002  1200
#> hat.par[2,2]    4.169   1.638  1.483  2.924  3.976  5.212  7.824 1.003  1000
#> hat.par[3,2]    7.852   1.412  4.756  6.936  7.973  8.906 10.197 1.005  2300
#> hat.par[4,2]   16.088   2.457 11.057 14.491 16.195 17.848 20.519 1.004   670
#> hat.par[5,2]    2.837   1.482  0.639  1.716  2.603  3.739  6.290 1.010   320
#> hat.par[6,2]    3.753   0.913  1.818  3.126  3.819  4.469  5.272 1.012   230
#> hat.par[7,2]    2.453   1.149  0.639  1.588  2.304  3.150  5.103 1.013   230
#> hat.par[8,2]   14.995   0.060 14.985 15.000 15.000 15.000 15.000 1.084  3000
#> hat.par[9,2]   16.900   2.741 11.513 14.966 16.981 18.815 22.168 1.004   520
#> hat.par[10,2]   3.283   1.315  1.004  2.319  3.199  4.213  5.989 1.037    73
#> hat.par[11,2]   6.822   1.564  3.884  5.741  6.789  7.888  9.901 1.020   110
#> hat.par[9,3]   21.305   2.246 16.756 19.831 21.407 22.853 25.463 1.024    97
#> hat.par[10,3]   8.397   1.356  5.486  7.530  8.497  9.373 10.733 1.001  2200
#> hat.par[11,3]  11.443   1.630  7.986 10.405 11.569 12.588 14.312 1.016   150
#> m.tau           0.682   0.510  0.001  0.275  0.606  0.997  1.882 1.311    18
#> resdev.o[1]     1.957   1.965  0.051  0.560  1.348  2.715  7.185 1.005  1100
#> resdev.o[2]     2.098   1.967  0.056  0.687  1.540  2.901  7.251 1.008   730
#> resdev.o[3]     2.248   2.250  0.049  0.615  1.578  3.128  8.212 1.013   160
#> resdev.o[4]     1.883   1.888  0.041  0.557  1.322  2.579  7.080 1.003   870
#> resdev.o[5]     2.098   2.073  0.056  0.588  1.461  2.960  7.671 1.002  2700
#> resdev.o[6]     2.234   1.814  0.098  0.970  1.852  2.959  6.884 1.024   140
#> resdev.o[7]     1.657   1.543  0.054  0.550  1.191  2.271  5.788 1.015   250
#> resdev.o[8]     0.016   0.183  0.000  0.000  0.000  0.000  0.053 1.073  3000
#> resdev.o[9]     3.084   2.359  0.254  1.334  2.515  4.249  9.106 1.017   160
#> resdev.o[10]    3.321   2.424  0.303  1.537  2.809  4.522  9.289 1.010   230
#> resdev.o[11]    3.079   2.402  0.259  1.331  2.527  4.150  9.265 1.001  3000
#> tau             0.639   0.430  0.062  0.280  0.590  0.891  1.642 1.180    16
#> totresdev.o    23.676   6.780 12.521 18.859 22.936 27.671 38.969 1.002  1000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> $m_tau
#>         mean           sd         2.5%          25%          50%          75% 
#> 6.823073e-01 5.101420e-01 5.061281e-04 2.753894e-01 6.061996e-01 9.971981e-01 
#>        97.5%         Rhat        n.eff 
#> 1.881547e+00 1.310647e+00 1.800000e+01 
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
#> attr(,"class")
#> [1] "run_ume"
# }