Perform the unrelated mean effects model

Performs the unrelated mean effects model of Dias et al. (2013) that has been refined (Spineli, 2021) and extended to address aggregate binary and continuous missing participant outcome data via the pattern-mixture model (Spineli et al. 2021; Spineli, 2019). This model offers a global evaluation of the plausibility of the consistency assumption in the network.

Usage

run_ume(full, n_iter, n_burnin, n_chains, n_thin, inits = NULL)

Arguments

full

An object of S3 class run_model. See 'Value' in run_model.

n_iter

Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin

Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains

Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin

Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

inits

A list with the initial values for the parameters; an argument of the jags function of the R-package R2jags. The default argument is NULL, and JAGS generates the initial values.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters:

EM

The summary effect estimate (according to the argument measure defined in run_model) for each pairwise comparison observed in the network.

dev_o

The deviance contribution of each trial-arm based on the observed outcome.

hat_par

The fitted outcome at each trial-arm.

tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau

The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements:

leverage_o

The leverage for the observed outcome at each trial-arm.

sign_dev_o

The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment

A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit

An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between the non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in

ume_plot function.

Details

run_ume inherits the arguments data, measure, model, assumption, heter_prior, mean_misspar, var_misspar, and ref from run_model. This prevents specifying a different Bayesian model from that considered in run_model.Therefore, the user needs first to apply run_model, and then use run_ume (see 'Examples').

The run_ume function also returns the arguments data, model, measure, assumption, n_chains, n_iter, n_burnin, and n_thin as specified by the user to be inherited by other relevant functions of the package.

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials not investigated in any two-arm or multi-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effects is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

For a binary outcome, when measure is "RR" (relative risk) or "RD" (risk difference) in run_model, run_ume currently considers the odds ratio as effect measure for being the base-case effect measure in run_model for a binary outcome (see also 'Details' in run_model).

run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see 'Details' in run_model). prepare_ume has also been refined to account for the multi-arm trials by assigning conditional univariate normal distributions on the underlying trial-specific effect size of comparisons with the baseline arm of the multi-arm trial (Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears on the R console. The model is updated until convergence using the autojags function of the R-package R2jags with 2 updates and number of iterations and thinning equal to n_iter and n_thin, respectively.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed on the R console.

References

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013;33(5):641--56. doi: 10.1177/0272989X12455847

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7(4):457--72. doi: 10.1214/ss/1177011136

Spineli LM. A Revised Framework to Evaluate the Consistency Assumption Globally in a Network of Interventions. Med Decis Making 2021. doi: 10.1177/0272989X211068005

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021;30(4):958--75. doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. doi: 10.1186/s12874-019-0731-y

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

# Read results from 'run_model' (using the default arguments)
res <- readRDS(system.file('extdata/res_liu.rds', package = 'rnmamod'))

# \donttest{
# Run random-effects unrelated mean effects model
# Note: Ideally, set 'n_iter' to 10000 and 'n_burnin' to 1000
run_ume(full = res,
        n_chains = 3,
        n_iter = 1000,
        n_burnin = 100,
        n_thin = 1)
#> JAGS generates initial values for the parameters.
#> Running the model ...
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 59
#>    Unobserved stochastic nodes: 80
#>    Total graph size: 1291
#> 
#> Initializing model
#> 
#> ... Updating the model until convergence
#> $EM
#>                mean        sd       2.5%        25%         50%        75%
#> EM[2,1]  0.56911443 0.7424087 -0.9442177  0.1939775  0.56943179  0.9463432
#> EM[3,1]  0.09710686 0.6422076 -1.0421827 -0.3012057  0.05253696  0.4456814
#> EM[4,1]  0.79865439 0.4928084 -0.1812311  0.4930031  0.80784454  1.0991819
#> EM[5,1]  1.85084302 0.6791482  0.5614965  1.4327493  1.81659200  2.2531572
#> EM[4,2] -0.39512817 0.8982306 -2.1058300 -0.9290301 -0.38989383  0.1306456
#> EM[6,2] -1.26814972 1.0429818 -3.3990740 -1.8812483 -1.27209101 -0.6484963
#> EM[4,3]  1.21488018 0.9862657 -0.9154718  0.6840412  1.24564383  1.7857711
#> EM[5,4] -1.01409085 1.2983241 -3.5798129 -1.8339243 -0.92055387 -0.1781273
#>             97.5%     Rhat n.eff
#> EM[2,1] 2.0497458 1.037441  2100
#> EM[3,1] 1.4510370 1.053738    64
#> EM[4,1] 1.7673321 1.024219   710
#> EM[5,1] 3.2841632 1.049211   140
#> EM[4,2] 1.3492224 1.042562   120
#> EM[6,2] 0.8439245 1.031494   160
#> EM[4,3] 3.0536641 1.012045   260
#> EM[5,4] 1.4468212 1.021800   390
#> 
#> $dev_o
#>                    mean         sd         2.5%        25%       50%
#> dev.o[1,1]  0.926088526 1.30907340 0.0008515154 0.09528389 0.4166993
#> dev.o[2,1]  0.993259987 1.24801276 0.0013635534 0.11296652 0.5209257
#> dev.o[3,1]  1.140517254 1.55672007 0.0011865393 0.11631916 0.5412035
#> dev.o[4,1]  1.015985535 1.45111453 0.0013630328 0.11876141 0.4921627
#> dev.o[5,1]  1.002588877 1.44171650 0.0004110865 0.10386998 0.4534699
#> dev.o[6,1]  1.134241818 1.47426736 0.0008570884 0.12851950 0.5906580
#> dev.o[7,1]  0.930006010 1.25176408 0.0010263542 0.10074683 0.4413111
#> dev.o[8,1]  0.006603266 0.07675424 0.0000000000 0.00000000 0.0000000
#> dev.o[9,1]  1.013505521 1.42600793 0.0010937181 0.10391326 0.4657980
#> dev.o[10,1] 1.032135171 1.33730147 0.0017917369 0.13592542 0.5349400
#> dev.o[11,1] 0.687808382 0.94459078 0.0007770346 0.07464218 0.3268722
#> dev.o[1,2]  0.996375444 1.43313746 0.0015157846 0.09908864 0.4537787
#> dev.o[2,2]  1.254541566 1.55559904 0.0010120808 0.14914891 0.6808240
#> dev.o[3,2]  1.046964378 1.42519425 0.0013454723 0.10715055 0.4865362
#> dev.o[4,2]  1.046677222 1.46959327 0.0009163646 0.10104649 0.4696106
#> dev.o[5,2]  0.978358793 1.43935767 0.0012427111 0.10251430 0.4208983
#> dev.o[6,2]  1.184474644 1.54434287 0.0015406375 0.14481194 0.5709247
#> dev.o[7,2]  0.810619682 1.16716077 0.0008185285 0.08477466 0.3606950
#> dev.o[8,2]  0.009634022 0.07837520 0.0000000000 0.00000000 0.0000000
#> dev.o[9,2]  0.955944133 1.30792822 0.0009651527 0.10105393 0.4601795
#> dev.o[10,2] 1.569978140 1.77743434 0.0029302742 0.24541495 0.9709812
#> dev.o[11,2] 1.145046671 1.40742031 0.0014982002 0.15197685 0.6643851
#> dev.o[9,3]  1.216559715 1.52924750 0.0014045809 0.15283191 0.6520426
#> dev.o[10,3] 0.852911951 1.19966277 0.0006842221 0.08634575 0.3902506
#> dev.o[11,3] 0.822840693 1.14507337 0.0013827600 0.09245421 0.3738131
#>                      75%      97.5%     Rhat n.eff
#> dev.o[1,1]  1.222545e+00 4.66661883 1.000854  3000
#> dev.o[2,1]  1.399484e+00 4.48155346 1.003555   660
#> dev.o[3,1]  1.499143e+00 5.71701544 1.014356   150
#> dev.o[4,1]  1.302990e+00 5.08462602 1.004767  3000
#> dev.o[5,1]  1.318120e+00 5.09635086 1.000700  3000
#> dev.o[6,1]  1.575219e+00 5.33899418 1.001040  3000
#> dev.o[7,1]  1.281755e+00 4.38330065 1.001731  1800
#> dev.o[8,1]  3.863576e-13 0.03092069 1.062385  3000
#> dev.o[9,1]  1.330739e+00 5.06658808 1.007961   270
#> dev.o[10,1] 1.404068e+00 4.84010029 1.001408  2200
#> dev.o[11,1] 9.222313e-01 3.50586701 1.004702   610
#> dev.o[1,2]  1.272430e+00 5.12304540 1.000960  3000
#> dev.o[2,2]  1.805199e+00 5.54766407 1.006020   360
#> dev.o[3,2]  1.434966e+00 5.03282613 1.004127   550
#> dev.o[4,2]  1.402855e+00 5.25048638 1.002798  3000
#> dev.o[5,2]  1.247183e+00 4.85488118 1.005385   410
#> dev.o[6,2]  1.676240e+00 5.65484650 1.001506  2000
#> dev.o[7,2]  1.023649e+00 4.27428230 1.001903  2800
#> dev.o[8,2]  9.658940e-13 0.05799271 1.056367  1400
#> dev.o[9,2]  1.283508e+00 4.51817817 1.004083   560
#> dev.o[10,2] 2.222685e+00 6.39048691 1.021790   100
#> dev.o[11,2] 1.615030e+00 5.13602176 1.002378  1300
#> dev.o[9,3]  1.690458e+00 5.54582573 1.004384   540
#> dev.o[10,3] 1.121319e+00 4.08309915 1.003146   760
#> dev.o[11,3] 1.084606e+00 4.18357627 1.004200   540
#> 
#> $hat_par
#>                    mean         sd       2.5%       25%       50%       75%
#> hat.par[1,1]  26.898931 4.42059221 18.7275282 23.823743 26.750648 29.763332
#> hat.par[2,1]   2.962723 1.34399420  0.8053603  1.944553  2.833350  3.834491
#> hat.par[3,1]  10.085034 1.22363592  7.1993896  9.394393 10.321199 11.000159
#> hat.par[4,1]  22.854010 2.53327196 17.3531827 21.176633 23.084673 24.724494
#> hat.par[5,1]   8.006743 2.06952864  4.0419513  6.550721  7.978648  9.398931
#> hat.par[6,1]   3.219836 0.96906178  1.3708064  2.504234  3.217294  3.940030
#> hat.par[7,1]   2.500851 1.18509385  0.7226296  1.614608  2.338106  3.204165
#> hat.par[8,1]  11.996759 0.03693807 11.9845496 12.000000 12.000000 12.000000
#> hat.par[9,1]  16.106206 2.62222754 10.8886579 14.349956 16.153600 17.867648
#> hat.par[10,1]  3.162415 1.15556027  1.2355367  2.335551  3.035938  3.863008
#> hat.par[11,1]  4.728240 1.34112233  2.3444710  3.768898  4.670464  5.618914
#> hat.par[1,2]  38.025700 5.09424203 28.3864960 34.521911 37.885781 41.390515
#> hat.par[2,2]   4.005340 1.60831598  1.5289867  2.822268  3.762443  4.980544
#> hat.par[3,2]   7.979670 1.42291934  4.7914668  7.079193  8.158633  9.064459
#> hat.par[4,2]  16.136163 2.52422276 11.0413391 14.440951 16.183770 17.910695
#> hat.par[5,2]   3.033550 1.49932698  0.7810689  1.954509  2.777249  3.913928
#> hat.par[6,2]   3.753348 0.91640421  1.8580666  3.141669  3.810473  4.457126
#> hat.par[7,2]   2.586112 1.20079179  0.7138016  1.687813  2.433438  3.300853
#> hat.par[8,2]  14.995234 0.03857903 14.9710317 15.000000 15.000000 15.000000
#> hat.par[9,2]  16.653265 2.72429872 11.3349916 14.772586 16.697172 18.542909
#> hat.par[10,2]  3.419261 1.29500244  1.0527209  2.502023  3.374132  4.268384
#> hat.par[11,2]  6.804395 1.46018648  4.0374112  5.805493  6.730915  7.802571
#> hat.par[9,3]  21.296711 2.18787750 16.8273922 19.859798 21.355018 22.809762
#> hat.par[10,3]  8.332110 1.31431073  5.6001907  7.474787  8.426055  9.290199
#> hat.par[11,3] 11.547117 1.43573311  8.4162562 10.632189 11.679687 12.530320
#>                   97.5%     Rhat n.eff
#> hat.par[1,1]  36.197304 1.007272   300
#> hat.par[2,1]   5.849454 1.036985    73
#> hat.par[3,1]  11.772682 1.013009  1200
#> hat.par[4,1]  27.158133 1.022595   120
#> hat.par[5,1]  12.161047 1.009115   230
#> hat.par[6,1]   5.055230 1.002040  2200
#> hat.par[7,1]   5.225126 1.003944   580
#> hat.par[8,1]  12.000000 1.062385  3000
#> hat.par[9,1]  21.122989 1.035195    71
#> hat.par[10,1]  5.775420 1.009180   230
#> hat.par[11,1]  7.580583 1.024433   160
#> hat.par[1,2]  48.454671 1.004991   600
#> hat.par[2,2]   7.694355 1.022580    93
#> hat.par[3,2]  10.222202 1.014090   590
#> hat.par[4,2]  20.930086 1.012906   160
#> hat.par[5,2]   6.649207 1.028523   120
#> hat.par[6,2]   5.324713 1.012280   210
#> hat.par[7,2]   5.412771 1.012084   190
#> hat.par[8,2]  15.000000 1.056367  1400
#> hat.par[9,2]  21.875635 1.010164   210
#> hat.par[10,2]  6.069128 1.051814    48
#> hat.par[11,2]  9.873541 1.011775   270
#> hat.par[9,3]  25.464892 1.010274   240
#> hat.par[10,3] 10.608983 1.017958   120
#> hat.par[11,3] 14.054206 1.009906   540
#> 
#> $leverage_o
#>  [1]  0.925612941  0.589166589  1.136077006  1.012712103  1.002579061
#>  [6]  0.715981600  0.819032122 -0.002414968  0.864824285  0.734756827
#> [11]  0.523329912  0.996350389  0.969744881  1.046775468  1.043825469
#> [16]  0.977905842  0.793662822  0.634043646  0.005277768  0.905678042
#> [21]  0.626037352  0.699104041  0.726319676  0.670996339  0.731870597
#> 
#> $sign_dev_o
#>  [1]  1 -1 -1  1 -1  1  1 -1 -1  1 -1 -1  1  1 -1 -1 -1 -1 -1 -1 -1  1  1  1 -1
#> 
#> $tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.54814596  0.39775683  0.06492175  0.26412469  0.45989756  0.74821929 
#>       97.5%        Rhat       n.eff 
#>  1.51559664  1.12258641 22.00000000 
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.64919257  0.53131485  0.02489219  0.22847195  0.52139340  0.94665864 
#>       97.5%        Rhat       n.eff 
#>  1.93941137  1.05011956 54.00000000 
#> 
#> $model_assessment
#>        DIC       pD      dev
#> 1 42.92292 19.14925 23.77367
#> 
#> $obs_comp
#> [1] "2vs1" "3vs1" "4vs1" "5vs1" "4vs2" "6vs2" "4vs3" "5vs4"
#> 
#> $jagsfit
#> Inference for Bugs model at "13", fit using jags,
#>  3 chains, each with 1000 iterations (first 0 discarded)
#>  n.sims = 3000 iterations saved
#>               mu.vect sd.vect   2.5%    25%    50%    75%  97.5%  Rhat n.eff
#> EM[2,1]         0.569   0.742 -0.944  0.194  0.569  0.946  2.050 1.037  2100
#> EM[3,1]         0.097   0.642 -1.042 -0.301  0.053  0.446  1.451 1.054    64
#> EM[4,1]         0.799   0.493 -0.181  0.493  0.808  1.099  1.767 1.024   710
#> EM[5,1]         1.851   0.679  0.561  1.433  1.817  2.253  3.284 1.049   140
#> EM[4,2]        -0.395   0.898 -2.106 -0.929 -0.390  0.131  1.349 1.043   120
#> EM[6,2]        -1.268   1.043 -3.399 -1.881 -1.272 -0.648  0.844 1.031   160
#> EM[4,3]         1.215   0.986 -0.915  0.684  1.246  1.786  3.054 1.012   260
#> EM[5,4]        -1.014   1.298 -3.580 -1.834 -0.921 -0.178  1.447 1.022   390
#> dev.o[1,1]      0.926   1.309  0.001  0.095  0.417  1.223  4.667 1.001  3000
#> dev.o[2,1]      0.993   1.248  0.001  0.113  0.521  1.399  4.482 1.004   660
#> dev.o[3,1]      1.141   1.557  0.001  0.116  0.541  1.499  5.717 1.014   150
#> dev.o[4,1]      1.016   1.451  0.001  0.119  0.492  1.303  5.085 1.005  3000
#> dev.o[5,1]      1.003   1.442  0.000  0.104  0.453  1.318  5.096 1.001  3000
#> dev.o[6,1]      1.134   1.474  0.001  0.129  0.591  1.575  5.339 1.001  3000
#> dev.o[7,1]      0.930   1.252  0.001  0.101  0.441  1.282  4.383 1.002  1800
#> dev.o[8,1]      0.007   0.077  0.000  0.000  0.000  0.000  0.031 1.062  3000
#> dev.o[9,1]      1.014   1.426  0.001  0.104  0.466  1.331  5.067 1.008   270
#> dev.o[10,1]     1.032   1.337  0.002  0.136  0.535  1.404  4.840 1.001  2200
#> dev.o[11,1]     0.688   0.945  0.001  0.075  0.327  0.922  3.506 1.005   610
#> dev.o[1,2]      0.996   1.433  0.002  0.099  0.454  1.272  5.123 1.001  3000
#> dev.o[2,2]      1.255   1.556  0.001  0.149  0.681  1.805  5.548 1.006   360
#> dev.o[3,2]      1.047   1.425  0.001  0.107  0.487  1.435  5.033 1.004   550
#> dev.o[4,2]      1.047   1.470  0.001  0.101  0.470  1.403  5.250 1.003  3000
#> dev.o[5,2]      0.978   1.439  0.001  0.103  0.421  1.247  4.855 1.005   410
#> dev.o[6,2]      1.184   1.544  0.002  0.145  0.571  1.676  5.655 1.002  2000
#> dev.o[7,2]      0.811   1.167  0.001  0.085  0.361  1.024  4.274 1.002  2800
#> dev.o[8,2]      0.010   0.078  0.000  0.000  0.000  0.000  0.058 1.056  1400
#> dev.o[9,2]      0.956   1.308  0.001  0.101  0.460  1.284  4.518 1.004   560
#> dev.o[10,2]     1.570   1.777  0.003  0.245  0.971  2.223  6.390 1.022   100
#> dev.o[11,2]     1.145   1.407  0.001  0.152  0.664  1.615  5.136 1.002  1300
#> dev.o[9,3]      1.217   1.529  0.001  0.153  0.652  1.690  5.546 1.004   540
#> dev.o[10,3]     0.853   1.200  0.001  0.086  0.390  1.121  4.083 1.003   760
#> dev.o[11,3]     0.823   1.145  0.001  0.092  0.374  1.085  4.184 1.004   540
#> hat.par[1,1]   26.899   4.421 18.728 23.824 26.751 29.763 36.197 1.007   300
#> hat.par[2,1]    2.963   1.344  0.805  1.945  2.833  3.834  5.849 1.037    73
#> hat.par[3,1]   10.085   1.224  7.199  9.394 10.321 11.000 11.773 1.013  1200
#> hat.par[4,1]   22.854   2.533 17.353 21.177 23.085 24.724 27.158 1.023   120
#> hat.par[5,1]    8.007   2.070  4.042  6.551  7.979  9.399 12.161 1.009   230
#> hat.par[6,1]    3.220   0.969  1.371  2.504  3.217  3.940  5.055 1.002  2200
#> hat.par[7,1]    2.501   1.185  0.723  1.615  2.338  3.204  5.225 1.004   580
#> hat.par[8,1]   11.997   0.037 11.985 12.000 12.000 12.000 12.000 1.062  3000
#> hat.par[9,1]   16.106   2.622 10.889 14.350 16.154 17.868 21.123 1.035    71
#> hat.par[10,1]   3.162   1.156  1.236  2.336  3.036  3.863  5.775 1.009   230
#> hat.par[11,1]   4.728   1.341  2.344  3.769  4.670  5.619  7.581 1.024   160
#> hat.par[1,2]   38.026   5.094 28.386 34.522 37.886 41.391 48.455 1.005   600
#> hat.par[2,2]    4.005   1.608  1.529  2.822  3.762  4.981  7.694 1.023    93
#> hat.par[3,2]    7.980   1.423  4.791  7.079  8.159  9.064 10.222 1.014   590
#> hat.par[4,2]   16.136   2.524 11.041 14.441 16.184 17.911 20.930 1.013   160
#> hat.par[5,2]    3.034   1.499  0.781  1.955  2.777  3.914  6.649 1.029   120
#> hat.par[6,2]    3.753   0.916  1.858  3.142  3.810  4.457  5.325 1.012   210
#> hat.par[7,2]    2.586   1.201  0.714  1.688  2.433  3.301  5.413 1.012   190
#> hat.par[8,2]   14.995   0.039 14.971 15.000 15.000 15.000 15.000 1.056  1400
#> hat.par[9,2]   16.653   2.724 11.335 14.773 16.697 18.543 21.876 1.010   210
#> hat.par[10,2]   3.419   1.295  1.053  2.502  3.374  4.268  6.069 1.052    48
#> hat.par[11,2]   6.804   1.460  4.037  5.805  6.731  7.803  9.874 1.012   270
#> hat.par[9,3]   21.297   2.188 16.827 19.860 21.355 22.810 25.465 1.010   240
#> hat.par[10,3]   8.332   1.314  5.600  7.475  8.426  9.290 10.609 1.018   120
#> hat.par[11,3]  11.547   1.436  8.416 10.632 11.680 12.530 14.054 1.010   540
#> m.tau           0.649   0.531  0.025  0.228  0.521  0.947  1.939 1.050    54
#> resdev.o[1]     1.922   1.952  0.047  0.562  1.311  2.631  7.311 1.001  3000
#> resdev.o[2]     2.248   1.824  0.089  0.881  1.859  3.154  6.815 1.020   120
#> resdev.o[3]     2.187   2.139  0.056  0.613  1.534  3.054  7.974 1.024    88
#> resdev.o[4]     2.063   2.058  0.053  0.632  1.432  2.895  7.515 1.005   460
#> resdev.o[5]     1.981   2.106  0.044  0.569  1.329  2.672  7.528 1.004   760
#> resdev.o[6]     2.319   1.922  0.129  1.020  1.919  3.014  7.554 1.013   250
#> resdev.o[7]     1.741   1.716  0.061  0.544  1.230  2.358  6.356 1.004   510
#> resdev.o[8]     0.016   0.132  0.000  0.000  0.000  0.000  0.105 1.020  2100
#> resdev.o[9]     3.186   2.412  0.287  1.359  2.662  4.397  9.346 1.008   270
#> resdev.o[10]    3.455   2.354  0.351  1.764  2.978  4.672  9.367 1.023    93
#> resdev.o[11]    2.656   1.989  0.260  1.230  2.214  3.531  7.795 1.010   290
#> tau             0.548   0.398  0.065  0.264  0.460  0.748  1.516 1.123    22
#> totresdev.o    23.774   6.263 12.947 19.358 23.256 27.558 37.346 1.014   160
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> $data
#>                study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#> 356    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#> 357     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> 358 Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#> 359      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#> 360      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#> 361   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA
#> 362     Barone, 2006  2  4 NA 23 16 NA  1  7 NA  33  34 NA
#> 363  Rektorova, 2003  2  6 NA  8  3 NA  3  2 NA  22  19 NA
#> 364  Leentjens, 2003  1  4 NA  4  3 NA  0  0 NA   6   6 NA
#> 365    Wermuth, 1998  1  4 NA  3  2 NA  2  5 NA  19  18 NA
#> 366      Rabey, 1996  4  5 NA 12 15 NA  8 12 NA  20  27 NA
#> 
#> $model
#> [1] "RE"
#> 
#> $measure
#> [1] "OR"
#> 
#> $assumption
#> [1] "IDE-ARM"
#> 
#> $phi
#> NULL
#> 
#> $n_chains
#> [1] 3
#> 
#> $n_iter
#> [1] 1000
#> 
#> $n_burnin
#> [1] 100
#> 
#> $n_thin
#> [1] 1
#> 
#> $m_tau
#>        mean          sd        2.5%         25%         50%         75% 
#>  0.64919257  0.53131485  0.02489219  0.22847195  0.52139340  0.94665864 
#>       97.5%        Rhat       n.eff 
#>  1.93941137  1.05011956 54.00000000 
#> 
#> $frail_comp
#> [1] "4vs3"
#> 
#> attr(,"class")
#> [1] "run_ume"
# }