A function to perform Bayesian unrelated mean effects model

Performs a Bayesian network meta-analysis based on the unrelated mean effects model of Dias et al. (2013a) extended to address aggregate binary and continuous participant outcome data via the pattern-mixture model (Spineli, 2019; Spineli et al., 2021). This model offers a global evaluation of the plausibility of the consistency assumption in the network (Dias et al. (2013b)).

run_ume(full, n_iter, n_burnin, n_chains, n_thin)

Arguments

full
An object of S3 class run_model. See 'Value' in run_model.

n_iter
Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 10000.

n_burnin
Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1000.

n_chains
Positive integer specifying the number of chains for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 2.

n_thin
Positive integer specifying the thinning rate for the MCMC sampling; an argument of the jags function of the R-package R2jags. The default argument is 1.

Value

An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:

EM	The summary effect estimate for each pairwise comparison observed in the network.

dev_o	The deviance contribution of each trial-arm based on the observed outcome.

hat_par	The fitted outcome at each trial-arm.

tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the whole network, when a random-effects model has been specified.

m_tau	The between-trial standard deviation (assumed common across the observed pairwise comparisons) for the subset of multi-arm trials, when a random-effects model has been specified.

The output also includes the following elements - the first three resulting from relevant monitored parameters:

leverage_o	The leverage for the observed outcome at each trial-arm.

sign_dev_o	The sign of the difference between observed and fitted outcome at each trial-arm.

model_assessment	A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.

jagsfit	An object of S3 class jags with the posterior results on all monitored parameters to be used in the mcmc_diagnostics function.

Furthermore, run_ume returns a character vector with the pairwise comparisons observed in the network, obs_comp, and a character vector with comparisons between non-baseline interventions observed in multi-arm trials only, frail_comp. Both vectors are used in ume_plot function.

Details

run_ume does not contain the arguments data, measure, model, assumption, heter_prior, mean_misspar, and var_misspar that are found in run_model. This is to prevent misspecifying the Bayesian model that has been considered in run_model. Instead, these arguments are contained in the argument full of the function. Therefore, the user needs first to apply run_model, and then usenrun_ume (see, 'Examples').

Initially, run_ume calls the improved_ume function to identify the frail comparisons, that is, comparisons between non-baseline interventions in multi-arm trials -- not investigated in any two-arm trial of the network (Spineli, 2021). The 'original' model of Dias et al. (2013a) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effect sizes is less than those obtained by performing separate pairwise meta-analyses (see run_series_meta).

Then, to perform the Bayesian unrelated effect measures model, run_ume calls the prepare_ume function which contains the WinBUGS code as written by Dias et al. (2013a) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively. prepare_ume has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see, 'Details' in run_model). prepare_ume has been also 'upgraded' to account for the multi-arm trials by assigning conditional univariate normal distributions on the basic parameters of these trials, that is, effect parameters between non-baseline and baseline arms (Dias et al., 2013b; Spineli, 2021).

run_ume runs Bayesian unrelated mean effects model in JAGS. The progress of the simulation appears in the R console.

The output of run_ume is not end-user-ready. The ume_plot function uses the output of run_ume as an S3 object and processes it further to provide an end-user-ready output.

run_ume can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.

References

Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. submitted 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. Stat Methods Med Res 2021. [doi: 10.1177/0962280220983544 ]

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. BMC Med Res Methodol 2019;19(1):86. [doi: 10.1186/s12874-019-0731-y ]

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. Med Decis Making 2013a;33(5):641--56. [doi: 10.1177/0272989X12455847 ]

Dias S, Sutton AJ, Ades AE, Welton NJ. Evidence synthesis for decision making 2: a generalized linear modeling framework for pairwise and network meta-analysis of randomized controlled trials. Med Decis Making 2013b;33(5):607--617. [doi: 10.1177/0272989X12458724 ]

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. Stat Sci 1992;7:457--472. [doi: 10.1214/ss/1177011136 ]

See also

Author

Loukia M. Spineli

Examples

data("nma.liu2013")

if (FALSE) {
# Perform a random-effects network meta-analysis (consistency model)
res <- run_model(data = nma.liu2013,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 1,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# Run random-effects unrelated mean effects model
run_ume(full = res,
        n_chains = 3,
        n_iter = 10000,
        n_burnin = 1000,
        n_thin = 1)
}